### 2025-07-31 #8678 @editorialbot
[PRE REVIEW]: FDTDX: High-Performance Open-Source FDTD Simulation with Automatic Differentiation
**Submitting author:** @ymahlau (<a href="http://orcid.org/0000-0003-0425-5003">Yannik Mahlau</a>)
**Repository:** https://github.com/ymahlau/fdtdx/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.2
