### 2025-07-31 #8678 @kyleniemeyer
For the record, the methodology of this software was described in a recent paper, but which did not focus on the software itself: https://pubs.acs.org/doi/10.1021/acsomega.4c10958. This is fine for JOSS since that paper focuses on the underlying methodology and research results.
