### 2025-07-31 #8678 @kyleniemeyer
@ymahlau I can see that you have tests and CI set up, but can you add a badge or something indicating the status to the README? This is a common way for people to quickly see that tests are passing.
