### 2025-07-31 4be2817
Readme.md: Added JOSS and CI build badge
