### 2025-07-31 645a673
paper.bib: Added missing DOIs
