+++ paper.bib
@@ -6,7 +6,8 @@ month = may,
 journal = {IEEE Transactions on Antennas and Propagation},
 volume = {14},
 number = {3},
-pages = {302--307}
+pages = {302--307},
+doi = {10.1109/TAP.1966.1138693}
 }
 
 @article{ceviche,
@@ -17,7 +18,8 @@ volume={6},
 number={11},
 pages={3010--3016},
 year={2019},
-publisher={ACS Publications}
+publisher={ACS Publications},
+doi = {10.1021/acsphotonics.9b01238}
 }
 
 @misc{fdtdz,
@@ -44,7 +46,8 @@ pages = {687-702},
 year = {2010},
 issn = {0010-4655},
 author = {Ardavan F. Oskooi and David Roundy and Mihai Ibanescu and Peter Bermel and J.D. Joannopoulos and Steven G. Johnson},
-keywords = {Computational electromagnetism, FDTD, Maxwell solver}
+keywords = {Computational electromagnetism, FDTD, Maxwell solver},
+doi = {10.1016/j.cpc.2009.11.008}
 }
 
 @misc{openEMS,
@@ -84,9 +87,9 @@ author = {Taflove, Allen and Hagness, Susan C.},
 edition = {3rd},
 keywords = {EM FDTD textbook},
 publisher = {Artech House},
-timestamp = {2011-05-09T14:57:54.000+0200},
 title = {Computational electrodynamics: the finite-difference time-domain method},
-year = 2005
+year = {2005},
+isbn = {9781580538329}
 }
 
 
@@ -104,14 +107,16 @@ booktitle={Photonic and Phononic Properties of Engineered Nanostructures XV},
 volume={13377},
 pages={40--52},
 year={2025},
-organization={SPIE}
+organization={SPIE},
+doi={10.1117/12.3052639}
 }
 
 @article{schubertmahlau2025quantized,
 title={Quantized Inverse Design for Photonic Integrated Circuits},
 author={Schubert, Frederik and Mahlau, Yannik and Bethmann, Konrad and Hartmann, Fabian and Caspary, Reinhard and Munderloh, Marco and Ostermann, J{\"o}rn and Rosenhahn, Bodo},
 journal={ACS Omega},
-year={2025}
+year={2025},
+doi={10.1021/acsomega.4c10958}
 }
 
 @article{larsen2025integrated,
@@ -120,7 +125,8 @@ author={Larsen, MV and Bourassa, JE and Kocsis, S and Tasker, JF and Chadwick, R
 journal={Nature},
 pages={1--5},
 year={2025},
-publisher={Nature Publishing Group}
+publisher={Nature Publishing Group},
+doi={10.1038/s41586-025-09044-5}
 }
 
 @article{augenstein2020inverse,
@@ -131,7 +137,8 @@ volume={7},
 number={8},
 pages={2190--2196},
 year={2020},
-publisher={ACS Publications}
+publisher={ACS Publications},
+doi={10.1021/acsphotonics.0c00699}
 }
 
 @INPROCEEDINGS{metamaterial,
@@ -141,7 +148,8 @@ title={Impact of unit cell variation on visible spectrum multiband metamaterial 
 year={2024},
 volume={},
 number={},
-pages={1-5}
+pages={1-5},
+doi={10.1109/MAPCON61407.2024.10923310}
 }
 
 @article{cpml,
@@ -151,6 +159,7 @@ journal = {Microwave and Optical Technology Letters},
 volume = {27},
 number = {5},
 pages = {334-339},
-year = {2000}
+year = {2000},
+doi={10.1002/1098-2760(20001205)27:5<334::AID-MOP14>3.0.CO;2-A}
 }
 
