### 2025-07-31 0131cb9
jnp.pyi stub: test if external works
