+++ pyproject.toml
@@ -20,7 +20,8 @@ dependencies = [
     "seaborn>=0.13.2",
     "matplotlib>=3.9.4",
     "rich>=13.9.4",
-    "plum-dispatch>=2.5.7"
+    "plum-dispatch>=2.5.7",
+    "fastcore>=1.8.7",
 ]
 
 [project.urls]
