### 2025-07-31 900728c
further pyi stub tests
