+++ pyproject.toml
@@ -32,6 +32,9 @@ Documentation = "https://ymahlau.github.io/fdtdx/"
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
+[tool.hatch.build.targets.wheel.sources]
+external_types = "typings"
+
 [tool.ruff]
 target-version = "py312"
 line-length = 120
