### 2025-07-31 b97c305
removed pyi stub
