+++ src/fdtdx/units/unitful.py
@@ -13,7 +13,7 @@ from plum import dispatch, overload
 from pytreeclass import tree_repr
 from fastcore.foundation import patch_to
 
-from fdtdx.units.utils import patch_fn_to_module
+from fdtdx.units.utils import handle_different_scales, patch_fn_to_module
 
 
 class SI(Enum):
@@ -48,11 +48,17 @@ class Unit:
 class Unitful(TreeClass):
     val: ArrayLike
     unit: Unit = frozen_field()
+    
+    def materialise(self) -> ArrayLike:
+        if self.unit.dim:
+            raise ValueError(f"Cannot materialise unitful array with a non-zero unit: {self.unit}")
+        factor = 10 ** self.unit.scale
+        return self.val * factor
 
-    def __str__(self):
+    def __str__(self) -> str:
         return f"Unitful [{self.unit}]: {tree_repr(self.val)}"
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return str(self)
     
     def __mul__(self, other: ArrayLike | "Unitful") -> "Unitful":
@@ -60,25 +66,46 @@ class Unitful(TreeClass):
     
     def __rmul__(self, other: ArrayLike | "Unitful") -> "Unitful":
         return multiply(self, other)
+    
+    def __add__(self, other: "Unitful") -> "Unitful":
+        return add(self, other)
+    
+    def __radd__(self, other: "Unitful") -> "Unitful":
+        return add(self, other)
+    
+    def __sub__(self, other: "Unitful") -> "Unitful":
+        return subtract(self, other)
+    
+    def __rsub__(self, other: "Unitful") -> "Unitful":
+        return subtract(other, self)
+    
+    
 
 
-s = Unitful(val=1, unit=Unit(0, {SI.s: 1}))
-ms = Unitful(val=1, unit=Unit(-3, {SI.s: 1}))
-µs = Unitful(val=1, unit=Unit(-6, {SI.s: 1}))
-ns = Unitful(val=1, unit=Unit(-9, {SI.s: 1}))
-ps = Unitful(val=1, unit=Unit(-12, {SI.s: 1}))
-fs = Unitful(val=1, unit=Unit(-15, {SI.s: 1}))
+def align_scales(
+    u1: Unitful,
+    u2: Unitful,
+) -> tuple[Unitful, Unitful]:
+    if u1.unit.dim != u2.unit.dim:
+        raise Exception(f"Cannot align arrays with different units")
+    new_scale, factor1, factor2 = handle_different_scales(
+        u1.unit.scale,
+        u2.unit.scale,
+    )
+    if new_scale != u1.unit.scale:
+        u1 = Unitful(
+            val=u1.val * factor1,
+            unit=Unit(new_scale, u1.unit.dim)
+        )
+    if new_scale != u2.unit.scale:
+        u2 = Unitful(
+            val=u2.val * factor2,
+            unit=Unit(new_scale, u2.unit.dim)
+        )
+    return u1, u2
 
-m_per_s = Unitful(val=1, unit=Unit(0, {SI.s: -1, SI.m: 1}))
 
-Hz = Unitful(val=1, unit=Unit(0, {SI.s: -1}))
-kHz = Unitful(val=1, unit=Unit(3, {SI.s: -1}))
-MHz = Unitful(val=1, unit=Unit(6, {SI.s: -1}))
-GHz = Unitful(val=1, unit=Unit(9, {SI.s: -1}))
-THz = Unitful(val=1, unit=Unit(12, {SI.s: -1}))
-PHz = Unitful(val=1, unit=Unit(15, {SI.s: -1}))
-
-
+## Multiplication ###########################
 @overload
 def multiply(  # type: ignore
[…]
