### 2025-07-31 613982f
Units: implemented lt, le, eq, ne, ge, gt
