### 2025-08-01 5018d65
Fraction class implemented
