### 2025-08-01 eb1b115
added Fraction to Unit dim, moved SI-Enum to typing
