### 2025-08-01 88b07b8
implemented best_scale, PhysicalArrayLike
