+++ src/fdtdx/units/unitful.py
@@ -1,5 +1,5 @@
+from dataclasses import dataclass
 import jax
-from jaxtyping import ArrayLike
 import jax.numpy as jnp
 
 from fdtdx.core.fraction import Fraction
@@ -8,18 +8,13 @@ from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field
 from plum import dispatch, overload
 from pytreeclass import tree_repr
 
-from fdtdx.typing import SI
-from fdtdx.units.utils import handle_different_scales, patch_fn_to_module
+from fdtdx.typing import SI, PhysicalArrayLike
+from fdtdx.units.utils import best_scale, handle_different_scales, patch_fn_to_module
 
-
-class Unit:
-    def __init__(
-        self,
-        scale: int,
-        dim: dict[SI, int | Fraction],
-    ):
-        self.scale = scale
-        self.dim = dim
+@autoinit
+class Unit(TreeClass):
+    scale: int = frozen_field()
+    dim: dict[SI, int | Fraction] = frozen_field()
 
     def __str__(self) -> str:
         res_str = f"10^{self.scale} " if self.scale != 0 else ""
@@ -33,10 +28,18 @@ class Unit:
 
 @autoinit
 class Unitful(TreeClass):
-    val: ArrayLike
+    val: PhysicalArrayLike
     unit: Unit = frozen_field()
     
-    def materialise(self) -> ArrayLike:
+    def __post_init__(self):
+        new_arr, power = best_scale(self.val)
+        self.val = new_arr
+        self.unit = Unit(
+            scale=self.unit.scale - power,
+            dim=self.unit.dim,
+        )
+    
+    def materialise(self) -> PhysicalArrayLike:
         if self.unit.dim:
             raise ValueError(f"Cannot materialise unitful array with a non-zero unit: {self.unit}")
         factor = 10 ** self.unit.scale
@@ -48,10 +51,10 @@ class Unitful(TreeClass):
     def __repr__(self) -> str:
         return str(self)
     
-    def __mul__(self, other: ArrayLike | "Unitful") -> "Unitful":
+    def __mul__(self, other: PhysicalArrayLike | "Unitful") -> "Unitful":
         return multiply(self, other)
     
-    def __rmul__(self, other: ArrayLike | "Unitful") -> "Unitful":
+    def __rmul__(self, other: PhysicalArrayLike | "Unitful") -> "Unitful":
         return multiply(self, other)
     
     def __add__(self, other: "Unitful") -> "Unitful":
@@ -66,22 +69,22 @@ class Unitful(TreeClass):
     def __rsub__(self, other: "Unitful") -> "Unitful":
         return subtract(other, self)
     
-    def __lt__(self, other: "Unitful") -> ArrayLike:
+    def __lt__(self, other: "Unitful") -> PhysicalArrayLike:
         return lt(self, other)
     
-    def __le__(self, other: "Unitful") -> ArrayLike:
+    def __le__(self, other: "Unitful") -> PhysicalArrayLike:
         return le(self, other)
     
-    def __eq__(self, other: "Unitful") -> ArrayLike:
+    def __eq__(self, other: "Unitful") -> PhysicalArrayLike:
         return eq(self, other)
     
-    def __ne__(self, other: "Unitful") -> ArrayLike:  # type: ignore
[…]
