### 2025-08-01 04aa55b
automatic scale choosing in unitful
