### 2025-08-01 e9b6008
Improved typing for non-unitful overridden methods
