+++ src/fdtdx/units/unitful.py
@@ -125,7 +125,7 @@ def align_scales(
 
 ## Multiplication ###########################
 @overload
-def multiply(  # type: ignore
+def multiply(
     x: Unitful, 
     y: Unitful
 ) -> Unitful:
@@ -142,7 +142,7 @@ def multiply(  # type: ignore
     return Unitful(val=new_val, unit=Unit(scale=new_scale, dim=unit_dict))
 
 @overload
-def multiply(  # type: ignore
+def multiply(
     x: PhysicalArrayLike, 
     y: Unitful
 ) -> Unitful:
@@ -150,19 +150,23 @@ def multiply(  # type: ignore
     return Unitful(val=new_val, unit=y.unit)
 
 @overload
-def multiply(  # type: ignore
+def multiply(
     x: Unitful, 
     y: PhysicalArrayLike
 ) -> Unitful:
     return multiply(y, x)
 
 @overload
-def multiply(  # type: ignore
-    x: PhysicalArrayLike, 
-    y: PhysicalArrayLike
-) -> PhysicalArrayLike:
-    return x * y
+def multiply(x: int, y: int) -> int: return x * y
 
+@overload
+def multiply(x: float, y: float) -> float: return x * y
+
+@overload
+def multiply(x: complex, y: complex) -> complex: return x * y
+
+@overload
+def multiply(x: jax.Array, y: jax.Array) -> jax.Array: return x * y
 
 @dispatch
 def multiply(x, y):  # type: ignore
@@ -180,18 +184,23 @@ def _same_dim_binary_fn(x: Unitful, y: Unitful, op_str: str) -> Unitful:
     return Unitful(val=new_val, unit=x_align.unit)
 
 @overload
-def add(  # type: ignore
+def add(
     x: Unitful, 
     y: Unitful
 ) -> Unitful:
     return _same_dim_binary_fn(x, y, "add")
 
 @overload
-def add(  # type: ignore
-    x: PhysicalArrayLike, 
-    y: PhysicalArrayLike
-) -> PhysicalArrayLike:
-    return x + y
+def add(x: int, y: int) -> int: return x + y
+
+@overload
+def add(x: float, y: float) -> float: return x + y
+
+@overload
+def add(x: complex, y: complex) -> complex: return x + y
+
+@overload
+def add(x: jax.Array, y: jax.Array) -> jax.Array: return x + y
 
 @dispatch
 def add(x, y):  # type: ignore
@@ -200,18 +209,23 @@ def add(x, y):  # type: ignore
 
 ## Subtractions ###########################
 @overload
-def subtract(  # type: ignore
+def subtract(
     x: Unitful, 
     y: Unitful
 ) -> Unitful:
     return _same_dim_binary_fn(x, y, "subtract")
 
 @overload
-def subtract(  # type: ignore
-    x: PhysicalArrayLike, 
-    y: PhysicalArrayLike
-) -> PhysicalArrayLike:
-    return x - y
+def subtract(x: int, y: int) -> int: return x - y
+
+@overload
+def subtract(x: float, y: float) -> float: return x - y
+
+@overload
+def subtract(x: complex, y: complex) -> complex: return x - y
+
+@overload
+def subtract(x: jax.Array, y: jax.Array) -> jax.Array: return x - y
 
 @dispatch
 def subtract(x, y):  # type: ignore
@@ -220,18 +234,20 @@ def subtract(x, y):  # type: ignore
 
 ## Remainder ###########################
 @overload
-def remainder(  # type: ignore
+def remainder(
     x: Unitful, 
     y: Unitful
 ) -> Unitful:
[…]
