+++ tests/units/test_unitful.py
@@ -428,3 +428,187 @@ def test_equality_fractional_dimensions():
     assert u1.unit.dim == {SI.A: Fraction(7, 11)}
     assert u2.unit.dim == {SI.A: Fraction(7, 11)}
 
+
+
+# Division Test Cases
+def test_divide_unitful_unitful_same_dimensions():
+    """Test division of two Unitful objects with same dimensions"""
+    distance1 = Unit(scale=0, dim={SI.m: 1})
+    distance2 = Unit(scale=0, dim={SI.m: 1})
+    u1 = Unitful(val=jnp.array(10.0), unit=distance1)
+    u2 = Unitful(val=jnp.array(2.0), unit=distance2)
+    
+    result = u1 / u2
+    
+    assert jnp.allclose(result.value(), 5.0)
+    assert result.unit.dim == {}  # m^1 / m^1 = dimensionless
+
+
+def test_divide_unitful_unitful_different_dimensions():
+    """Test division of two Unitful objects with different dimensions"""
+    distance = Unit(scale=0, dim={SI.m: 1})
+    time = Unit(scale=0, dim={SI.s: 1})
+    u1 = Unitful(val=jnp.array(20.0), unit=distance)  # 20 meters
+    u2 = Unitful(val=jnp.array(4.0), unit=time)       # 4 seconds
+    
+    result: Unitful = jnp.divide(u1, u2)  # type: ignore
+    
+    assert jnp.allclose(result.value(), 5.0)  # 20/4 = 5
+    assert result.unit.dim == {SI.m: 1, SI.s: -1}  # m/s
+
+
+def test_divide_unitful_unitful_with_fractional_dimensions():
+    """Test division creating fractional dimensions"""
+    # Create m^2 (area) and m^(1/2) 
+    area_unit = Unit(scale=0, dim={SI.m: 2})
+    sqrt_unit = Unit(scale=0, dim={SI.m: Fraction(1, 2)})
+    u1 = Unitful(val=jnp.array(16.0), unit=area_unit)
+    u2 = Unitful(val=jnp.array(4.0), unit=sqrt_unit)
+    
+    # m^2 / m^(1/2) = m^(3/2)
+    result: Unitful = jnp.divide(u1, u2)  # type: ignore
+    
+    assert jnp.allclose(result.value(), 4.0)
+    assert result.unit.dim == {SI.m: Fraction(3, 2)}
+
+
+def test_divide_magic_method():
+    """Test division using magic method (__truediv__)"""
+    velocity_unit = Unit(scale=0, dim={SI.m: 1, SI.s: -1})
+    time_unit = Unit(scale=0, dim={SI.s: 1})
+    velocity = Unitful(val=jnp.array(30.0), unit=velocity_unit)  # 30 m/s
+    time = Unitful(val=jnp.array(6.0), unit=time_unit)          # 6 s
+    
+    # This would require implementing __truediv__ in the Unitful class
+    # result = velocity / time  # Should give acceleration (m/s^2)
+    result: Unitful = jnp.divide(velocity, time)  # type: ignore
+    
+    assert jnp.allclose(result.value(), 5.0)
+    assert result.unit.dim == {SI.m: 1, SI.s: -2}  # m/s^2 (acceleration)
+
+
+def test_len_scalar_unitful():
+    """Test __len__ method with scalar Unitful object"""
+    time_unit = Unit(scale=0, dim={SI.s: 1})
+    scalar_time = Unitful(val=5.0, unit=time_unit)
+    
+    result = len(scalar_time)
+    
+    assert result == 1
+
+
+def test_len_array_unitful():
+    """Test __len__ method with array Unitful object"""
+    time_unit = Unit(scale=0, dim={SI.s: 1})
[…]
