### 2025-08-04 9291511
Initial work on stopping conditions
