### 2025-08-04 1a36030
Hooked up stopping conditions to custom_fdtd_forward function
