### 2025-08-05 a5149fc
More tweaks, docstrings
