### 2025-08-11 20d6ec7
Removing stopping conditions from custom_fdtd_forward and adding them to checkpointed_fdtd
