### 2025-08-11 873d9ed
Adding validation detector shape and key presence checks, making shape of arrays output by each branch be the same
