+++ src/fdtdx/fdtd/stop_conditions.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from typing import Optional
 
 import jax
 import jax.numpy as jnp
@@ -15,6 +16,11 @@ class StoppingCondition(TreeClass, ABC):
     This class provides the interface for implementing custom termination criteria
     that can depend on simulation state, detector readings, field values, etc.
     """
+
+    @abstractmethod
+    def validate(self, arrays: ArrayContainer) -> None:
+        """Optional pre-run validation; override in subclasses."""
+        return
 
     @abstractmethod
     def __call__(self, state: SimulationState) -> jax.Array:
@@ -37,7 +43,10 @@ class TimeStepCondition(StoppingCondition):
     a specified end time is reached.
     """
 
-    end_time: int = frozen_field()
+    end_time_step: int = frozen_field()
+
+    def validate(self, arrays: ArrayContainer) -> None:
+        pass
 
     def __call__(self, state: SimulationState) -> jax.Array:
         """Check if simulation should continue based on time step.
@@ -49,18 +58,54 @@ class TimeStepCondition(StoppingCondition):
             jax.Array: Boolean scalar - True if current time < end_time, False otherwise
         """
         time_step, _ = state
-        return self.end_time > time_step
+        return self.end_time_step > time_step
 
 
 @autoinit
 class DetectorThresholdCondition(StoppingCondition):
-    """Stopping condition based on detector readings reaching a threshold.
-    """
+    """Stopping condition based on detector readings reaching a threshold."""
 
     detector_name: str = frozen_field()
+    state_key: Optional[str] = frozen_field(default=None)
     threshold: float = frozen_field()
-    end_time: int = frozen_field()
+    end_time_step: int = frozen_field()
     comparison: str = frozen_field(default="less_than")  # "less_than", "greater_than", "equal"
+    expected_per_step_shape: tuple[int, ...] | None = frozen_field(default=None)
+
+    def validate(self, arrays: ArrayContainer) -> None:
+        if self.detector_name not in arrays.detector_states:
+            available = tuple(arrays.detector_states.keys())
+            raise KeyError(f"Detector '{self.detector_name}' not found. Available detectors: {available}")
+        det_state: DetectorState = arrays.detector_states[self.detector_name]
+        if self.state_key is not None:
+            if self.state_key not in det_state:
+                available = tuple(det_state.keys())
+                raise KeyError(
+                    f"Key '{self.state_key}' not found in DetectorState('{self.detector_name}'). "
+                    f"Available keys: {available}"
+                )
+            readings = det_state[self.state_key]
+        else:
+            if len(det_state) == 0:
+                raise ValueError(f"DetectorState('{self.detector_name}') is empty.")
+            if len(det_state) > 1:
+                available = tuple(det_state.keys())
+                raise ValueError(
+                    f"DetectorState('{self.detector_name}') has multiple arrays {available}. "
+                    "Specify which one via 'state_key'."
+                )
+            readings = next(iter(det_state.values()))
+
+        if readings.ndim < 1:
+            raise ValueError(
+                f"DetectorState('{self.detector_name}') array must have a time axis; got ndim={readings.ndim}."
+            )
+        per_step_shape = tuple(readings.shape[1:])
+        if self.expected_per_step_shape is not None and per_step_shape != self.expected_per_step_shape:
+            raise ValueError(
+                f"Per-step shape mismatch for detector '{self.detector_name}'. "
+                f"Expected {self.expected_per_step_shape}, got {per_step_shape}."
+            )
 
     def __call__(self, state: SimulationState) -> jax.Array:
[…]
