### 2025-08-11 #8678 @fruzsinaagocs
:wave: @a-eghrari and @esf0, would any of you be willing to review this submission for JOSS? Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.
