### 2025-08-18 2f42223
fdtdx.Unitful: pos and abs magic method
