### 2025-08-18 1057441
fdtdx.Unitful: matmul magic method
