### 2025-08-19 2659278
fdtdx.Unitful: implemented pow
