### 2025-08-19 2a0a6da
Moved patching to functional module
