### 2025-08-19 276c1a5
functional: implemented unitful sqrt
