### 2025-08-19 244ce30
removed mu as variable character
