+++ src/fdtdx/units/unitful.py
@@ -1,12 +1,13 @@
 from dataclasses import dataclass
+from typing import Any, Callable, NamedTuple, Self
 import jax
 import jax.numpy as jnp
 
 from fdtdx.core.fraction import Fraction
-from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field
+from fdtdx.core.jax.pytrees import TreeClass, autoinit, field, frozen_field
 
 from plum import dispatch, overload
-from pytreeclass import tree_repr
+from pytreeclass import AtIndexer, BaseKey, tree_repr
 
 from fdtdx.typing import SI, PhysicalArrayLike
 from fdtdx.units.utils import best_scale, handle_different_scales
@@ -26,17 +27,116 @@ class Unit(TreeClass):
         return str(self)
 
 
+@dataclass(frozen=True)
+class UnitfulIndexer:
+    unitful: "Unitful"
+    where: Any | None = None
+
+    def __getitem__(self, where: Any) -> "UnitfulIndexer":
+        if self.where is not None:
+            raise Exception("Already called [] on Unitful.at! Double Indexing [][] is currently not supported")
+        return UnitfulIndexer(self.unitful, where)
+    
+    def get(self) -> "Unitful":
+        """Get the leaf values at the specified location."""
+        if self.where is None:
+            return self.unitful
+        if isinstance(self.unitful.val, int | float | complex):
+            raise Exception(f"Cannot index scalar value: {self.unitful}")
+        new_val = self.unitful.val[self.where]
+        return Unitful(val=new_val, unit=self.unitful.unit)
+    
+    def set(self, value: "Unitful") -> "Unitful":
+        """Set the leaf values at the specified location."""
+        if self.where is None:
+            raise Exception(f"Cannot update value if no where clause is given")
+        if isinstance(self.unitful.val, int | float | complex):
+            raise Exception(f"Cannot index scalar value: {self.unitful}")
+        if self.unitful.unit.dim != value.unit.dim:
+            raise Exception(
+                f"Cannot update array value with different unit: {self.unitful.unit.dim} != {value.unit.dim}"
+            )
+        align_self, align_other = align_scales(self.unitful, value)
+        align_self_arr = align_self.val
+        assert not isinstance(align_self_arr, int | complex | float)
+        new_val: jax.Array = align_self_arr.at[self.where].set(align_other.val)
+        return Unitful(val=new_val, unit=align_self.unit)
+    
+    def add(self, value: "Unitful") -> "Unitful":
+        """Set the leaf values at the specified location."""
+        if self.where is None:
+            raise Exception(f"Cannot update value if no where clause is given")
+        if isinstance(self.unitful.val, int | float | complex):
+            raise Exception(f"Cannot index scalar value: {self.unitful}")
+        if self.unitful.unit.dim != value.unit.dim:
+            raise Exception(
+                f"Cannot update array value with different unit: {self.unitful.unit.dim} != {value.unit.dim}"
+            )
+        align_self, align_other = align_scales(self.unitful, value)
+        align_self_arr = align_self.val
+        assert not isinstance(align_self_arr, int | complex | float)
+        new_val: jax.Array = align_self_arr.at[self.where].add(align_other.val)
+        return Unitful(val=new_val, unit=align_self.unit)
+    
+    def subtract(self, value: "Unitful") -> "Unitful":
+        """Set the leaf values at the specified location."""
+        if self.where is None:
+            raise Exception(f"Cannot update value if no where clause is given")
+        if isinstance(self.unitful.val, int | float | complex):
+            raise Exception(f"Cannot index scalar value: {self.unitful}")
[…]
