+++ tests/units/test_unitful.py
@@ -750,3 +750,295 @@ def test_pow_overload_jax_arrays():
     # Should return a regular JAX array, not a Unitful object
     assert isinstance(result, jax.Array)
     assert not isinstance(result, Unitful)
+
+
+def test_at_get_single_index():
+    """Test getting a single value using .at[].get()"""
+    distance_unit = Unit(scale=0, dim={SI.m: 1})
+    distances = Unitful(val=jnp.array([10.0, 20.0, 30.0, 40.0]), unit=distance_unit)
+    
+    result = distances.at[2].get()
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), 30.0)
+    assert result.unit.dim == {SI.m: 1}
+
+
+def test_at_get_slice():
+    """Test getting a slice using .at[].get()"""
+    time_unit = Unit(scale=-3, dim={SI.s: 1})  # milliseconds
+    times = Unitful(val=jnp.array([1.0, 2.0, 3.0, 4.0, 5.0]), unit=time_unit)
+    
+    result = times.at[1:4].get()
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), jnp.array([2.0e-3, 3.0e-3, 4.0e-3]))
+    assert result.unit.dim == {SI.s: 1}
+
+
+def test_at_set_single_index_same_units():
+    """Test setting a single value with same units using .at[].set()"""
+    mass_unit = Unit(scale=0, dim={SI.kg: 1})
+    masses = Unitful(val=jnp.array([1.0, 2.0, 3.0, 4.0]), unit=mass_unit)
+    new_mass = Unitful(val=jnp.array(99.0), unit=mass_unit)
+    
+    result = masses.at[2].set(new_mass)
+    
+    assert isinstance(result, Unitful)
+    expected = jnp.array([1.0, 2.0, 99.0, 4.0])
+    assert jnp.allclose(result.value(), expected)
+    assert result.unit.dim == {SI.kg: 1}
+
+
+def test_at_set_different_scales_same_dimensions():
+    """Test setting values with different scales but same dimensions"""
+    time_unit_s = Unit(scale=0, dim={SI.s: 1})    # seconds
+    time_unit_ms = Unit(scale=-3, dim={SI.s: 1})  # milliseconds
+    
+    times = Unitful(val=jnp.array([1.0, 2.0, 3.0]), unit=time_unit_s)
+    new_time = Unitful(val=jnp.array(5000.0), unit=time_unit_ms)  # 5000 ms = 5 s
+    
+    result = times.at[1].set(new_time)
+    
+    assert isinstance(result, Unitful)
+    expected = jnp.array([1.0, 5.0, 3.0])  # 5000 ms converted to 5 s
+    assert jnp.allclose(result.value(), expected)
+    assert result.unit.dim == {SI.s: 1}
+
+
+def test_at_add_single_index():
+    """Test adding to a single value using .at[].add()"""
+    force_unit = Unit(scale=0, dim={SI.kg: 1, SI.m: 1, SI.s: -2})  # Newtons
+    forces = Unitful(val=jnp.array([10.0, 20.0, 30.0]), unit=force_unit)
+    additional_force = Unitful(val=jnp.array(5.0), unit=force_unit)
+    
+    result = forces.at[1].add(additional_force)
+    
+    assert isinstance(result, Unitful)
+    expected = jnp.array([10.0, 25.0, 30.0])  # 20 + 5 = 25
+    assert jnp.allclose(result.value(), expected)
+    assert result.unit.dim == {SI.kg: 1, SI.m: 1, SI.s: -2}
+
+
+def test_at_add_with_scale_alignment():
+    """Test adding values that require scale alignment"""
[…]
