### 2025-08-19 4a77b11
Fix mutable objects in default argument values
