### 2025-08-20 200edc3
fdtdx.Unitful: implemented min, max, mean
