+++ tests/units/test_unitful.py
@@ -1042,3 +1042,221 @@ def test_at_get_preserves_complex_units():
         SI.s: Fraction(1, 2)
     }
 
+
+def test_min_magic_method():
+    """Test min magic method on Unitful objects"""
+    temperature_unit = Unit(scale=0, dim={SI.K: 1})
+    temperatures = Unitful(val=jnp.array([300.0, 250.0, 400.0, 275.0]), unit=temperature_unit)
+    
+    result = temperatures.min()
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), 250.0)
+    assert result.unit.dim == {SI.K: 1}
+
+
+def test_min_overload_unitful():
+    """Test min function with Unitful objects"""
+    # Create pressure values with scale
+    pressure_unit = Unit(scale=3, dim={SI.kg: 1, SI.m: -1, SI.s: -2})  # kilopascals
+    pressures = Unitful(val=jnp.array([101.0, 95.5, 110.2, 88.7, 105.3]), unit=pressure_unit)
+    
+    result = jnp.min(pressures)  # type: ignore
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), 88.7e3)  # 88.7 kPa = 88700 Pa
+    assert result.unit.dim == {SI.kg: 1, SI.m: -1, SI.s: -2}
+
+
+def test_min_overload_jax_arrays():
+    """Test min function with regular JAX arrays"""
+    array = jnp.array([15.0, 3.2, 42.1, 8.9, 25.6])
+    
+    result = jnp.min(array)
+    
+    assert jnp.allclose(result, 3.2)
+    assert isinstance(result, jax.Array)
+    assert not isinstance(result, Unitful)
+
+
+def test_min_with_axis_parameter():
+    """Test min method with axis parameter on 2D Unitful array"""
+    energy_unit = Unit(scale=-3, dim={SI.kg: 1, SI.m: 2, SI.s: -2})  # millijoules
+    energies = Unitful(val=jnp.array([[10.0, 20.0, 30.0], [5.0, 15.0, 25.0]]), unit=energy_unit)
+    
+    # Min along axis 0 (columns)
+    result = energies.min(axis=0)
+    
+    assert isinstance(result, Unitful)
+    expected_vals = jnp.array([5.0, 15.0, 25.0])
+    assert jnp.allclose(result.value(), expected_vals * 1e-3)  # Convert to joules
+    assert result.unit.dim == {SI.kg: 1, SI.m: 2, SI.s: -2}
+
+
+def test_min_with_fractional_dimensions():
+    """Test min with Unitful objects containing fractional dimensions"""
+    # Create a unit with fractional dimension: m^(3/2)
+    fractional_unit = Unit(scale=0, dim={SI.m: Fraction(3, 2)})
+    values = Unitful(val=jnp.array([8.0, 27.0, 64.0, 125.0]), unit=fractional_unit)
+    
+    result = values.min()
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), 8.0)
+    assert result.unit.dim == {SI.m: Fraction(3, 2)}
+
+
+def test_max_magic_method():
+    """Test max magic method on Unitful objects"""
+    current_unit = Unit(scale=-3, dim={SI.A: 1})  # milliamps
+    currents = Unitful(val=jnp.array([150.0, 200.0, 180.0, 250.0, 120.0]), unit=current_unit)
+    
+    result = currents.max()
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), 250.0e-3)  # 250 mA = 0.25 A
[…]
