### 2025-08-20 9f38b0e
Update docstrings to reflect new default value types
