+++ src/fdtdx/utils/plot_setup.py
@@ -32,13 +32,13 @@ def plot_setup(
     Args:
         config (SimulationConfig): Configuration object containing simulation parameters like resolution
         objects (ObjectContainer): Container holding all simulation objects to be plotted
-        exclude_object_list (list[SimulationObject], optional): List of objects to exclude from all plots
+        exclude_object_list (list[SimulationObject] | None, optional): List of objects to exclude from all plots
         filename (str | Path | None, optional): If provided, saves the plot to this file instead of displaying
         axs (Sequence[Any] | None, optional): Optional matplotlib axes to plot on. If None, creates new figure
         plot_legend (bool, optional): Whether to add a legend showing object names/types
-        exclude_xy_plane_object_list (list[SimulationObject], optional): Objects to exclude from XY plane plot
-        exclude_yz_plane_object_list (list[SimulationObject], optional): Objects to exclude from YZ plane plot
-        exclude_xz_plane_object_list (list[SimulationObject], optional): Objects to exclude from XZ plane plot
+        exclude_xy_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from XY plane plot
+        exclude_yz_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from YZ plane plot
+        exclude_xz_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from XZ plane plot
 
     Returns:
         Figure: The generated figure object
