### 2025-08-21 d5f7e3a
fdtdx.Unitful: implemented sum, shape, dtype, ndim, size
