+++ tests/units/test_unitful.py
@@ -1260,3 +1260,366 @@ def test_mean_preserves_all_unit_properties():
     assert jnp.allclose(result.value(), 20.0e-6)  # 20 microCoulombs
     assert result.unit.dim == {SI.A: 1, SI.s: 1}
 
+
+def test_sum_magic_method():
+    """Test sum magic method on Unitful objects"""
+    energy_unit = Unit(scale=0, dim={SI.kg: 1, SI.m: 2, SI.s: -2})  # Joules
+    energies = Unitful(val=jnp.array([10.0, 20.0, 30.0, 40.0]), unit=energy_unit)
+    
+    result = energies.sum()
+    
+    assert isinstance(result, Unitful)
+    assert jnp.allclose(result.value(), 100.0)  # 10+20+30+40 = 100
+    assert result.unit.dim == {SI.kg: 1, SI.m: 2, SI.s: -2}
+
+
+def test_sum_overload_with_axis_and_scale():
+    """Test sum function with axis parameter and unit scale conversion"""
+    # Create power values with kilowatt scale
+    power_unit = Unit(scale=3, dim={SI.kg: 1, SI.m: 2, SI.s: -3})  # kilowatts
+    powers = Unitful(val=jnp.array([
+        [1.0, 2.0, 3.0],
+        [4.0, 5.0, 6.0],
+        [7.0, 8.0, 9.0]
+    ]), unit=power_unit)
+    
+    # Sum along axis 1 (rows)
+    result = jnp.sum(powers, axis=1)  # type: ignore
+    
+    assert isinstance(result, Unitful)
+    expected_vals = jnp.array([6.0, 15.0, 24.0])  # Row sums
+    assert jnp.allclose(result.value(), expected_vals * 1e3)  # Convert to watts
+    assert result.unit.dim == {SI.kg: 1, SI.m: 2, SI.s: -3}
+    assert result.val.shape == (3,)  # type: ignore
+
+
+def test_sum_with_fractional_dimensions_and_keepdims():
+    """Test sum with complex fractional dimensions and keepdims parameter"""
+    # Create a unit with mixed fractional dimensions: m^(5/4) * kg^(-1/2) * s^(3/7)
+    complex_unit = Unit(scale=-2, dim={
+        SI.m: Fraction(5, 4), 
+        SI.kg: Fraction(-1, 2), 
+        SI.s: Fraction(3, 7)
+    })
+    values = Unitful(val=jnp.array([
+        [2.0, 4.0], 
+        [6.0, 8.0], 
+        [10.0, 12.0]
+    ]), unit=complex_unit)
+    
+    # Sum along axis 0 with keepdims=True
+    result = values.sum(axis=0, keepdims=True)
+    
+    assert isinstance(result, Unitful)
+    expected_vals = jnp.array([[18.0, 24.0]])  # Column sums: [2+6+10, 4+8+12]
+    assert jnp.allclose(result.value(), expected_vals * 1e-2)  # scale=-2 means *0.01
+    assert result.unit.dim == {
+        SI.m: Fraction(5, 4), 
+        SI.kg: Fraction(-1, 2), 
+        SI.s: Fraction(3, 7)
+    }
+    assert result.val.shape == (1, 2)  # type: ignore
+
+
+def test_shape_scalar_python_int():
+    """Test shape property with Python scalar integer"""
+    time_unit = Unit(scale=0, dim={SI.s: 1})
+    scalar_time = Unitful(val=5, unit=time_unit)
+    
+    assert scalar_time.shape == ()
+
+
+def test_shape_scalar_python_float():
+    """Test shape property with Python scalar float"""
[…]
