### 2025-08-21 3b456fd
pre-defined composite units: Unit and Unitful entries
