### 2025-08-22 edc7698
fdtdx.functional: implemented roll
