### 2025-08-22 7b5ac1a
SimulationConfig: attributes are unitfuls
