+++ tests/test_config.py
@@ -1,0 +1,152 @@
+import math
+import pytest
+import jax.numpy as jnp
+from fdtdx.core.fraction import Fraction
+from fdtdx.units.typing import SI
+from fdtdx.units.unitful import Unit, Unitful
+from fdtdx.units.composite import s, m, Hz
+from fdtdx.config import SimulationConfig
+from fdtdx import constants
+
+
+def test_simulation_config_time_step_duration_unitful():
+    """Test that time_step_duration returns correct Unitful with time dimensions"""
+    # Create simulation with 1 nanosecond resolution and 10 picoseconds simulation time
+    resolution = Unitful(val=1e-9, unit=Unit(scale=0, dim={SI.m: 1}))  # 1 nm
+    time = Unitful(val=10e-12, unit=Unit(scale=0, dim={SI.s: 1}))  # 10 ps
+    
+    config = SimulationConfig(
+        time=time,
+        resolution=resolution,
+        courant_factor=0.5
+    )
+    
+    time_step = config.time_step_duration
+    
+    assert isinstance(time_step, Unitful)
+    assert time_step.unit.dim == {SI.s: 1}  # Should have time dimension
+    
+    # Calculate expected value: courant_number * resolution / c
+    expected_courant = 0.5 / math.sqrt(3)  # courant_factor / sqrt(3)
+    expected_duration = expected_courant * 1e-9 / constants.c.value()
+    assert jnp.allclose(time_step.value(), expected_duration, rtol=1e-10)
+
+
+def test_simulation_config_max_travel_distance_unitful():
+    """Test that max_travel_distance returns correct Unitful with length dimensions"""
+    # Create simulation with 100 micrometers resolution and 1 femtosecond time
+    resolution = Unitful(val=100e-6, unit=Unit(scale=0, dim={SI.m: 1}))  # 100 μm
+    time = Unitful(val=1e-15, unit=Unit(scale=0, dim={SI.s: 1}))  # 1 fs
+    
+    config = SimulationConfig(
+        time=time,
+        resolution=resolution
+    )
+    
+    max_distance = config.max_travel_distance
+    
+    assert isinstance(max_distance, Unitful)
+    assert max_distance.unit.dim == {SI.m: 1}  # Should have length dimension
+    
+    # Expected: speed of light * time
+    expected_distance = constants.c.value() * 1e-15
+    assert jnp.allclose(max_distance.value(), expected_distance, rtol=1e-10)
+
+
+def test_simulation_config_unitful_with_different_scales():
+    """Test SimulationConfig with Unitful inputs having different unit scales"""
+    # Use millimeters for resolution (scale -3) and microseconds for time (scale -6)
+    resolution = Unitful(val=0.5, unit=Unit(scale=-3, dim={SI.m: 1}))  # 0.5 mm
+    time = Unitful(val=2.0, unit=Unit(scale=-6, dim={SI.s: 1}))  # 2 μs
+    
+    config = SimulationConfig(
+        time=time,
+        resolution=resolution,
+        courant_factor=0.8
+    )
+    
+    # Test time_step_duration preserves unit consistency
+    time_step = config.time_step_duration
+    assert isinstance(time_step, Unitful)
+    assert time_step.unit.dim == {SI.s: 1}
+    
+    # Test max_travel_distance preserves unit consistency  
+    max_distance = config.max_travel_distance
+    assert isinstance(max_distance, Unitful)
+    assert max_distance.unit.dim == {SI.m: 1}
+    
+    # Verify the calculations are dimensionally consistent
+    # time_step should be: (0.8/sqrt(3)) * 0.5e-3 / c
+    expected_courant = 0.8 / math.sqrt(3)
[…]

[…]
