### 2025-08-22 254a32e
SimulatioConfig: Updated docstring to Unitful
