### 2025-08-22 2f91221
WaveCharacter: attributes are unitfuls
