+++ tests/core/test_wave_character.py
@@ -1,181 +1,187 @@
-import math
-
 import pytest
+import jax.numpy as jnp
+from fdtdx.core.wavelength import WaveCharacter
+from fdtdx.units.unitful import Unit, Unitful
+from fdtdx.units.typing import SI
+from fdtdx.units.composite import Hz, s, m
+from fdtdx import constants
 
-from fdtdx import WaveCharacter, constants
+
+def test_wavecharacter_frequency_initialization_with_phase_shift():
+    """Test WaveCharacter initialization with frequency and custom phase shift"""
+    # Create a 1 GHz frequency with π/4 phase shift
+    frequency_val = 1e9 * Hz  # 1 GHz
+    phase_shift = jnp.pi / 4
+    x = 1.0 / frequency_val
+    
+    wave_char = WaveCharacter(frequency=frequency_val, phase_shift=phase_shift)
+    
+    # Check that frequency is preserved
+    assert jnp.allclose(wave_char.get_frequency().value(), 1e9)
+    assert wave_char.get_frequency().unit.dim == {SI.s: -1}
+    
+    # Check that phase shift is preserved
+    assert jnp.allclose(wave_char.phase_shift, jnp.pi / 4)
+    
+    # Check that period and wavelength are correctly calculated
+    expected_period = 1e-9  # 1/1e9 = 1 ns
+    assert jnp.allclose(wave_char.get_period().value(), expected_period)
+    assert wave_char.get_period().unit.dim == {SI.s: 1}
+    
+    # Check wavelength calculation: λ = c/f
+    expected_wavelength = constants.c.value() / 1e9  # ~0.3 m
+    assert jnp.allclose(wave_char.get_wavelength().value(), expected_wavelength, rtol=1e-10)
+    assert wave_char.get_wavelength().unit.dim == {SI.m: 1}
 
 
-class TestWaveCharacter:
-    """Test cases for WaveCharacter dataclass."""
+def test_wavecharacter_wavelength_initialization_infrared():
+    """Test WaveCharacter initialization with infrared wavelength"""
+    # Create a 1550 nm wavelength (common telecom wavelength)
+    wavelength_val = 1550e-9 * m  # 1550 nm
+    
+    wave_char = WaveCharacter(wavelength=wavelength_val)
+    
+    # Check that wavelength is preserved
+    assert jnp.allclose(wave_char.get_wavelength().value(), 1550e-9)
+    assert wave_char.get_wavelength().unit.dim == {SI.m: 1}
+    
+    # Check default phase shift
+    assert wave_char.phase_shift == 0.0
+    
+    # Check frequency calculation: f = c/λ
+    expected_frequency = constants.c.value() / (1550e-9)  # ~193.5 THz
+    assert jnp.allclose(wave_char.get_frequency().value(), expected_frequency, rtol=1e-10)
+    assert wave_char.get_frequency().unit.dim == {SI.s: -1}
+    
+    # Check period calculation: T = λ/c
+    expected_period = (1550e-9) / constants.c.value()  # ~5.17 fs
+    assert jnp.allclose(wave_char.get_period().value(), expected_period, rtol=1e-10)
+    assert wave_char.get_period().unit.dim == {SI.s: 1}
 
-    def test_initialization_with_period(self):
-        """Test initialization with period only."""
-        period = 1e-9  # 1 nanosecond
-        wave = WaveCharacter(period=period)
 
-        assert wave.get_period() == period
-        assert wave.get_wavelength() == period * constants.c
-        assert wave.get_frequency() == 1.0 / period
-        assert wave.phase_shift == 0.0
+def test_wavecharacter_period_initialization_microseconds():
[…]
