+++ src/fdtdx/units/composite.py
@@ -7,7 +7,6 @@ us_unit = Unit(scale=-6, dim={SI.s: 1})
 ns_unit = Unit(scale=-9, dim={SI.s: 1})
 ps_unit = Unit(scale=-12, dim={SI.s: 1})
 fs_unit = Unit(scale=-15, dim={SI.s: 1})
-
 s = Unitful(val=1, unit=s_unit)
 ms = Unitful(val=1, unit=ms_unit)
 us = Unitful(val=1, unit=us_unit)
@@ -21,7 +20,6 @@ MHz_unit = Unit(scale=6, dim={SI.s: -1})
 GHz_unit = Unit(scale=9, dim={SI.s: -1})
 THz_unit = Unit(scale=12, dim={SI.s: -1})
 PHz_unit = Unit(scale=15, dim={SI.s: -1})
-
 Hz = Unitful(val=1, unit=Hz_unit)
 kHz = Unitful(val=1, unit=kHz_unit)
 MHz = Unitful(val=1, unit=MHz_unit)
@@ -35,7 +33,6 @@ mm_unit = Unit(scale=-3, dim={SI.m: 1})
 um_unit = Unit(scale=-6, dim={SI.m: 1})
 nm_unit = Unit(scale=-9, dim={SI.m: 1})
 pm_unit = Unit(scale=-12, dim={SI.m: 1})
-
 km = Unitful(val=1, unit=km_unit)
 m = Unitful(val=1, unit=m_unit)
 mm = Unitful(val=1, unit=mm_unit)
@@ -43,4 +40,28 @@ um = Unitful(val=1, unit=um_unit)
 nm = Unitful(val=1, unit=nm_unit)
 pm = Unitful(val=1, unit=pm_unit)
 
-m_per_s = Unitful(val=1, unit=Unit(scale=0, dim={SI.s: -1, SI.m: 1}))
+MW_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3})
+kW_unit = Unit(scale=3, dim={SI.kg: 1, SI.m: 2, SI.s: -3})
+W_unit = Unit(scale=0, dim={SI.kg: 1, SI.m: 2, SI.s: -3})
+mW_unit = Unit(scale=-3, dim={SI.kg: 1, SI.m: 2, SI.s: -3})
+uW_unit = Unit(scale=-6, dim={SI.kg: 1, SI.m: 2, SI.s: -3})
+MW = Unitful(val=1, unit=MW_unit)
+kW = Unitful(val=1, unit=kW_unit)
+W = Unitful(val=1, unit=W_unit)
+mW = Unitful(val=1, unit=mW_unit)
+uW = Unitful(val=1, unit=uW_unit)
+
+kV_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+V_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+mV_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+uV_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+kV = Unitful(val=1, unit=kV_unit)
+V = Unitful(val=1, unit=V_unit)
+mV = Unitful(val=1, unit=mV_unit)
+uV = Unitful(val=1, unit=uV_unit)
+
+m_per_s_unit = Unit(scale=0, dim={SI.s: -1, SI.m: 1})
+m_per_s = m / s
+
+V_per_m_unit = Unit(scale=0, dim={SI.kg: 1, SI.m: 1, SI.s: -3, SI.A: -1})
+V_per_m = V / m
