+++ src/fdtdx/units/composite.py
@@ -51,14 +51,39 @@ W = Unitful(val=1, unit=W_unit)
 mW = Unitful(val=1, unit=mW_unit)
 uW = Unitful(val=1, unit=uW_unit)
 
-kV_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
-V_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
-mV_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
-uV_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+kV_unit = Unit(scale=3, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+V_unit = Unit(scale=0, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+mV_unit = Unit(scale=-3, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
+uV_unit = Unit(scale=-6, dim={SI.kg: 1, SI.m: 2, SI.s: -3, SI.A: -1})
 kV = Unitful(val=1, unit=kV_unit)
 V = Unitful(val=1, unit=V_unit)
 mV = Unitful(val=1, unit=mV_unit)
 uV = Unitful(val=1, unit=uV_unit)
+
+kA_unit = Unit(scale=3, dim={SI.A: 1})
+A_unit = Unit(scale=0, dim={SI.A: 1})
+mA_unit = Unit(scale=-3, dim={SI.A: 1})
+uA_unit = Unit(scale=-6, dim={SI.A: 1})
+kA = Unitful(val=1, unit=kA_unit)
+A = Unitful(val=1, unit=A_unit)
+mA = Unitful(val=1, unit=mA_unit)
+uA = Unitful(val=1, unit=uA_unit)
+
+F_unit = Unit(scale=0, dim={SI.s: 4, SI.A: 2, SI.m: -2, SI.kg: -2})
+mF_unit = Unit(scale=-3, dim={SI.s: 4, SI.A: 2, SI.m: -2, SI.kg: -2})
+uF_unit = Unit(scale=-6, dim={SI.s: 4, SI.A: 2, SI.m: -2, SI.kg: -2})
+nF_unit = Unit(scale=-9, dim={SI.s: 4, SI.A: 2, SI.m: -2, SI.kg: -2})
+pF_unit = Unit(scale=-12, dim={SI.s: 4, SI.A: 2, SI.m: -2, SI.kg: -2})
+fF_unit = Unit(scale=-15, dim={SI.s: 4, SI.A: 2, SI.m: -2, SI.kg: -2})
+F = Unitful(val=1, unit=F_unit)
+mF = Unitful(val=1, unit=mF_unit)
+uF = Unitful(val=1, unit=uF_unit)
+nF = Unitful(val=1, unit=nF_unit)
+pF = Unitful(val=1, unit=pF_unit)
+fF = Unitful(val=1, unit=fF_unit)
+
+N_unit = Unit(scale=0, dim={SI.kg: 1, SI.m: 1, SI.s: -2})
+N = Unitful(val=1, unit=N_unit)
 
 m_per_s_unit = Unit(scale=0, dim={SI.s: -1, SI.m: 1})
 m_per_s = m / s
