### 2025-08-22 e4704a0
functional: implemented square
