+++ src/fdtdx/functional/patching.py
@@ -24,6 +24,7 @@ from fdtdx.units.unitful import (
 from fdtdx.functional.numpy import (
     sqrt,
     roll,
+    square,
 )
 
 def patch_fn_to_module(
@@ -65,6 +66,7 @@ def patch_all_functions():
         (mean, None),
         (sum, None),
         (roll, None),
+        (square, None),
     ]
     for fn, orig in _full_patch_list_numpy:
         patch_fn_to_module(
