### 2025-08-28 3d321f6
Add comments to examples
