+++ examples/optimize_ceviche_corner.py
@@ -1,3 +1,4 @@
+# Imports: standard libraries and required packages
 import sys
 import time
 
@@ -14,17 +15,22 @@ def main(
     evaluation: bool,
     backward: bool,
 ):
+    # Log the random seed for reproducibility
     logger.info(f"{seed=}")
 
+    # Initialize experiment logger for tracking results and saving outputs
     exp_logger = fdtdx.Logger(
         experiment_name="ceviche_corner",
         name=None,
     )
+    # Create a JAX random key for stochastic operations
     key = jax.random.PRNGKey(seed=seed)
 
+    # Set simulation wavelength and compute corresponding period
     wavelength = 1.55e-6
     period = fdtdx.constants.wavelength_to_period(wavelength)
 
+    # Define simulation configuration (time, resolution, data type, etc.)
     config = fdtdx.SimulationConfig(
         time=50e-15,
         resolution=20e-9,
@@ -32,12 +38,14 @@ def main(
         courant_factor=0.99,
     )
 
+    # Calculate number of time steps per period and total time steps
     period_steps = round(period / config.time_step_duration)
     all_time_steps = list(range(config.time_steps_total))
     logger.info(f"{config.time_steps_total=}")
     logger.info(f"{period_steps=}")
     logger.info(f"{config.max_travel_distance=}")
 
+    # If not in evaluation mode or if backward simulation is requested, set up gradient recording
     if not evaluation or backward:
         gradient_config = fdtdx.GradientConfig(
             recorder=fdtdx.Recorder(
@@ -49,16 +57,20 @@ def main(
         )
         config = config.aset("gradient_config", gradient_config)
 
+    # List to hold placement constraints for simulation objects
     placement_constraints = []
 
+    # Define the simulation volume (physical size)
     volume = fdtdx.SimulationVolume(
         partial_real_shape=(2.7e-6, 2.7e-6, 1.5e-6),
     )
 
+    # Add boundary objects and constraints to the simulation
     bound_cfg = fdtdx.BoundaryConfig.from_uniform_bound(thickness=10)
     _, c_list = fdtdx.boundary_objects_from_config(bound_cfg, volume)
     placement_constraints.extend(c_list)
 
+    # Add substrate object (base layer) and its placement constraint
     substrate = fdtdx.UniformMaterialObject(
         partial_real_shape=(None, None, 0.5e-6),
         material=fdtdx.Material(permittivity=fdtdx.constants.relative_permittivity_silica),
@@ -73,6 +85,7 @@ def main(
         )
     )
 
+    # Define device geometry and material configuration
     height = 400e-9
     width = 400e-9
     material_config = {
@@ -85,6 +98,7 @@ def main(
         partial_real_shape=(1.6e-6, 1.6e-6, height),
         materials=material_config,
         param_transforms=[
+            # Optional symmetry and smoothing transforms
             # DiagonalSymmetry2D(min_min_to_max_max=False),
             fdtdx.GaussianSmoothing2D(std_discrete=3),
             fdtdx.SubpixelSmoothedProjection(),
@@ -92,6 +106,7 @@ def main(
         ],
         partial_voxel_real_shape=(voxel_size, voxel_size, height),
     )
+    # Place device relative to substrate with specified margins
     placement_constraints.append(
         device.place_relative_to(
             substrate,
@@ -103,6 +118,7 @@ def main(
         )
     )
 
+    # Add input waveguide and its placement constraints
     waveguide_in = fdtdx.UniformMaterialObject(
         partial_real_shape=(None, width, height),
         material=material_config["Silicon"],
@@ -123,6 +139,7 @@ def main(
         ]
     )
 
+    # Add source object and its placement constraint
     source = fdtdx.ModePlaneSource(
         partial_grid_shape=(1, None, None),
[…]
