### 2025-08-28 03de532
Composite Units: Added Joules, Ampere per meter
