### 2025-08-28 35a3a59
unitful adaptation of compute_energy, normalize_by_energy
