+++ src/fdtdx/core/physics/metrics.py
@@ -9,51 +9,66 @@ FDTD field array shapes.
 import jax
 import jax.numpy as jnp
 
+from fdtdx.units.composite import J, V_per_m_unit, A_per_m_unit
+from fdtdx.units.typing import SI
+from fdtdx.units.unitful import Unitful
+import fdtdx.functional as ff
+from fdtdx.constants import eps0, mu0
 
 def compute_energy(
-    E: jax.Array,
-    H: jax.Array,
+    E: Unitful,
+    H: Unitful,
     inv_permittivity: jax.Array | float,
     inv_permeability: jax.Array | float,
     axis: int = 0,
-) -> jax.Array:
+) -> Unitful:
     """Computes the total electromagnetic energy density of the field.
 
     Args:
-        E (jax.Array): Electric field array with shape (3, nx, ny, nz)
-        H (jax.Array): Magnetic field array with shape (3, nx, ny, nz)
+        E (Unitful): Electric field array with shape (3, nx, ny, nz)
+        H (Unitful): Magnetic field array with shape (3, nx, ny, nz)
         inv_permittivity (jax.Array | float): Inverse of the electric permittivity array
         inv_permeability (jax.Array | float): Inverse of the magnetic permeability array
         axis (int, optional): Axis index of the X,Y,Z component for the E and H field. Defaults to 0.
     Returns:
-        jax.Array: Total energy density array with shape (nx, ny, nz)
+        Unitful: Total energy density array with shape (nx, ny, nz)
     """
-    abs_E = jnp.sum(jnp.square(jnp.abs(E)), axis=axis)
-    energy_E = 0.5 * (1 / inv_permittivity) * abs_E
+    assert E.unit.dim == V_per_m_unit.dim
+    assert H.unit.dim == A_per_m_unit.dim
+    
+    patched_E = ff.square(ff.abs(E))
+    sum_fields_E = ff.sum(patched_E, axis=axis)
+    energy_E = 0.5 * eps0 * (1 / inv_permittivity) * sum_fields_E
+    assert isinstance(energy_E, Unitful)
 
-    abs_H = jnp.sum(jnp.square(jnp.abs(H)), axis=axis)
-    energy_H = 0.5 * (1 / inv_permeability) * abs_H
+    patched_H = ff.square(ff.abs(H))
+    sum_fields_H = ff.sum(patched_H, axis=axis)
+    energy_H = 0.5 * mu0 * (1 / inv_permeability) * sum_fields_H
+    assert isinstance(energy_H, Unitful)
 
     total_energy = energy_E + energy_H
     return total_energy
 
 
 def normalize_by_energy(
-    E: jax.Array,
-    H: jax.Array,
+    E: Unitful,
+    H: Unitful,
     inv_permittivity: jax.Array | float,
     inv_permeability: jax.Array | float,
-) -> tuple[jax.Array, jax.Array]:
+    resolution: Unitful,
+    normalization_target: Unitful = 1*J,
+) -> tuple[Unitful, Unitful]:
     """Normalizes electromagnetic fields by their total energy.
 
     Args:
-        E (jax.Array): Electric field array with shape (3, nx, ny, nz)
-        H (jax.Array): Magnetic field array with shape (3, nx, ny, nz)
+        E (Unitful): Electric field array with shape (3, nx, ny, nz)
+        H (Unitful): Magnetic field array with shape (3, nx, ny, nz)
         inv_permittivity (jax.Array | float): Inverse of the electric permittivity array
         inv_permeability (jax.Array | float): Inverse of the magnetic permeability array
+        resolution (Unitful): Spatial resolution of grid points in metre.
 
     Returns:
[…]
