+++ tests/core/physics/test_metrics.py
@@ -1,39 +1,44 @@
 """Test cases for electromagnetic field metrics and normalization utilities."""
 
+import jax
 import jax.numpy as jnp
 
+from fdtdx.config import SimulationConfig
 from fdtdx.core.physics.metrics import (
     compute_energy,
     compute_poynting_flux,
     normalize_by_energy,
     normalize_by_poynting_flux,
 )
+from fdtdx.units.composite import V, m, A, J, s
+from fdtdx.units.typing import SI
+from fdtdx.units.unitful import Unitful
+from fdtdx.constants import eps0, mu0
+import fdtdx.functional as ff
 
 
 def test_compute_energy_basic():
     """Test basic energy computation with simple field configuration."""
     # Create simple uniform fields
-    E = jnp.ones((3, 4, 4, 4))
-    H = jnp.ones((3, 4, 4, 4))
+    E = (V / m) * jnp.ones((3, 4, 4, 4)) * 1e8
+    H = (A / m) * jnp.ones((3, 4, 4, 4))
     inv_permittivity = 1.0
     inv_permeability = 1.0
 
     energy = compute_energy(E, H, inv_permittivity, inv_permeability)
 
-    # Energy should be 0.5 * (1/inv_eps) * |E|^2 + 0.5 * (1/inv_mu) * |H|^2
-    # For uniform fields: |E|^2 = 3, |H|^2 = 3
-    expected_energy = 0.5 * (1 / inv_permittivity) * 3 + 0.5 * (1 / inv_permeability) * 3
-    expected_energy = 3.0
+    expected_energy = 0.5 * eps0.value() * (1 / inv_permittivity) * 3 * 1e16 + 0.5 * mu0.value() * (1 / inv_permeability) * 3
 
     assert energy.shape == (4, 4, 4)
-    assert jnp.allclose(energy, expected_energy)
+    assert energy.unit.dim == {SI.kg: 1, SI.m: -1, SI.s: -2} # J / m^3
+    assert jnp.allclose(energy.value(), expected_energy)
 
 
 def test_compute_energy_complex_fields():
     """Test energy computation with complex fields."""
     # Create complex fields
-    E = jnp.array([[[1 + 1j]], [[2 + 0j]], [[0 + 2j]]])
-    H = jnp.array([[[1 - 1j]], [[1 + 1j]], [[2 + 0j]]])
+    E = (V / m) * jnp.array([[[1 + 1j]], [[2 + 0j]], [[0 + 2j]]])
+    H = (A / m) * jnp.array([[[1 - 1j]], [[1 + 1j]], [[2 + 0j]]])
     inv_permittivity = 2.0
     inv_permeability = 0.5
 
@@ -41,70 +46,39 @@ def test_compute_energy_complex_fields():
 
     # |E|^2 = |1+1j|^2 + |2+0j|^2 + |0+2j|^2 = 2 + 4 + 4 = 10
     # |H|^2 = |1-1j|^2 + |1+1j|^2 + |2+0j|^2 = 2 + 2 + 4 = 8
-    expected_energy = 0.5 * (1 / 2.0) * 10 + 0.5 * (1 / 0.5) * 8
-    expected_energy = 2.5 + 8.0
+    expected_energy = 0.5 * eps0.value() * (1 / 2.0) * 10 + mu0.value() * 0.5 * (1 / 0.5) * 8
 
-    assert jnp.allclose(energy, expected_energy)
-
-
-def test_compute_energy_spatially_varying_materials():
-    """Test energy computation with spatially varying material properties."""
-    E = jnp.ones((3, 2, 2, 2))
-    H = jnp.ones((3, 2, 2, 2))
-
-    # Create spatially varying permittivity and permeability
[…]
