+++ src/fdtdx/functional/patching.py
@@ -19,6 +19,7 @@ from fdtdx.units.unitful import (
     max,
     mean,
     sum,
+    abs,
 )
 
 from fdtdx.functional.numpy import (
@@ -67,6 +68,8 @@ def patch_all_functions():
         (sum, None),
         (roll, None),
         (square, None),
+        (abs, None),
+        (abs, "absolute")
     ]
     for fn, orig in _full_patch_list_numpy:
         patch_fn_to_module(
