### 2025-08-29 6e82ff7
Unitful: implemented jnp.conjugate
