### 2025-08-29 b486d11
physics.metrics: poynting vector with unitful
