### 2025-08-29 0e81fa5
fdtdx.functional: implemented jnp.dot
