### 2025-08-29 ffbba87
fdtdx.functional: implemented transpose
