### 2025-08-29 e647cc9
physics.metrics: implemented poynting flux with Unitful
