### 2025-08-29 68e5b03
fdtdx.Object: Added Unitful typing, but not functional yet
