+++ docs/api/public/SimulationVolume.md
@@ -1,5 +1,0 @@
-# fdtdx.SimulationVolume
-
-::: fdtdx.SimulationVolume
-    options:
-        heading: " "
