+++ docs/source/api.rst
@@ -1,0 +1,82 @@
+API
+===    
+
+.. autosummary::
+    :toctree: api/
+
+    fdtdx.Detector
+    fdtdx.export_stl
+    fdtdx.TreeClass
+    fdtdx.autoinit
+    fdtdx.field
+    fdtdx.private_field
+    fdtdx.frozen_field
+    fdtdx.frozen_private_field
+    fdtdx.metric_efficiency
+    fdtdx.compute_energy
+    fdtdx.normalize_by_energy
+    fdtdx.compute_poynting_flux
+    fdtdx.normalize_by_poynting_flux
+    fdtdx.OnOffSwitch
+    fdtdx.WaveCharacter
+    fdtdx.run_fdtd
+    fdtdx.place_objects
+    fdtdx.apply_params
+    fdtdx.full_backward
+    fdtdx.ArrayContainer
+    fdtdx.ParameterContainer
+    fdtdx.ObjectContainer
+    fdtdx.SimulationState
+    fdtdx.Recorder
+    fdtdx.RecordingState
+    fdtdx.DtypeConversion
+    fdtdx.LinearReconstructEveryK
+    fdtdx.SimulationObject
+    fdtdx.PositionConstraint
+    fdtdx.SizeConstraint
+    fdtdx.SizeExtensionConstraint
+    fdtdx.GridCoordinateConstraint
+    fdtdx.RealCoordinateConstraint
+    fdtdx.PerfectlyMatchedLayer
+    fdtdx.PeriodicBoundary
+    fdtdx.BoundaryConfig
+    fdtdx.boundary_objects_from_config
+    fdtdx.EnergyDetector
+    fdtdx.PoyntingFluxDetector
+    fdtdx.FieldDetector
+    fdtdx.PhasorDetector
+    fdtdx.ModeOverlapDetector
+    fdtdx.Device
+    fdtdx.StandardToInversePermittivityRange
+    fdtdx.StandardToCustomRange
+    fdtdx.StandardToPlusOneMinusOneRange
+    fdtdx.GaussianSmoothing2D
+    fdtdx.RemoveFloatingMaterial
+    fdtdx.ConnectHolesAndStructures
+    fdtdx.BinaryMedianFilterModule
+    fdtdx.ClosestIndex
+    fdtdx.BrushConstraint2D
+    fdtdx.PillarDiscretization
+    fdtdx.TanhProjection
+    fdtdx.SubpixelSmoothedProjection
+    fdtdx.DiagonalSymmetry2D
+    fdtdx.ParameterTransformation
+    fdtdx.circular_brush
+    fdtdx.BOTTOM_Z_PADDING_CONFIG_REPEAT
+    fdtdx.GaussianPlaneSource
+    fdtdx.UniformPlaneSource
+    fdtdx.ModePlaneSource
+    fdtdx.SingleFrequencyProfile
+    fdtdx.GaussianPulseProfile
+    fdtdx.Cylinder
+    fdtdx.Sphere
+    fdtdx.ExtrudedPolygon
+    fdtdx.UniformMaterialObject
+    fdtdx.SimulationVolume
+    fdtdx.Logger
+    fdtdx.plot_setup
+    fdtdx.SimulationConfig
+    fdtdx.GradientConfig
+    fdtdx.wavelength_to_period
+    fdtdx.Material
+
