+++ docs/api/public/EnergyDetector.md
@@ -1,5 +1,0 @@
-# fdtdx.EnergyDetector
-
-::: fdtdx.EnergyDetector
-    options:
-        heading: " "
