+++ docs/api/public/Recorder.md
@@ -1,5 +1,0 @@
-# fdtdx.Recorder
-
-::: fdtdx.Recorder
-    options:
-        heading: " "
