+++ docs/api/public/PositionConstraint.md
@@ -1,5 +1,0 @@
-# fdtdx.PositionConstraint
-
-::: fdtdx.PositionConstraint
-    options:
-        heading: " "
