+++ docs/api/public/ClosestIndex.md
@@ -1,5 +1,0 @@
-# fdtdx.ClosestIndex
-
-::: fdtdx.ClosestIndex
-    options:
-        heading: " "

[…]
