### 2025-09-01 ffedf15
object.py: minor formatting changes
