+++ pyproject.toml
@@ -54,10 +54,10 @@ dev = [
     "ruff>=0.8.2",
     "tox-uv>=1.16.1",
     "pydoclint>=0.6.6",
-    "sphinx>=8.1.3",
+    "sphinx>=8.0",
     "sphinx-autobuild>=2025.8.25",
     "sphinx-book-theme>=1.1.4",
-    "nbsphinx>=0.9.7",
+    "nbsphinx>=0.9",
     "sphinx-autodoc-typehints>=3.2.0",
 ]
 cuda12 = [
