### 2025-09-01 fb2c45f
sync_notebooks.sh: Updated notebooks path
