### 2025-09-01 7c38be8
CI/CD: updated pipeline for sphinx
