### 2025-09-01 db22f8b
docs/conf.py: Updated version
