### 2025-09-01 184566a
pyproject.toml: added myst_nb dependency in dev group
