+++ pyproject.toml
@@ -59,6 +59,7 @@ dev = [
     "sphinx-book-theme>=1.1.4",
     "nbsphinx>=0.9",
     "sphinx-autodoc-typehints>=3.2.0",
+    "myst_nb>=1.3.0",
 ]
 cuda12 = [
     "jax[cuda12]>=0.4.36",
