### 2025-09-01 cf6283c
Added .readthedocs.yaml
