### 2025-09-01 1710574
Added .readthedocs.yaml
