### 2025-09-01 62cf524
.readthedocs.yaml minor path fix
