+++ .readthedocs.yaml
@@ -1,7 +1,7 @@
 version: 2
 formats: [htmlzip]
 sphinx:
-   configuration: docs/conf.py
+   configuration: docs/source/conf.py
 python:
    install:
      - method: pip
