### 2025-09-01 d5d0b1f
physics/metrics.py: Normalize_poynting flux with Unitful
