### 2025-09-01 3ae99dd
functional: implemented pad and stack with Unitful
