+++ tests/functional/test_numpy.py
@@ -442,3 +442,162 @@ def test_transpose_jax_array():
     assert not isinstance(result, Unitful)
     # Shape should be transposed: (4, 3) -> (3, 4)
     assert result.shape == (3, 4)
+    
+
+def test_pad_unitful_temperature_field():
+    """Test pad with Unitful object containing a temperature field with fractional dimensions"""
+    # Create temperature unit: Kelvin with fractional dimension and scale=-3 (millikelvin)
+    temp_unit = Unit(scale=-3, dim={SI.K: 1, SI.m: Fraction(-1, 2)})  # K/sqrt(m)
+    # 2x3 temperature field array
+    temp_field = Unitful(val=jnp.array([
+        [100.0, 200.0, 300.0],
+        [400.0, 500.0, 600.0]
+    ]), unit=temp_unit)
+    
+    # Pad with constant value and symmetric padding
+    result: Unitful = jnp.pad(temp_field, pad_width=((1, 1), (2, 1)), mode='constant', constant_values=0.0)  # type: ignore
+    
+    assert isinstance(result, Unitful)
+    # Expected shape: original (2, 3) + padding ((1,1), (2,1)) = (4, 6)
+    assert result.shape == (4, 6)
+    # Check padded values (should be 0 with proper scale factor)
+    assert jnp.allclose(result.value()[0, :], 0.0)  # type: ignore First row should be zeros
+    assert jnp.allclose(result.value()[-1, :], 0.0)  # type: ignore Last row should be zeros
+    assert jnp.allclose(result.value()[:, :2], 0.0)  # type: ignore First two columns should be zeros
+    assert jnp.allclose(result.value()[:, -1], 0.0)  # type: ignore Last column should be zeros
+    # Check original values are preserved in the center with scale factor 10^(-3)
+    assert jnp.allclose(result.value()[1, 2], 100.0 * 1e-3)  # type: ignore Original [0,0] -> [1,2]
+    assert jnp.allclose(result.value()[2, 4], 600.0 * 1e-3)  # type: ignore Original [1,2] -> [2,4]
+    # Units should be preserved
+    assert result.unit.dim == {SI.K: 1, SI.m: Fraction(-1, 2)}
+
+
+def test_pad_jax_array():
+    """Test pad with regular JAX array (non-Unitful) using edge mode"""
+    # Create a 3x2 matrix
+    matrix = jnp.array([
+        [1.0, 2.0],
+        [3.0, 4.0],
+        [5.0, 6.0]
+    ])
+    
+    # Pad with edge values (replicate border values)
+    result = jnp.pad(matrix, pad_width=((1, 2), (1, 1)), mode='edge')  # type: ignore
+    
+    # Expected result: 3x2 -> 6x4 with edge padding
+    # Top padding: replicate first row
+    # Bottom padding: replicate last row  
+    # Left padding: replicate first column
+    # Right padding: replicate last column
+    expected = jnp.array([
+        [1.0, 1.0, 2.0, 2.0],  # Top padding: replicate [1, 2]
+        [1.0, 1.0, 2.0, 2.0],  # Original row 0 with side padding
+        [3.0, 3.0, 4.0, 4.0],  # Original row 1 with side padding
+        [5.0, 5.0, 6.0, 6.0],  # Original row 2 with side padding
+        [5.0, 5.0, 6.0, 6.0],  # Bottom padding: replicate [5, 6]
+        [5.0, 5.0, 6.0, 6.0]   # Bottom padding: replicate [5, 6]
+    ])
+    
+    assert jnp.allclose(result, expected)
+    # Should return a regular JAX array, not a Unitful object
[…]
