### 2025-09-01 3619182
functional: implemented jnp.isfinite
