### 2025-09-01 6ad80e2
functional: implemented jnp.roll
