### 2025-09-01 a943e1b
physics/curl: interpolate_fields, curl_E, curl_H with Unitful
