+++ src/fdtdx/core/physics/curl.py
@@ -1,12 +1,14 @@
 import jax
 import jax.numpy as jnp
+import fdtdx.functional as ff
+from fdtdx.units.unitful import Unitful
 
 
 def interpolate_fields(
-    E_field: jax.Array,
-    H_field: jax.Array,
+    E_field: Unitful,
+    H_field: Unitful,
     periodic_axes: tuple[bool, bool, bool] = (False, False, False),
-) -> tuple[jax.Array, jax.Array]:
+) -> tuple[Unitful, Unitful]:
     """Interpolates E and H fields onto E_z in a FDTD grid with PEC/periodic boundary conditions.
 
     Performs spatial interpolation of the electric and magnetic fields to align them
@@ -14,15 +16,15 @@ def interpolate_fields(
     naturally staggered in the Yee grid.
 
     Args:
-        E_field (jax.Array): 4D tensor representing the electric field.
+        E_field (Unitful): 4D tensor representing the electric field.
                 Dimensions are (width, depth, height, direction).
-        H_field (jax.Array): 4D tensor representing the magnetic field.
+        H_field (Unitful): 4D tensor representing the magnetic field.
                 Dimensions are (width, depth, height, direction).
         periodic_axes (tuple[bool, bool, bool], optional): Tuple of booleans indicating which axes use periodic
             boundaries (periodic_x, periodic_y, periodic_z). Defaults to (False, False, False).
 
     Returns:
-        tuple[jax.Array, jax.Array]: A tuple (E_interp, H_interp) containing:
+        tuple[Unitful, Unitful]: A tuple (E_interp, H_interp) containing:
             - E_interp: Interpolated electric field as 4D tensor
             - H_interp: Interpolated magnetic field as 4D tensor
 
@@ -40,8 +42,8 @@ def interpolate_fields(
             pad_width = ((0, 0), (0, 0), (1, 1), (0, 0))
         else:  # i == 2
             pad_width = ((0, 0), (0, 0), (0, 0), (1, 1))
-        E_field = jnp.pad(E_field, pad_width, mode=pad_mode)
-        H_field = jnp.pad(H_field, pad_width, mode=pad_mode)
+        E_field = ff.pad(E_field, pad_width, mode=pad_mode)
+        H_field = ff.pad(H_field, pad_width, mode=pad_mode)
 
     E_x, E_y, E_z = E_field[0], E_field[1], E_field[2]
     H_x, H_y, H_z = H_field[0], H_field[1], H_field[2]
@@ -64,13 +66,17 @@ def interpolate_fields(
     ) / 8.0
 
     # Constructing the interpolated fields
-    E_interp = jnp.stack([E_x, E_y, E_z], axis=0)
-    H_interp = jnp.stack([H_x, H_y, H_z], axis=0)
+    E_interp = ff.stack([E_x, E_y, E_z], axis=0)
+    H_interp = ff.stack([H_x, H_y, H_z], axis=0)
 
     return E_interp, H_interp
 
 
-def curl_E(E: jax.Array, periodic_axes: tuple[bool, bool, bool] = (False, False, False)) -> jax.Array:
+def curl_E(
+    E: Unitful,
+    resolution: Unitful,
+    periodic_axes: tuple[bool, bool, bool] = (False, False, False),
+) -> Unitful:
     """Transforms an E-type field into an H-type field by performing a curl operation.
 
     Computes the discrete curl of the electric field to obtain the corresponding
[…]
