+++ tests/core/physics/test_curl.py
@@ -1,13 +1,14 @@
 import jax.numpy as jnp
-
+import fdtdx.functional as ff
 from fdtdx.core.physics.curl import curl_E, curl_H, interpolate_fields
+from fdtdx.units.composite import V, m, A
 
 
 def test_interpolate_fields_basic():
     """Test basic interpolation functionality with simple field configuration."""
     # Create simple test fields
-    E_field = jnp.ones((3, 5, 5, 5))
-    H_field = jnp.ones((3, 5, 5, 5)) * 0.5
+    E_field = (V / m) * jnp.ones((3, 5, 5, 5))
+    H_field = (A / m) * jnp.ones((3, 5, 5, 5)) * 0.5
 
     E_interp, H_interp = interpolate_fields(E_field, H_field)
 
@@ -16,9 +17,9 @@ def test_interpolate_fields_basic():
     assert H_interp.shape == (3, 5, 5, 5)
 
     # Check that interpolated values are reasonable
-    assert jnp.all(E_interp[2] == 1.0)  # E_z should remain unchanged
-    assert jnp.all(jnp.isfinite(E_interp))
-    assert jnp.all(jnp.isfinite(H_interp))
+    assert jnp.all(E_interp[2].value() == 1.0)  # E_z should remain unchanged
+    assert jnp.all(ff.isfinite(E_interp))
+    assert jnp.all(ff.isfinite(H_interp))
 
 
 def test_interpolate_fields_periodic_boundaries():
@@ -29,53 +30,54 @@ def test_interpolate_fields_periodic_boundaries():
     z = jnp.linspace(0, 2 * jnp.pi, 6)
     X, Y, Z = jnp.meshgrid(x, y, z, indexing="ij")
 
-    E_field = jnp.stack([jnp.sin(X), jnp.cos(Y), jnp.sin(Z)], axis=0)
-    H_field = jnp.stack([jnp.cos(X), jnp.sin(Y), jnp.cos(Z)], axis=0)
+    E_field = (V / m) * jnp.stack([jnp.sin(X), jnp.cos(Y), jnp.sin(Z)], axis=0)
+    H_field = (A / m) * jnp.stack([jnp.cos(X), jnp.sin(Y), jnp.cos(Z)], axis=0)
 
     # Test with all periodic boundaries
     E_interp, H_interp = interpolate_fields(E_field, H_field, periodic_axes=(True, True, True))
 
     assert E_interp.shape == (3, 6, 6, 6)
     assert H_interp.shape == (3, 6, 6, 6)
-    assert jnp.all(jnp.isfinite(E_interp))
-    assert jnp.all(jnp.isfinite(H_interp))
+    assert jnp.all(ff.isfinite(E_interp))
+    assert jnp.all(ff.isfinite(H_interp))
 
 
 def test_interpolate_fields_mixed_boundaries():
     """Test interpolation with mixed periodic and PEC boundary conditions."""
-    E_field = jnp.ones((3, 6, 6, 6))
-    H_field = jnp.ones((3, 6, 6, 6)) * 2.0
+    E_field = (V / m) * jnp.ones((3, 6, 6, 6))
+    H_field = (A / m) * jnp.ones((3, 6, 6, 6)) * 2.0
 
     # Test with periodic x, PEC y,z
     E_interp, H_interp = interpolate_fields(E_field, H_field, periodic_axes=(True, False, False))
 
     assert E_interp.shape == (3, 6, 6, 6)
     assert H_interp.shape == (3, 6, 6, 6)
-    assert jnp.all(jnp.isfinite(E_interp))
-    assert jnp.all(jnp.isfinite(H_interp))
+    assert jnp.all(ff.isfinite(E_interp))
[…]
