### 2025-09-02 df88f03
Fleshed out docstrings
