### 2025-09-02 ddb9556
functional: implemented jnp.real, jnp.imag
