### 2025-09-02 9da2c9b
physics/modes.py: Unitful mode computation
