### 2025-09-02 #8678 @fruzsinaagocs
> Hi [@fruzsinaagocs](https://github.com/fruzsinaagocs), it seems that the other potential reviewers are not responding. Do you want to suggest and ping other potential candidates?

Hi @ymahlau,
We usually give at least a week for potential reviewers to respond before we ping another set.
If the review / pre-review of the paper is stalled, the handling editor will ping me here.
