### 2025-09-03 3b080ab
adjusted main fdtd functionality for unitful
