### 2025-09-03 b19f29f
paper.md: more acknowledgements
