### 2025-09-03 328ab26
paper.md: fixed typo
