### 2025-09-03 46b704c
Removed DetectorThresholdCondition, worked more on FieldConvergenceCondition
