+++ src/fdtdx/fdtd/stop_conditions.py
@@ -66,127 +66,6 @@ class TimeStepCondition(StoppingCondition):
 
 
 @autoinit
-class DetectorThresholdCondition(StoppingCondition):
-    """Stopping condition based on detector readings reaching a threshold.
-
-    Attributes:
-        detector_name (str): Unique identifier for the detector. Used for
-            logging and reference.
-        state_key (Optional[str]): Name of the state variable to monitor.
-            Defaults to None. If None, the detector operates without
-            referencing a specific state.
-        threshold (float): The numerical threshold against which values are
-            compared.
-        end_time_step (int): The last time step (inclusive) up to which
-            simulation progresses if the threshold is not reached.
-            Can be found at SimulationConfig.time_step_duration
-        comparison (str): The comparison operation used to evaluate values
-            against the threshold. Must be one of: `"less_than"`,
-            `"greater_than"`, or `"equal"`. Defaults to `"less_than"`.
-        expected_per_step_shape (tuple[int, ...] | None): The expected shape of
-            the per-step input values. If None, no shape check is enforced.
-            Defaults to None.
-    """
-
-    detector_name: str = frozen_field()
-    state_key: Optional[str] = frozen_field(default=None)
-    threshold: float = frozen_field()
-    end_time_step: int = frozen_field()
-    comparison: str = frozen_field(default="less_than")  # "less_than", "greater_than", "equal"
-    expected_per_step_shape: tuple[int, ...] | None = frozen_field(default=None)
-
-    def validate(self, arrays: ArrayContainer) -> None:
-        if self.detector_name not in arrays.detector_states:
-            available = tuple(arrays.detector_states.keys())
-            raise KeyError(f"Detector '{self.detector_name}' not found. Available detectors: {available}")
-        det_state: DetectorState = arrays.detector_states[self.detector_name]
-        if self.state_key is not None:
-            if self.state_key not in det_state:
-                available = tuple(det_state.keys())
-                raise KeyError(
-                    f"Key '{self.state_key}' not found in DetectorState('{self.detector_name}'). "
-                    f"Available keys: {available}"
-                )
-            readings = det_state[self.state_key]
-        else:
-            if len(det_state) == 0:
-                raise ValueError(f"DetectorState('{self.detector_name}') is empty.")
-            if len(det_state) > 1:
-                available = tuple(det_state.keys())
-                raise ValueError(
-                    f"DetectorState('{self.detector_name}') has multiple arrays {available}. "
-                    "Specify which one via 'state_key'."
-                )
-            readings = next(iter(det_state.values()))
-
-        if readings.ndim < 1:
-            raise ValueError(
-                f"DetectorState('{self.detector_name}') array must have a time axis; got ndim={readings.ndim}."
-            )
-        per_step_shape = tuple(readings.shape[1:])
-        if self.expected_per_step_shape is not None and per_step_shape != self.expected_per_step_shape:
-            raise ValueError(
-                f"Per-step shape mismatch for detector '{self.detector_name}'. "
-                f"Expected {self.expected_per_step_shape}, got {per_step_shape}."
-            )
-
-    def __call__(self, state: SimulationState) -> jax.Array:
-        """Check if simulation should continue based on detector reading.
-
-        Args:
-            state: Current simulation state (SimulationState)
-
-        Returns:
-            jax.Array: Boolean scalar - True if condition not met and time < end_time
-        """
-        time_step, arrays = state
-        time_condition = time_step < self.end_time_step
-        detector_condition = self._check_detector_condition(arrays)
[…]
