### 2025-09-03 8b74c66
Renamed condition to clarify, specified what it intakes in docstring
