### 2025-09-03 c95b23f
Adding ObjectContainer to parameters ingested by all conditions, adding logic to verify if energy detector is valid
