### 2025-09-03 83d4901
Passing objects to cond_fun
