### 2025-09-03 7d1101f
Fleshed out EnergyConvergenceCondition
