### 2025-09-04 8cf5204
Raising NotImplementedError instead of passing
