### 2025-09-06 e776e42
adding test_fdtd for fdtd module
