+++ tests/fdtd/test_fdtd.py
@@ -1,0 +1,123 @@
+import jax
+import jax.numpy as jnp
+import pytest
+
+from fdtdx.fdtd.fdtd import reversible_fdtd, checkpointed_fdtd, custom_fdtd_forward
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
+from fdtdx.config import SimulationConfig
+from fdtdx.objects.object import SimulationObject
+
+
+class DummySimulationObject(SimulationObject):
+    """Minimal mock implementation of a SimulationObject."""
+    def __init__(self, name="dummy"):
+        self.name = name
+
+    def update(self, *args, **kwargs):
+        return args, kwargs
+
+
+@pytest.fixture
+def dummy_arrays():
+    field_shape = (3, 2, 2, 2)   # (components, nx, ny, nz)
+    mat_shape   = (2, 2, 2)      # scalar per voxel
+
+    return ArrayContainer(
+        E=jnp.zeros(field_shape),
+        H=jnp.zeros(field_shape),
+        inv_permittivities=jnp.ones(mat_shape),
+        inv_permeabilities=jnp.ones(mat_shape),
+        boundary_states={},
+        detector_states={},
+        recording_state=None,
+        electric_conductivity=None,
+        magnetic_conductivity=None,
+    )
+
+
+@pytest.fixture
+def dummy_objects():
+    # Include one mock SimulationObject in the object list
+    return ObjectContainer(object_list=[DummySimulationObject()], volume_idx=0)
+
+
+@pytest.fixture
+def empty_objects():
+    return ObjectContainer(object_list=[], volume_idx=0)
+
+
+@pytest.fixture
+def dummy_config():
+    return SimulationConfig(
+        time=1.0,
+        resolution=1.0,
+        backend="gpu",
+        dtype=jnp.float32,
+        courant_factor=0.99,
+        gradient_config=None,
+    )
+
+
+def test_reversible_fdtd_runs(dummy_arrays, dummy_objects, dummy_config):
+    key = jax.random.PRNGKey(0)
+    t, arrs = reversible_fdtd(dummy_arrays, dummy_objects, dummy_config, key)
+
+    assert isinstance(t, jax.Array)
+    assert isinstance(arrs, ArrayContainer)
+    assert arrs.E.shape == dummy_arrays.E.shape
+
+
+def test_checkpointed_fdtd_runs(dummy_arrays, dummy_objects, dummy_config):
+    key = jax.random.PRNGKey(0)
+    t, arrs = checkpointed_fdtd(dummy_arrays, dummy_objects, dummy_config, key)
+
+    assert isinstance(t, jax.Array)
+    assert isinstance(arrs, ArrayContainer)
+
+
+def test_custom_fdtd_forward_runs(dummy_arrays, dummy_objects, dummy_config):
+    key = jax.random.PRNGKey(0)
+    t, arrs = custom_fdtd_forward(
+        dummy_arrays, dummy_objects, dummy_config, key,
+        reset_container=True, record_detectors=False,
+        start_time=0, end_time=1
+    )
+
+    assert isinstance(t, jax.Array)
+    assert isinstance(arrs, ArrayContainer)
+
+
+# --- Edge case tests ---
+
+def test_zero_time(dummy_arrays, dummy_objects):
+    config = SimulationConfig(
+        time=0.0,
+        resolution=1.0,
+        backend="gpu",
+        dtype=jnp.float32,
+        courant_factor=0.99,
+        gradient_config=None,
+    )
+    key = jax.random.PRNGKey(0)
+    t, arrs = reversible_fdtd(dummy_arrays, dummy_objects, config, key)
+    assert int(t) == 0
[…]
