### 2025-09-06 758bda7
adding test_fdtd for fdtd module
