### 2025-09-07 166c39d
Added objects param to docstrings, removed unnecessary comments
