### 2025-09-07 16330ab
Initial version of tests on stopping conditions
