+++ tests/fdtd/test_condition.py
@@ -7,14 +7,12 @@ import pytest
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 
-# Import the module to test
-from fdtdx.fdtd.forward import forward, forward_single_args_wrapper
 from fdtdx.interfaces.state import RecordingState
 from fdtdx.objects.boundaries.boundary import BaseBoundaryState
 from fdtdx.objects.detectors.detector import DetectorState
 
 # Import the module to test
-from fdtdx.fdtd.stop_conditions import StoppingCondition, TimeStepCondition, EnergyConvergenceCondition
+from fdtdx.fdtd.stop_conditions import TimeStepCondition, EnergyThresholdCondition, DetectorConvergenceCondition
 
 
 class TestCondition:
@@ -58,22 +56,153 @@ class TestCondition:
 
     def test_condition_function_signature(self, setup_simulation_state):
         """Test that stopping condition function has the correct signature and returns expected types."""
+        sim_data = setup_simulation_state
+        state = sim_data["state"]
+        objects = sim_data["objects"]
+        arrays = sim_data["arrays"]
+        end_step = 100
+
+        # 1. Test TimeStepCondition
+        time_step_cond = TimeStepCondition(end_step=end_step)
+        ts_result = time_step_cond(state, objects)
+        
+        assert isinstance(ts_result, jax.Array)
+        assert ts_result.dtype == jnp.bool_
+        assert ts_result.shape == ()
+
+        # 2. Test EnergyThresholdCondition
+        energy_threshold_cond = EnergyThresholdCondition(
+            threshold=1e-6,
+            end_step=end_step,
+            min_steps=10,
+        )
+        et_result = energy_threshold_cond(state, objects)
+
+        assert isinstance(et_result, jax.Array)
+        assert et_result.dtype == jnp.bool_
+        assert et_result.shape == ()
+
+        # 3. Test DetectorConvergenceCondition
+        # Mock dependencies for DetectorConvergenceCondition
+        detector_name = "test_detector"
+        arrays.detector_states[detector_name] = {"energy": jnp.zeros((end_step, 1))}
+        detector_object_mock = Mock()
+        detector_object_mock._time_step_to_arr_idx = jnp.arange(end_step)
+        objects[detector_name] = detector_object_mock
+
+        detector_cond = DetectorConvergenceCondition(
+            threshold=1e-5,
+            end_step=end_step,
+            min_steps=10,
+            detector_name=detector_name,
+        )
+        dc_result = detector_cond(state, objects)
+
+        assert isinstance(dc_result, jax.Array)
+        assert dc_result.dtype == jnp.bool_
+        assert dc_result.shape == ()
+
+    def test_time_step_condition(self, setup_simulation_state):
+        """Test the TimeStepCondition stopping condition."""
         state = setup_simulation_state["state"]
         objects = setup_simulation_state["objects"]
+        end_step = 100
 
-        # Test with different combinations of flags
-        result_time_step_condition = TimeStepCondition(state, objects)
-        result_energy_convergence_condition = EnergyConvergenceCondition(state, objects)
+        # Create the condition function
+        cond_fun = TimeStepCondition(end_step=end_step)
 
-        # Check return type
-        assert isinstance(result_time_step_condition, tuple)
-        assert isinstance(result_energy_convergence_condition, tuple)
+        # Test when time_step < end_step
+        assert cond_fun(state, objects)
 
-        assert len(result_time_step_condition) == 2
-        assert len(result_energy_convergence_condition) == 2
+        # Test when time_step == end_step
+        state = (jnp.array(end_step), state[1])
+        assert not cond_fun(state, objects)
 
[…]
