### 2025-09-07 8899e8d
Simplfied energy threshold checking, made another condition for checking convergence in detector reading
