### 2025-09-08 4bbd7f9
fixing config time for test_fdtd in fdtd module
