### 2025-09-08 b436a8f
fixing config time for test_fdtd in fdtd module
