### 2025-09-08 75045c9
unitful: static arrays for optimization within jit context
