### 2025-09-08 3a59d85
fdtdx.core: adjusted functions for untiful
