### 2025-09-09 21c4c68
unitful: jax.jit, linalg.norm, sin, cos, tan
