+++ src/fdtdx/functional/numpy.py
@@ -1,3 +1,5 @@
+import cmath
+from functools import partial
 import math
 from typing import Sequence
 import jax
@@ -8,8 +10,8 @@ from plum import dispatch, overload
 
 from fdtdx.core.fraction import Fraction
 from fdtdx.core.jax.utils import is_currently_jitting, is_traced
-from fdtdx.units.typing import PHYSICAL_DTYPES, SI
-from fdtdx.units.unitful import MAX_OPTIMIZED_ARR_SIZE, Unit, Unitful, get_static_operand
+from fdtdx.units.typing import PHYSICAL_DTYPES, SI, PhysicalArrayLike
+from fdtdx.units.unitful import EMPTY_UNIT, MAX_OPTIMIZED_ARR_SIZE, Unit, Unitful, get_static_operand, output_unitful_for_array, unary_fn
 from fdtdx.units.utils import dim_after_multiplication, handle_n_scales, is_struct_optimizable
 
 ## Square Root ###########################
@@ -74,7 +76,7 @@ def roll(
 
 @overload
 def roll(
-    x: int | float | complex | jax.Array,
+    x: jax.Array,
     *args,
     **kwargs, 
 ) -> jax.Array: 
@@ -150,6 +152,15 @@ def cross(
 ) -> jax.Array: 
     return jnp._orig_cross(a, b, *args, **kwargs)  # type: ignore
 
+@overload
+def cross(  # type: ignore
+    a,
+    b,
+    *args,
+    **kwargs,
+):
+    raise NotImplementedError(f"Currently not implemented for {a}, {b}, {args}, {kwargs}")
+
 @dispatch
 def cross(  # type: ignore
     a,
@@ -176,7 +187,25 @@ def conj(
 
 @overload
 def conj(
-    x: int | float | complex | jax.Array,
+    x: int,
+) -> int:
+    return x
+
+@overload
+def conj(
+    x: float,
+) -> float: 
+    return x
+
+@overload
+def conj(
+    x: complex,
+) -> complex: 
+    return x.conjugate()
+
+@overload
+def conj(
+    x: jax.Array,
 ) -> jax.Array: 
     return jnp._orig_conj(x)  # type: ignore
 
@@ -214,9 +243,16 @@ def dot(
     b: jax.Array | np.ndarray,
     *args,
     **kwargs,
-):
+) -> Unitful:
     new_val = jnp._orig_dot(a.val, b, *args, **kwargs)  # type: ignore
-    return Unitful(val=new_val, unit=a.unit)
+    # static computation
+    new_static_arr = None
+    if is_traced(new_val):
+        x_arr = get_static_operand(a)
+        y_arr = get_static_operand(b)
+        if x_arr is not None and y_arr is not None:
+            new_static_arr = np.dot(x_arr, y_arr, *args, **kwargs)
+    return Unitful(val=new_val, unit=a.unit, static_arr=new_static_arr)
 
 @overload
 def dot(
@@ -224,9 +260,16 @@ def dot(
     b: Unitful,
     *args,
     **kwargs,
-):
+) -> Unitful:
     new_val = jnp._orig_dot(a, b.val, *args, **kwargs)  # type: ignore
-    return Unitful(val=new_val, unit=b.unit)
+    # static computation
+    new_static_arr = None
+    if is_traced(new_val):
+        x_arr = get_static_operand(a)
+        y_arr = get_static_operand(b)
+        if x_arr is not None and y_arr is not None:
+            new_static_arr = np.dot(x_arr, y_arr, *args, **kwargs)
+    return Unitful(val=new_val, unit=b.unit, static_arr=new_static_arr)
 
 @overload
 def dot(
[…]
