+++ src/fdtdx/units/unitful.py
@@ -8,14 +8,20 @@ from jax import core
 from fdtdx.core.fraction import Fraction
 from fdtdx.core.jax.pytrees import TreeClass, autoinit, field, frozen_field
 
-from plum import dispatch, overload
+from plum import add_conversion_method, dispatch, overload
 from pytreeclass import tree_repr
 
-from fdtdx.core.jax.utils import is_traced
-from fdtdx.units.typing import SI, PhysicalArrayLike, StaticPhysicalArrayLike
+from fdtdx.core.jax.utils import is_currently_jitting, is_traced
+from fdtdx.units.typing import PHYSICAL_DTYPES, SI, PhysicalArrayLike, StaticPhysicalArrayLike
 from fdtdx.units.utils import best_scale, dim_after_multiplication, handle_different_scales
 
+""" This determines the maximum size where an array can be saved statically for optimization during jit tracing"""
 MAX_OPTIMIZED_ARR_SIZE = int(1e5)
+
+""" Global optimization stop flag. If True, then no arrays are statically saved. Additionally, no jax functions 
+like jnp.ones will return a Unitful instead of Array for scale optimization.
+"""
+OPTIMIZATION_STOP_FLAG = False
 
 @autoinit
 class Unit(TreeClass):
@@ -30,6 +36,8 @@ class Unit(TreeClass):
 
     def __repr__(self) -> str:
         return str(self)
+    
+EMPTY_UNIT = Unit(scale=0, dim={})
 
 
 @dataclass(frozen=True)
@@ -266,9 +274,18 @@ class Unitful(TreeClass):
         **kwargs
     ) -> "Unitful":
         return astype(self, *args, **kwargs)
+    
+    def squeeze(
+        self,
+        axis: int | None = None,
+    ) -> "Unitful":
+        return squeeze(self, axis)
 
     def __str__(self) -> str:
-        return f"Unitful [{self.unit}]: {tree_repr(self.val)}"
+        try:
+            return f"Unitful [{self.unit}]: {tree_repr(self.val)}"
+        except:
+            return f"Unitful [{self.unit}]: {self.shape}"
 
     def __repr__(self) -> str:
         return str(self)
@@ -394,6 +411,8 @@ def align_scales(
 def get_static_operand(
     x: Unitful | PhysicalArrayLike, 
 ) -> StaticPhysicalArrayLike | None:
+    if OPTIMIZATION_STOP_FLAG:
+        return None
     # Physical arraylike without a unit
     if isinstance(x, PhysicalArrayLike):
         if is_traced(x):
@@ -416,6 +435,22 @@ def get_static_operand(
             x_arr = np.asarray(x_arr, copy=True)
     return x_arr
 
+def output_unitful_for_array(x: jax.Array | jax.ShapeDtypeStruct) -> bool:
+    if x.dtype not in PHYSICAL_DTYPES or x.size > MAX_OPTIMIZED_ARR_SIZE:
+        return False
+    return is_currently_jitting() and not OPTIMIZATION_STOP_FLAG
+
+
+def unitful_to_array_conversion(obj: Unitful) -> jax.Array:
+    assert obj.unit.dim == {}
+    if is_currently_jitting():
+        return obj  # type: ignore
+    return obj.array_materialise()
+
+# This conversion method is necessary, because within jit-context we lie to the dispatcher.
+# Specifically, functions that are supposed to return a jax array will return a unitful to be able to perform
[…]
