### 2025-09-09 cad866e
Fleshed out DetectorConvergenceCondition using Fourier criterion
