+++ src/fdtdx/fdtd/stop_conditions.py
@@ -116,25 +116,47 @@ class EnergyThresholdCondition(StoppingCondition):
 @autoinit
 class DetectorConvergenceCondition(StoppingCondition):
     """Stopping condition based on convergence of values read off by a
-    Detector with `reduce_volume=True`.
+    Detector with `reduce_volume=True` and number of readings equal to
+    number of time steps.
 
-    This condition stops a simulation when the relative change in field
-    energy between time steps falls below a specified threshold. A minimum
-    number of steps can be enforced before convergence checks are applied,
-    and a hard cutoff is imposed at ``end_time``.
+    This stopping condition starts off estimating convergence of values read off by a
+    user-specified Detector by considering two subsets of its readings:
+      (i) the k previous full periods [t - (k+1)P, t - P), and
+      (ii) the last full period [t - P, t),
+    where t is the current time step, P is the number of samples of the period
+    over which the reading changes (it's usually half the period of the CW
+    source in the case of PoyntingFluxDetector and EnergyDetector).
+    Once it has them, it computes the mean over all periods of (i) such that the
+    result has the same shape as (ii). After this, their Fourier transform is
+    computed, and then the L2 norm of these two transforms, and stops the
+    simulation once this norm falls below a specified threshold.
+    A minimum number of steps can be enforced before convergence
+    checks are applied, and a hard cutoff is imposed when the time step is equal
+    to ``end_step``.
 
     Attributes:
+        k (int): Number of *previous* full periods used as a reference (>=1).
+        spp (int): Number of samples per period. The period depends on the detector used.
+            That is, if using a PoyntingFluxDetector or EnergyDetector, this period
+            is half that of the CW source used. So, to obtain spp, do
+            (cw_source_period / 2) / time_step_duration. The latter can be found
+            at SimulationConfig.time_step_duration.
         threshold (float): Relative change threshold for determining
             convergence. Defaults to ``1e-6``.
         end_step (int): The maximum number of time steps before the
             simulation is stopped, regardless of convergence.
+            Can be obtained from SimulationConfig.time_steps_total.
         min_steps (int): The minimum number of time steps that must be
-            completed before convergence checking begins. Defaults to
+            completed before convergence checking begins. A good value would be
+            (k_periods + 1) * spp, where spp is samples per period. Defaults to
             ``100``.
-        detector_name (str): The name of the EnergyDetector from which
-            energy values will be obtained in order to determine the stopping condition.
+        detector_name (str): The name of the Detector from which
+            detector readings will be obtained in order to determine when to
+            halt the simulation.
     """
 
+    k: int = frozen_field(default=4)
+    spp: int = frozen_field()
     threshold: float = frozen_field(default=1e-6)
     end_step: int = frozen_field()
     min_steps: int = frozen_field(default=100)
@@ -142,32 +164,36 @@ class DetectorConvergenceCondition(StoppingCondition):
 
     def validate(self, state: SimulationState, objects: ObjectContainer) -> None:
         _, arrays = state
+
         if self.detector_name not in arrays.detector_states:
             available = tuple(arrays.detector_states.keys())
             raise KeyError(f"Detector '{self.detector_name}' not found. Available detectors: {available}")
         det_state: DetectorState = arrays.detector_states[self.detector_name]
-        if ("energy" not in det_state) or ("poynting_flux" not in det_state) or ("fields" not in det_state):
+
[…]
