### 2025-09-09 be22c89
Added a test for DetectorConvergenceCondition
