### 2025-09-09 8f3f7c6
unitful: asarray, array, exp, expand_dims, where
