+++ src/fdtdx/units/unitful.py
@@ -12,7 +12,7 @@ from plum import add_conversion_method, dispatch, overload
 from pytreeclass import tree_repr
 
 from fdtdx.core.jax.utils import is_currently_jitting, is_traced
-from fdtdx.units.typing import PHYSICAL_DTYPES, SI, PhysicalArrayLike, StaticPhysicalArrayLike
+from fdtdx.units.typing import PHYSICAL_DTYPES, SI, PhysicalArrayLike, RealPhysicalArrayLike, StaticPhysicalArrayLike
 from fdtdx.units.utils import best_scale, dim_after_multiplication, handle_different_scales
 
 """ This determines the maximum size where an array can be saved statically for optimization during jit tracing"""
@@ -280,6 +280,13 @@ class Unitful(TreeClass):
         axis: int | None = None,
     ) -> "Unitful":
         return squeeze(self, axis)
+    
+    def reshape(
+        self,
+        *args,
+        **kwargs,
+    ) -> "Unitful":
+        return reshape(self, args, **kwargs)
 
     def __str__(self) -> str:
         try:
@@ -302,34 +309,34 @@ class Unitful(TreeClass):
     def __rtruediv__(self, other: PhysicalArrayLike | "Unitful") -> "Unitful":
         return divide(other, self)
     
-    def __add__(self, other: "Unitful") -> "Unitful":
+    def __add__(self, other: "Unitful | PhysicalArrayLike") -> "Unitful":
         return add(self, other)
     
-    def __radd__(self, other: "Unitful") -> "Unitful":
+    def __radd__(self, other: "Unitful | PhysicalArrayLike") -> "Unitful":
         return add(self, other)
     
-    def __sub__(self, other: "Unitful") -> "Unitful":
+    def __sub__(self, other: "Unitful | PhysicalArrayLike") -> "Unitful":
         return subtract(self, other)
     
-    def __rsub__(self, other: "Unitful") -> "Unitful":
+    def __rsub__(self, other: "Unitful | PhysicalArrayLike") -> "Unitful":
         return subtract(other, self)
     
-    def __lt__(self, other: "Unitful") -> jax.Array | bool:
+    def __lt__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
         return lt(self, other)
     
-    def __le__(self, other: "Unitful") -> jax.Array | bool:
+    def __le__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
         return le(self, other)
     
-    def __eq__(self, other: "Unitful") -> jax.Array | bool:
+    def __eq__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
         return eq(self, other)
     
-    def __ne__(self, other: "Unitful") -> jax.Array | bool:  # type: ignore
+    def __ne__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:  # type: ignore
         return ne(self, other)
     
-    def __ge__(self, other: "Unitful") -> jax.Array | bool:
+    def __ge__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
         return ge(self, other)
     
-    def __gt__(self, other: "Unitful") -> jax.Array | bool:
+    def __gt__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
         return gt(self, other)
     
     def __len__(self):
[…]
