### 2025-09-10 0711f49
adding sphinx docstrings for conversion and core modules
