+++ src/fdtdx/core/physics/modes.py
@@ -12,33 +12,18 @@ from fdtdx.core.physics.metrics import normalize_by_poynting_flux
 
 ModeTupleType = namedtuple("Mode", ["neff", "Ex", "Ey", "Ez", "Hx", "Hy", "Hz"])
 """A named tuple containing the mode fields and effective index.
-
-Attributes:
-    neff: Complex effective refractive index of the mode
-    Ex: x-component of the electric field
-    Ey: y-component of the electric field
-    Ez: z-component of the electric field
-    Hx: x-component of the magnetic field
-    Hy: y-component of the magnetic field
-    Hz: z-component of the magnetic field
 """
 
 
 def compute_mode_polarization_fraction(
+    #: a ModeTupleType instance
     mode: ModeTupleType,
+    #: indices of transverse E-field component axes.
     tangential_axes: tuple[int, int],
+    #: "te" or "tm" determines which axis is 'E1'
     pol: Literal["te", "tm"],
 ) -> float:
-    """Mode polarization fraction.
-
-    Args:
-        mode (ModeTupleType): a ModeTupleType instance
-        tangential_axes (tuple[int, int]): indices of transverse E-field component axes.
-        pol (Literal["te", "tm"]): "te" or "tm" determines which axis is 'E1'
-
-    Returns:
-        float: Polarization fraction between 0 and 1.
-    """
+    """Mode polarization fraction."""
 
     E_fields = [mode.Ex, mode.Ey, mode.Ez]
     E1 = E_fields[tangential_axes[0]]
@@ -52,24 +37,21 @@ def compute_mode_polarization_fraction(
         raise ValueError(f"pol must be 'te' or 'tm', but got {pol}")
 
     denominator = np.sum(np.abs(E1) ** 2 + np.abs(E2) ** 2) + 1e-18
+
+    #: Polarization fraction between 0 and 1.
     return numerator / denominator
 
 
 def sort_modes(
+    #: list of modes.
     modes: list[ModeTupleType],
+    #: If not none, sort by polarization specificaton.
     filter_pol: Literal["te", "tm"] | None,
+    #: indices of transverse E-field component axes.
     tangential_axes: tuple[int, int],
 ) -> list[ModeTupleType]:
     """
     Sort modes by polarization.
-
-    Args:
-        modes (list[ModeTupleType]): list of modes.
-        filter_pol (Literal["te", "tm"] | None): If not none, sort by polarization specificaton.
-        tangential_axes (tuple[int, int]): indices of transverse E-field component axes.
-
-    Returns:
-        list[ModeTupleType]: sorted list of modes.
     """
     if filter_pol is None:
         return sorted(modes, key=lambda m: float(np.real(m.neff)), reverse=True)
@@ -84,6 +66,7 @@ def sort_modes(
     matching_sorted = sorted(matching, key=lambda m: float(np.real(m.neff)), reverse=True)
     non_matching_sorted = sorted(non_matching, key=lambda m: float(np.real(m.neff)), reverse=True)
 
+    #: sorted list of modes.
     return matching_sorted + non_matching_sorted
 
 
@@ -189,15 +172,25 @@ def compute_mode(
 
 
 def tidy3d_mode_computation_wrapper(
+    #: Operating frequency in Hz
     frequency: float,
+    #: 2D array of relative permittivity values
     permittivity_cross_section: np.ndarray,
+    #: List of coordinate arrays [x, y] defining the grid
     coords: List[np.ndarray],
+    #: Propagation direction, either "+" or "-"
     direction: Literal["+", "-"],
+    #: 2D array of relative permeability values. Defauts to None.
     permeability_cross_section: np.ndarray | None = None,
[…]
