+++ src/fdtdx/functional/numpy.py
@@ -1,4 +1,3 @@
-import cmath
 from functools import partial
 import math
 from typing import Sequence
@@ -9,9 +8,9 @@ import numpy as np
 from plum import dispatch, overload
 
 from fdtdx.core.fraction import Fraction
-from fdtdx.core.jax.utils import is_currently_jitting, is_traced
-from fdtdx.units.typing import PHYSICAL_DTYPES, SI, PhysicalArrayLike
-from fdtdx.units.unitful import EMPTY_UNIT, MAX_OPTIMIZED_ARR_SIZE, Unit, Unitful, align_scales, get_static_operand, output_unitful_for_array, unary_fn
+from fdtdx.core.jax.utils import is_traced
+from fdtdx.units.typing import SI, PhysicalArrayLike
+from fdtdx.units.unitful import EMPTY_UNIT, MAX_STATIC_OPTIMIZED_SIZE, Unit, Unitful, align_scales, can_perform_scatic_ops, get_static_operand, output_unitful_for_array
 from fdtdx.units.utils import dim_after_multiplication, handle_n_scales, is_struct_optimizable
 
 ## Square Root ###########################
@@ -398,21 +397,21 @@ def stack(
 
 @overload
 def stack(
-    x: jax.Array | np.ndarray | Sequence[ArrayLike],
+    x: ArrayLike | Sequence[ArrayLike],
     *args,
     **kwargs,
-) -> jax.Array:
-    all_physical = (
-        True 
-        if isinstance(x, jax.Array | np.ndarray) 
-        else all([isinstance(x_i, PhysicalArrayLike) for x_i in x])
-    )
+) -> ArrayLike:
+    if isinstance(x, Sequence):
+        all_physical = all([output_unitful_for_array(x_i) for x_i in x])
+    else:
+        all_physical = output_unitful_for_array(x)
     # axis/dtype args/kwargs needs to be static for eval_shape
     partial_orig_fn = partial(jnp._orig_stack, x, *args, **kwargs)  # type: ignore
     result_shape_dtype = jax.eval_shape(partial_orig_fn)
+    # if we cannot convert to unitful just call original function
     if not output_unitful_for_array(result_shape_dtype) or not all_physical:
         return jnp._orig_stack(x, *args, **kwargs)  # type: ignore        
-    
+    # convert inputs to unitful without dimension
     if isinstance(x, Sequence):
         unit_input = [Unitful(val=x_i, unit=EMPTY_UNIT) for x_i in x] # type: ignore
     else:
@@ -640,21 +639,22 @@ def asarray(
     *args,
     **kwargs,
 ) -> jax.Array:
-    materialised = jax.tree_map(
+    materialised = jax.tree.map(
         lambda x: x.materialise() if isinstance(x, Unitful) else x,
         a,
         is_leaf=lambda x: isinstance(x, Unitful),
     )
+    partial_orig_fn = partial(jnp._orig_asarray, materialised, *args, **kwargs)  # type: ignore
+    result_shape_dtype = jax.eval_shape(partial_orig_fn)
     result: jax.Array = jnp._orig_asarray(materialised, *args, **kwargs)  # type: ignore
-        
-    if not output_unitful_for_array(result):
+    if not output_unitful_for_array(result_shape_dtype):
         # cannot use this as Unitful, wrong dtype
         return result
     
     # try to get a static version of the array and save to trace metadata
     static_arr = None
     result_size = math.prod(result.shape)
-    if is_struct_optimizable(a) and result_size <= MAX_OPTIMIZED_ARR_SIZE:
+    if is_struct_optimizable(a) and result_size <= MAX_STATIC_OPTIMIZED_SIZE:
[…]

[…]
