### 2025-09-10 1b0fade
added unitful manifesto
