### 2025-09-11 699a21c
Made stopping_condition optional again to fix breakage
