### 2025-09-11 e3f391a
adding sphinx docstrings for attributes
