### 2025-09-11 d62ad10
functional: matmul respects static arrs
