### 2025-09-11 c886ab3
initialization.py: Added is_compiling check
