### 2025-09-11 ff1c8e5
unitful: renaming of is_currently_compiling
