### 2025-09-12 1d2afbc
unitful: implemented prod, floor, ceil, improved comparison ops
