+++ src/fdtdx/units/__init__.py
@@ -7,12 +7,12 @@ us_unit = Unit(scale=-6, dim={SI.s: 1})
 ns_unit = Unit(scale=-9, dim={SI.s: 1})
 ps_unit = Unit(scale=-12, dim={SI.s: 1})
 fs_unit = Unit(scale=-15, dim={SI.s: 1})
-s = Unitful(val=1, unit=s_unit)
-ms = Unitful(val=1, unit=ms_unit)
-us = Unitful(val=1, unit=us_unit)
-ns = Unitful(val=1, unit=ns_unit)
-ps = Unitful(val=1, unit=ps_unit)
-fs = Unitful(val=1, unit=fs_unit)
+s = Unitful(val=1.0, unit=s_unit)
+ms = Unitful(val=1.0, unit=ms_unit)
+us = Unitful(val=1.0, unit=us_unit)
+ns = Unitful(val=1.0, unit=ns_unit)
+ps = Unitful(val=1.0, unit=ps_unit)
+fs = Unitful(val=1.0, unit=fs_unit)
 
 Hz_unit = Unit(scale=0, dim={SI.s: -1})
 kHz_unit = Unit(scale=3, dim={SI.s: -1})
@@ -20,12 +20,12 @@ MHz_unit = Unit(scale=6, dim={SI.s: -1})
 GHz_unit = Unit(scale=9, dim={SI.s: -1})
 THz_unit = Unit(scale=12, dim={SI.s: -1})
 PHz_unit = Unit(scale=15, dim={SI.s: -1})
-Hz = Unitful(val=1, unit=Hz_unit)
-kHz = Unitful(val=1, unit=kHz_unit)
-MHz = Unitful(val=1, unit=MHz_unit)
-GHz = Unitful(val=1, unit=GHz_unit)
-THz = Unitful(val=1, unit=THz_unit)
-PHz = Unitful(val=1, unit=PHz_unit)
+Hz = Unitful(val=1.0, unit=Hz_unit)
+kHz = Unitful(val=1.0, unit=kHz_unit)
+MHz = Unitful(val=1.0, unit=MHz_unit)
+GHz = Unitful(val=1.0, unit=GHz_unit)
+THz = Unitful(val=1.0, unit=THz_unit)
+PHz = Unitful(val=1.0, unit=PHz_unit)
 
 km_unit = Unit(scale=3, dim={SI.m: 1})
 m_unit = Unit(scale=0, dim={SI.m: 1})
@@ -33,41 +33,41 @@ mm_unit = Unit(scale=-3, dim={SI.m: 1})
 um_unit = Unit(scale=-6, dim={SI.m: 1})
 nm_unit = Unit(scale=-9, dim={SI.m: 1})
 pm_unit = Unit(scale=-12, dim={SI.m: 1})
-km = Unitful(val=1, unit=km_unit)
-m = Unitful(val=1, unit=m_unit)
-mm = Unitful(val=1, unit=mm_unit)
-um = Unitful(val=1, unit=um_unit)
-nm = Unitful(val=1, unit=nm_unit)
-pm = Unitful(val=1, unit=pm_unit)
+km = Unitful(val=1.0, unit=km_unit)
+m = Unitful(val=1.0, unit=m_unit)
+mm = Unitful(val=1.0, unit=mm_unit)
+um = Unitful(val=1.0, unit=um_unit)
+nm = Unitful(val=1.0, unit=nm_unit)
+pm = Unitful(val=1.0, unit=pm_unit)
 
 MW_unit = Unit(scale=6, dim={SI.kg: 1, SI.m: 2, SI.s: -3})
[…]
