+++ src/fdtdx/units/unitful.py
@@ -175,9 +175,17 @@ class Unitful(TreeClass):
         )
         if isinstance(self.val, NonPhysicalArrayLike) or bad_dtype:
             if self.unit.scale != 0:
-                raise Exception(f"Cannot have non-zero scale for non-physical {self}")
+                if isinstance(self.val, int):
+                    raise Exception(f"Cannot have non-zero scale for integer {self}. Consider using float as input.")
+                else:
+                    raise Exception(f"Cannot have non-zero scale for non-physical {self}")
             if self.unit.dim != {}:
-                raise Exception(f"Cannot have non-empty dimension for non-physical {self}")
+                if isinstance(self.val, int):
+                    raise Exception(
+                        f"Cannot have non-empty dimension for integer {self}. Consider using float as input."
+                    )
+                else:
+                    raise Exception(f"Cannot have non-empty dimension for non-physical {self}")
     
     def __post_init__(self):
         self._validate()
@@ -235,6 +243,13 @@ class Unitful(TreeClass):
         v = self.value()
         assert isinstance(v, float), f"safe float_value called on Unitful with non-float value: {self}"
         return v
+    
+    def static_value(self) -> StaticArrayLike | None:
+        if self.static_arr is None: return None
+        scale = self.unit.scale
+        if isinstance(scale, Fraction):
+            scale = scale.value()
+        return self.static_arr * (10 ** scale)
     
     @property
     def at(self) -> UnitfulIndexer:  # type: ignore
@@ -335,22 +350,22 @@ class Unitful(TreeClass):
     def __rsub__(self, other: "Unitful | PhysicalArrayLike") -> "Unitful":
         return subtract(other, self)
     
-    def __lt__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
+    def __lt__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
         return lt(self, other)
     
-    def __le__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
+    def __le__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
         return le(self, other)
     
-    def __eq__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
+    def __eq__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
         return eq(self, other)
     
-    def __ne__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:  # type: ignore
+    def __ne__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:  # type: ignore
         return ne(self, other)
     
-    def __ge__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
+    def __ge__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
         return ge(self, other)
     
-    def __gt__(self, other: "Unitful | RealPhysicalArrayLike") -> jax.Array | bool:
[…]
