### 2025-09-12 ec8e18c
bugfix in matmul with unitful
