### 2025-09-12 8c16cb4
unitful: subtractions with integers
