### 2025-09-12 50841f5
unitful: implemented .T property
