### 2025-09-12 ad6913d
unitful: indexing with unitless unitful
