### 2025-09-12 069bac0
docs: moved docs to readthedocs
