+++ docs/source/quickstart.rst
@@ -2,9 +2,12 @@ Quickstart
 ==========
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 1
    :caption: Contents:
 
    notebooks/JAX_in_FDTDX.ipynb
    notebooks/Object_placement_guide.ipynb
    notebooks/Materials_guide.ipynb
+
+
+More guides will follow shortly!
