### 2025-09-12 5895ca0
adding test_misc for fdtd module
