### 2025-09-12 1ebb25d
docs: bugfix in tutorial fetching
