### 2025-09-15 07c9763
compute_modes: exported from fdtdx, added docs
