### 2025-09-15 b83f672
detector.py: moved detector states to numpy as memory optimization
