### 2025-09-15 160909e
docs: added basic simulation script to quickstart
