+++ docs/source/quickstart.rst
@@ -8,6 +8,7 @@ Quickstart
    notebooks/jax_introduction.ipynb
    notebooks/basic_materials.ipynb
    notebooks/object_placement_guide.ipynb
+   notebooks/basic_simulation.ipynb
 
 
 More guides will follow shortly!
