### 2025-09-15 a48bbb0
adding interface_slice for test_misc in fdtd module
