### 2025-09-15 da001b3
Added new method, docstrings, using wave_character and more clear attribute names,
