+++ src/fdtdx/fdtd/stop_conditions.py
@@ -1,10 +1,13 @@
 from abc import ABC, abstractmethod
+from typing import Self
 
 import jax
 import jax.numpy as jnp
 
-from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field
-from fdtdx.fdtd.container import ArrayContainer, DetectorState, SimulationState, ObjectContainer
+from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field, frozen_private_field
+from fdtdx.fdtd.container import DetectorState, SimulationState, ObjectContainer
+from fdtdx.core.wavelength import WaveCharacter
+from fdtdx.config import SimulationConfig
 
 
 @autoinit
@@ -16,17 +19,23 @@ class StoppingCondition(TreeClass, ABC):
     """
 
     @abstractmethod
-    def validate(self, state: SimulationState, objects: ObjectContainer) -> None:
-        """Optional pre-run validation; override in subclasses."""
-        return
+    def setup(self, state: SimulationState, config: SimulationConfig, objects: ObjectContainer) -> Self:
+        """Preparing condition for its use in the simulation."""
+        return self
 
     @abstractmethod
-    def __call__(self, state: SimulationState, objects: ObjectContainer) -> jax.Array:
+    def _validate(self, state: SimulationState, config: SimulationConfig, objects: ObjectContainer) -> None:
+        """Pre-run validation; override in subclasses."""
+        pass
+
+    @abstractmethod
+    def __call__(self, state: SimulationState, config: SimulationConfig, objects: ObjectContainer) -> jax.Array:
         """Evaluate the stopping condition.
 
         Args:
-            state: Current simulation state (SimulationState)
-            objects: Objects in simulation (ObjectContainer)
+            state (SimulationState): Current simulation state
+            config (SimulationConfig): Configuration of the simulation
+            objects (ObjectContainer): Objects in simulation
 
         Returns:
             jax.Array: Boolean scalar - True if simulation should continue, False if it should stop
@@ -43,29 +52,28 @@ class TimeStepCondition(StoppingCondition):
 
     This recreates the original behavior where simulation continues until
     a specified end time is reached.
-
-    Attributes:
-        end_step (int): Time step at which simulation ends.
-            Can be found at SimulationConfig.time_step_duration
     """
 
-    end_step: int = frozen_field()
+    def setup(self, state: SimulationState, config: SimulationConfig, objects: ObjectContainer) -> Self:
+        self._validate(state, config, objects)
+        return self
 
-    def validate(self, state: SimulationState, objects: ObjectContainer) -> None:
+    def _validate(self, state: SimulationState, config: SimulationConfig, objects: ObjectContainer) -> None:
         pass
 
-    def __call__(self, state: SimulationState, objects: ObjectContainer) -> jax.Array:
+    def __call__(self, state: SimulationState, config: SimulationConfig, objects: ObjectContainer) -> jax.Array:
         """Check if simulation should continue based on time step.
 
         Args:
-            state: Current simulation state (SimulationState)
-            objects: Objects in simulation (ObjectContainer)
+            state (SimulationState): Current simulation state.
+            config (SimulationConfig): Configuration of the simulation.
+            objects (ObjectContainer): Objects in simulation.
 
         Returns:
             jax.Array: Boolean scalar - True if current time < end_time, False otherwise
         """
-        time_step, _ = state
-        return self.end_step > time_step
+        curr_time_step, _ = state
+        return config.time_steps_total > curr_time_step
 
 
 @autoinit
@@ -73,40 +81,46 @@ class EnergyThresholdCondition(StoppingCondition):
     """This condition stops a simulation when the total energy in the simulation volume
     falls below a specified threshold. A minimum number of steps can be enforced
[…]
