### 2025-09-15 e419e1e
Made some tests work, others saw progress but still fail
