+++ tests/fdtd/test_condition.py
@@ -10,9 +10,13 @@ from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 from fdtdx.interfaces.state import RecordingState
 from fdtdx.objects.boundaries.boundary import BaseBoundaryState
 from fdtdx.objects.detectors.detector import DetectorState
+from fdtdx.core.wavelength import WaveCharacter
+from fdtdx.objects.detectors.energy import EnergyDetector
+from fdtdx.objects.static_material.static import SimulationVolume
 
 # Import the module to test
 from fdtdx.fdtd.stop_conditions import TimeStepCondition, EnergyThresholdCondition, DetectorConvergenceCondition
+from fdtdx.objects.detectors.energy import EnergyDetector
 
 
 class TestCondition:
@@ -26,8 +30,8 @@ class TestCondition:
         inv_permeabilities = jnp.ones((10, 10, 10))
 
         # Create mock boundary and detector states
-        boundary_states = {"pml": Mock(spec=BaseBoundaryState)}
-        detector_states = {"detector1": Mock(spec=DetectorState)}
+        boundary_states = {"pml": BaseBoundaryState()}
+        detector_states = {"detector1": DetectorState()}
         recording_state = Mock(spec=RecordingState)
 
         # Create array container
@@ -56,212 +60,217 @@ class TestCondition:
 
     def test_condition_function_signature(self, setup_simulation_state):
         """Test that stopping condition function has the correct signature and returns expected types."""
-        sim_data = setup_simulation_state
-        state = sim_data["state"]
-        objects = sim_data["objects"]
-        arrays = sim_data["arrays"]
-        end_step = 100
+        state = setup_simulation_state["state"]
+        arrays = setup_simulation_state["arrays"]
+        cw_source_period = 5e-11  # 20 GHz
+        wave_character = WaveCharacter(period=cw_source_period)
+        threshold = 1e-6
+        min_steps = 10
+
+        detector_name = "test_detector"
+        detector_object = EnergyDetector(name=detector_name)
+        config = SimulationConfig(time=100e-11, resolution=1e-4)
+        volume = SimulationVolume(partial_real_shape=(1e-3, 1e-3, 1e-3))
+        objects = ObjectContainer(
+            object_list=[volume, detector_object], volume_idx=0,
+        )
+        arrays.detector_states[detector_name] = {"energy": jnp.zeros((config.time_steps_total, 1))}
 
         # TimeStepCondition
-        time_step_cond = TimeStepCondition(end_step=end_step)
-        ts_result = time_step_cond(state, objects)
-        
+        time_step_cond = TimeStepCondition()
+        ts_result = time_step_cond(state, config, objects)
+
         assert isinstance(ts_result, jax.Array)
         assert ts_result.dtype == jnp.bool_
         assert ts_result.shape == ()
 
         # EnergyThresholdCondition
-        energy_threshold_cond = EnergyThresholdCondition(
-            threshold=1e-6,
-            end_step=end_step,
-            min_steps=10,
+        energy_thresh_cond = EnergyThresholdCondition(
+            threshold=threshold,
+            min_steps=min_steps,
         )
-        et_result = energy_threshold_cond(state, objects)
+        energy_thresh_cond = energy_thresh_cond.setup(state, config, objects)
+        et_result = energy_thresh_cond(state, config, objects)
 
         assert isinstance(et_result, jax.Array)
         assert et_result.dtype == jnp.bool_
         assert et_result.shape == ()
 
-        # Test DetectorConvergenceCondition
-        # Mock dependencies for DetectorConvergenceCondition
-        detector_name = "test_detector"
[…]
