### 2025-09-16 65437b2
wavecharacter: refactored post_init for json import/export
