### 2025-09-16 7fefbe1
typing: added jax dtype list
