### 2025-09-16 4c53d8b
added null value for undefined including None values
