### 2025-09-16 d6a5561
implemented json import / export
