+++ src/fdtdx/core/null.py
@@ -1,15 +1,14 @@
-
 class Null:
-    
     __slots__ = ()
-    
+
     def __repr__(self) -> str:
         return "null"
-    
+
     def __str__(self) -> str:
         return repr(self)
-    
-    __bool__ = lambda _: False
+
+    def __bool__(self):
+        return False
 
 
 NULL = Null()
