### 2025-09-17 ce6cc41
config & wavelength: updated attributes to unitful
