### 2025-09-17 1f8be8e
null: renamed to custom_null to avoid naming problems
