+++ tests/fdtd/test_initialization.py
@@ -1,0 +1,282 @@
+from unittest.mock import MagicMock, Mock, patch
+
+import jax
+import jax.numpy as jnp
+import pytest
+
+from fdtdx.config import SimulationConfig
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
+
+# Import the functions to test
+from fdtdx.fdtd.initialization import (
+    _init_arrays,
+    _init_params,
+    _resolve_object_constraints,
+    apply_params,
+    place_objects,
+)
+from fdtdx.objects.device.parameters.transform import ParameterType
+from fdtdx.objects.object import (
+    GridCoordinateConstraint,
+    PositionConstraint,
+    RealCoordinateConstraint,
+    SimulationObject,
+)
+from fdtdx.objects.static_material.static import UniformMaterialObject
+
+
+class TestInitialization:
+    @pytest.fixture
+    def mock_config(self):
+        """Create a mock SimulationConfig for testing."""
+        config = Mock(spec=SimulationConfig)
+        config.dtype = jnp.float32
+        config.backend = "cpu"
+        config.resolution = 1.0
+        config.time_steps_total = 100
+        config.gradient_config = None
+        return config
+
+    @pytest.fixture
+    def mock_volume(self):
+        """Create a mock volume object for testing."""
+        volume = Mock(spec=SimulationObject)
+        volume.name = "volume"
+        volume.grid_shape = (10, 10, 10)
+        volume.partial_grid_shape = (10, 10, 10)
+        volume.partial_real_shape = (10.0, 10.0, 10.0)
+        volume.placement_order = 0
+        return volume
+
+    @pytest.fixture
+    def mock_object(self):
+        """Create a mock simulation object for testing."""
+        obj = Mock(spec=SimulationObject)
+        obj.name = "test_object"
+        obj.grid_shape = (5, 5, 5)
+        obj.partial_grid_shape = (5, 5, 5)
+        obj.partial_real_shape = (5.0, 5.0, 5.0)
+        obj.placement_order = 1
+        return obj
+
+    @pytest.fixture
+    def mock_objects(self):
+        """Create a mock ObjectContainer for testing."""
+        objects = Mock(spec=ObjectContainer)
+        objects.devices = []
+        return objects
+
+    @pytest.fixture
+    def mock_uniform_material_object(self):
+        """Create a mock UniformMaterialObject for testing."""
+        material = Mock()
+        material.permittivity = 2.0
+        material.permeability = 1.0
+        material.electric_conductivity = 0.0
+        material.magnetic_conductivity = 0.0
+
+        obj = Mock(spec=UniformMaterialObject)
+        obj.name = "material_object"
+        obj.grid_shape = (5, 5, 5)
+        obj.grid_slice = (slice(2, 7), slice(2, 7), slice(2, 7))
+        obj.material = material
+        obj.placement_order = 1
+        return obj
+
+    @pytest.fixture
+    def random_key(self):
+        """Create a JAX random key for testing."""
+        return jax.random.PRNGKey(42)
+
+    def test_place_objects(self, mock_volume, mock_config, random_key):
+        """Test the place_objects function."""
+        # Create a simple position constraint
+        constraint = Mock(spec=PositionConstraint)
+        constraint.object = mock_volume
+        constraint.other_object = None
+
+        # Mock the internal functions
+        with (
+            patch("fdtdx.fdtd.initialization._resolve_object_constraints") as mock_resolve,
+            patch("fdtdx.fdtd.initialization._init_params") as mock_init_params,
+            patch("fdtdx.fdtd.initialization._init_arrays") as mock_init_arrays,
+        ):
+            # Set up mock returns
[…]
