### 2025-09-18 961a224
Bug fix: Use getter methods to access WaveCharacter properties.
