### 2025-09-24 3b1d2e5
Fix code formatting issues
- Run pre-commit hooks to fix ruff-format violations
- Break long lines in source.py for better readability
