### 2025-09-24 40d9978
Setting attributes in setup(), validating in the right method, docstring, fixing return value of DCC
