+++ tests/fdtd/test_condition.py
@@ -1,4 +1,4 @@
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
 import jax
 import jax.numpy as jnp
@@ -7,7 +7,6 @@ import pytest
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 
-from fdtdx.interfaces.state import RecordingState
 from fdtdx.objects.boundaries.boundary import BaseBoundaryState
 from fdtdx.objects.detectors.detector import DetectorState
 from fdtdx.core.wavelength import WaveCharacter
@@ -16,7 +15,6 @@ from fdtdx.objects.static_material.static import SimulationVolume
 
 # Import the module to test
 from fdtdx.fdtd.stop_conditions import TimeStepCondition, EnergyThresholdCondition, DetectorConvergenceCondition
-from fdtdx.objects.detectors.energy import EnergyDetector
 
 
 class TestCondition:
@@ -32,7 +30,6 @@ class TestCondition:
         # Create mock boundary and detector states
         boundary_states = {"pml": BaseBoundaryState()}
         detector_states = {"detector1": DetectorState()}
-        recording_state = Mock(spec=RecordingState)
 
         # Create array container
         arrays = ArrayContainer(
@@ -42,15 +39,20 @@ class TestCondition:
             inv_permeabilities=inv_permeabilities,
             boundary_states=boundary_states,
             detector_states=detector_states,
-            recording_state=recording_state,
+            recording_state=None,
+            electric_conductivity=None,
+            magnetic_conductivity=None,
         )
 
         # Create simulation state
         time_step = jnp.array(0)
         state = (time_step, arrays)
 
-        # Create mock config and objects
-        config = Mock(spec=SimulationConfig)
+        config = SimulationConfig(
+            time=100e-11,
+            resolution=1e-4,
+            courant_factor=0.99,
+        )
         objects = Mock(spec=ObjectContainer)
 
         # Create random key
@@ -65,7 +67,8 @@ class TestCondition:
         cw_source_period = 5e-11  # 20 GHz
         wave_character = WaveCharacter(period=cw_source_period)
         threshold = 1e-6
-        min_steps = 10
+        min_steps = 1572 # Approximately 5 periods at 20 GHz with dt ~ 1.906e-13 s.
+        # It's so specific because we ensure min_steps > spp * (prev_periods + 1), otherwise a ValueError is raised
 
         detector_name = "test_detector"
         detector_object = EnergyDetector(name=detector_name)
@@ -115,11 +118,11 @@ class TestCondition:
 
     def test_time_step_condition(self, setup_simulation_state):
         state = setup_simulation_state["state"]
+        config = setup_simulation_state["config"]
         objects = setup_simulation_state["objects"]
-        config = SimulationConfig(time=10e-11, resolution=1e-4)
 
-        # Create the condition function
         cond_fun = TimeStepCondition()
+        cond_fun = cond_fun.setup(state, config, objects)
 
         # Test when time_step < end_step
         assert cond_fun(state, config, objects)
@@ -134,11 +137,10 @@ class TestCondition:
 
     def test_energy_threshold_condition(self, setup_simulation_state):
         objects = setup_simulation_state["objects"]
+        config = setup_simulation_state["config"]
         arrays = setup_simulation_state["arrays"]
-        config = SimulationConfig(time=10e-11, resolution=1e-4)
-
-        min_steps = 100
         threshold = 1e-5
+        min_steps = 10
 
         cond_fun = EnergyThresholdCondition(
             threshold=threshold,
@@ -147,11 +149,13 @@ class TestCondition:
 
         # Test before min_steps -> should continue
[…]
