### 2025-09-25 2584ecc
adding test_modes for core physics module
