+++ tests/core/physics/test_modes.py
@@ -1,0 +1,357 @@
+from unittest.mock import patch
+
+import jax.numpy as jnp
+import numpy as np
+import pytest
+
+from fdtdx.core.physics.modes import (
+    ModeTupleType,
+    compute_mode,
+    compute_mode_polarization_fraction,
+    sort_modes,
+    tidy3d_mode_computation_wrapper,
+)
+
+
+class TestModeTupleType:
+    """Test the ModeTupleType named tuple."""
+
+    def test_mode_tuple_creation(self):
+        """Test creating a ModeTupleType instance."""
+        mode = ModeTupleType(
+            neff=1.5 + 0.1j,
+            Ex=np.array([1, 2, 3]),
+            Ey=np.array([4, 5, 6]),
+            Ez=np.array([7, 8, 9]),
+            Hx=np.array([10, 11, 12]),
+            Hy=np.array([13, 14, 15]),
+            Hz=np.array([16, 17, 18]),
+        )
+
+        assert mode.neff == 1.5 + 0.1j
+        assert np.array_equal(mode.Ex, np.array([1, 2, 3]))
+        assert isinstance(mode, tuple)  # Should behave like a tuple
+
+
+class TestComputeModePolarizationFraction:
+    """Test the compute_mode_polarization_fraction function."""
+
+    def test_te_polarization(self):
+        """Test TE polarization fraction calculation."""
+        # Create a mode with stronger E1 component (TE-like)
+        mode = ModeTupleType(
+            neff=1.5,
+            Ex=np.array([[2.0, 2.0], [2.0, 2.0]]),  # E1 component (axis 0)
+            Ey=np.array([[1.0, 1.0], [1.0, 1.0]]),  # E2 component (axis 1)
+            Ez=np.array([[0.0, 0.0], [0.0, 0.0]]),
+            Hx=np.array([[0.0, 0.0], [0.0, 0.0]]),
+            Hy=np.array([[0.0, 0.0], [0.0, 0.0]]),
+            Hz=np.array([[0.0, 0.0], [0.0, 0.0]]),
+        )
+
+        fraction = compute_mode_polarization_fraction(mode, (0, 1), "te")
+
+        # Expected: |E1|^2 / (|E1|^2 + |E2|^2) = 16 / (16 + 4) = 0.8
+        expected = 16.0 / (16.0 + 4.0)
+        assert fraction == pytest.approx(expected)
+
+    def test_tm_polarization(self):
+        """Test TM polarization fraction calculation."""
+        # Create a mode with stronger E2 component (TM-like)
+        mode = ModeTupleType(
+            neff=1.5,
+            Ex=np.array([[1.0, 1.0], [1.0, 1.0]]),  # E1 component (axis 0)
+            Ey=np.array([[3.0, 3.0], [3.0, 3.0]]),  # E2 component (axis 1)
+            Ez=np.array([[0.0, 0.0], [0.0, 0.0]]),
+            Hx=np.array([[0.0, 0.0], [0.0, 0.0]]),
+            Hy=np.array([[0.0, 0.0], [0.0, 0.0]]),
+            Hz=np.array([[0.0, 0.0], [0.0, 0.0]]),
+        )
+
+        fraction = compute_mode_polarization_fraction(mode, (0, 1), "tm")
+
+        # Expected: |E2|^2 / (|E1|^2 + |E2|^2) = 36 / (4 + 36) = 0.9
+        expected = 36.0 / (4.0 + 36.0)
+        assert fraction == pytest.approx(expected)
+
+    def test_invalid_pol_raises_error(self):
+        """Test that invalid polarization raises ValueError."""
+        mode = ModeTupleType(
+            neff=1.5,
[…]
