### 2025-09-25 aa16855
added argmax into unit, added test
