### 2025-09-25 7e5cbf4
Fixed the way period is obtained, fixed JIT compat, also formatting
