+++ src/fdtdx/fdtd/fdtd.py
@@ -8,10 +8,10 @@ from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.backward import backward
 from fdtdx.fdtd.container import ArrayContainer, ObjectContainer, SimulationState, reset_array_container
 from fdtdx.fdtd.forward import forward, forward_single_args_wrapper
+from fdtdx.fdtd.stop_conditions import StoppingCondition, TimeStepCondition
 from fdtdx.interfaces.state import RecordingState
 from fdtdx.objects.boundaries.boundary import BaseBoundaryState
 from fdtdx.objects.detectors.detector import DetectorState
-from fdtdx.fdtd.stop_conditions import StoppingCondition, TimeStepCondition
 
 
 def reversible_fdtd(
