### 2025-09-26 6805bc9
Fix method call to get frequency in PhasorDetector
