### 2025-10-01 a7476d3
Adjust function behavior and add tests
