+++ src/fdtdx/units/unitful.py
@@ -447,7 +447,7 @@ class Unitful(TreeClass):
     def prod(self, **kwargs) -> "Unitful":
         return prod(self, **kwargs)
     
-    def argmax(self, **kwargs) -> np.number | np.ndarray | jax.Array:
+    def argmax(self, **kwargs) -> "Unitful":
         return argmax(self, **kwargs)
         
 
@@ -1211,7 +1211,7 @@ def pow(x, y):  # type: ignore
 
 
 ## min #######################################
-def unary_fn(x: Unitful, op_str: str, *args, **kwargs) -> Unitful | ArrayLike:
+def unary_fn(x: Unitful, op_str: str, *args, **kwargs) -> Unitful:
     # TODO: handling of numpy arrays
     orig_fn = getattr(jax.numpy, f"_orig_{op_str}")
     new_val = orig_fn(x.val, *args, **kwargs)
@@ -1223,12 +1223,12 @@ def unary_fn(x: Unitful, op_str: str, *args, **kwargs) -> Unitful | ArrayLike:
         if x_arr is not None:
             np_fn = getattr(np, f"{op_str}")
             new_static_arr = np_fn(x_arr, *args, **kwargs)
-    if not output_unitful_for_array(new_static_arr):
-        return new_val
+    if new_val.dtype not in PHYSICAL_DTYPES:
+        return Unitful(val=new_val, unit=EMPTY_UNIT, static_arr=new_static_arr)
     return Unitful(val=new_val, unit=x.unit, static_arr=new_static_arr)
 
 @overload
-def min(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "min", *args, **kwargs) # type: ignore
+def min(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "min", *args, **kwargs)
 
 @overload
 def min(x: PhysicalArrayLike, *args, **kwargs) -> jax.Array:
@@ -1245,7 +1245,7 @@ def min(x, *args, **kwargs):  # type: ignore
 
 ## max #######################################
 @overload
-def max(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "max", *args, **kwargs) # type: ignore
+def max(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "max", *args, **kwargs)
 
 @overload
 def max(x: PhysicalArrayLike, *args, **kwargs) -> jax.Array:
@@ -1262,7 +1262,7 @@ def max(x, *args, **kwargs):  # type: ignore
 
 ## mean #######################################
 @overload
-def mean(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "mean", *args, **kwargs) # type: ignore
+def mean(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "mean", *args, **kwargs)
 
 @overload
 def mean(x: PhysicalArrayLike, *args, **kwargs) -> jax.Array:
@@ -1279,7 +1279,7 @@ def mean(x, *args, **kwargs):  # type: ignore
 
 ## sum #######################################
 @overload
-def sum(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "sum", *args, **kwargs) # type: ignore
+def sum(x: Unitful, *args, **kwargs) -> Unitful: return unary_fn(x, "sum", *args, **kwargs)
 
 @overload
 def sum(x: PhysicalArrayLike, *args, **kwargs) -> jax.Array:
@@ -1304,7 +1304,7 @@ def prod(x: Unitful, *args, **kwargs) -> Unitful:
     new_static_arr = x.static_value()
     new_val = x.value()
[…]
