### 2025-10-01 208d53e
added same_position wrapper for place_at_center, modified test_placement.py to use same_position instead of place_at_center
