### 2025-10-01 3838872
Implement handling of numpy arrays in unary_fn for Unitful
