+++ src/fdtdx/units/unitful.py
@@ -1212,7 +1212,17 @@ def pow(x, y):  # type: ignore
 
 ## min #######################################
 def unary_fn(x: Unitful, op_str: str, *args, **kwargs) -> Unitful:
-    # TODO: handling of numpy arrays
+    # handling of numpy arrays
+    if isinstance(x.val, np.ndarray | np.number):
+        np_fn = getattr(np, f"{op_str}")
+        new_val = np_fn(x.val, *args, **kwargs)
+        if not isinstance(new_val, np.ndarray) and not isinstance(new_val, np.number):
+            raise Exception(f"This is an internal error: {op_str} produced {type(new_val)}")
+        if new_val.dtype not in PHYSICAL_DTYPES:
+            return Unitful(val=new_val, unit=EMPTY_UNIT, static_arr=None)
+        return Unitful(val=new_val, unit=x.unit, static_arr=None)
+    
+    # handling of other jax array or other types
     orig_fn = getattr(jax.numpy, f"_orig_{op_str}")
     new_val = orig_fn(x.val, *args, **kwargs)
     if not isinstance(new_val, jax.Array):
