### 2025-10-01 dfae408
added non isotropic permetivity, permeability, and conductivity
