+++ src/fdtdx/fdtd/initialization.py
@@ -145,17 +145,26 @@ def apply_params(
     # apply parameter to devices
     for device in objects.devices:
         cur_material_indices = device(params[device.name], expand_to_sim_grid=True, **transform_kwargs)
+        # allowed_perm_list is now list of tuples: [(εx, εy, εz), ...]
         allowed_perm_list = compute_allowed_permittivities(device.materials)
-        if device.output_type == ParameterType.CONTINUOUS:
-            first_term = (1 - cur_material_indices) * (1 / allowed_perm_list[0])
-            second_term = cur_material_indices * (1 / allowed_perm_list[1])
-            new_perm_slice = first_term + second_term
-        else:
-            new_perm_slice = jnp.asarray(allowed_perm_list)[cur_material_indices.astype(jnp.int32)]
-            new_perm_slice = straight_through_estimator(cur_material_indices, new_perm_slice)
-            new_perm_slice = 1 / new_perm_slice
-        new_perm = arrays.inv_permittivities.at[*device.grid_slice].set(new_perm_slice)
-        arrays = arrays.at["inv_permittivities"].set(new_perm)
+        allowed_perm_array = jnp.asarray(allowed_perm_list)  # shape: (num_materials, 3)
+
+        # Process each component separately
+        for i in range(3):
+            if device.output_type == ParameterType.CONTINUOUS:
+                # Linear interpolation between two materials for component i
+                first_term = (1 - cur_material_indices) * (1.0 / allowed_perm_array[0, i])
+                second_term = cur_material_indices * (1.0 / allowed_perm_array[1, i])
+                new_perm_slice = first_term + second_term
+            else:
+                # Discrete material selection for component i
+                component_values = allowed_perm_array[:, i][cur_material_indices.astype(jnp.int32)]
+                component_values = straight_through_estimator(cur_material_indices, component_values)
+                new_perm_slice = 1 / component_values
+
+            # Update component i of inv_permittivities array
+            new_perm = arrays.inv_permittivities.at[i, *device.grid_slice].set(new_perm_slice)
+            arrays = arrays.at["inv_permittivities"].set(new_perm)
 
     # apply random key to sources
     new_objects = []
@@ -213,8 +222,9 @@ def _init_arrays(
         backend=config.backend,
     )
 
-    # permittivity
-    shape_params = volume_shape
+    # permittivity - shape (3, Nx, Ny, Nz) for non-isotropic materials
+    # First dimension is for (x, y, z) components
+    shape_params = (3, *volume_shape)
     inv_permittivities = create_named_sharded_matrix(
         shape_params,
         value=0.0,
@@ -223,7 +233,8 @@ def _init_arrays(
         backend=config.backend,
     )
 
-    # permeability
+    # permeability - shape (3, Nx, Ny, Nz) for non-isotropic materials
+    # Use scalar 1.0 only if all materials are non-magnetic AND isotropic
     if objects.all_objects_non_magnetic:
         inv_permeabilities = 1.0
     else:
@@ -235,7 +246,7 @@ def _init_arrays(
             backend=config.backend,
         )
 
-    # electric conductivity
+    # electric conductivity - shape (3, Nx, Ny, Nz) for non-isotropic materials
     electric_conductivity = None
     if not objects.all_objects_non_electrically_conductive:
         electric_conductivity = create_named_sharded_matrix(
@@ -246,7 +257,7 @@ def _init_arrays(
             backend=config.backend,
         )
 
-    # magnetic conductivity
+    # magnetic conductivity - shape (3, Nx, Ny, Nz) for non-isotropic materials
     magnetic_conductivity = None
     if not objects.all_objects_non_magnetically_conductive:
         magnetic_conductivity = create_named_sharded_matrix(
[…]
