+++ src/fdtdx/materials.py
@@ -3,50 +3,124 @@ import math
 from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field
 
 
+def _normalize_material_property(value: float | tuple[float, float, float]) -> tuple[float, float, float]:
+    """Normalize material property to a 3-tuple for (x, y, z) components.
+
+    Args:
+        value: Either a scalar (isotropic) or 3-tuple (non-isotropic) material property
+
+    Returns:
+        tuple[float, float, float]: Material property as (x, y, z) components
+    """
+    if isinstance(value, tuple):
+        if len(value) != 3:
+            raise ValueError(f"Material property tuple must have exactly 3 elements, got {len(value)}")
+        return value
+    else:
+        # Isotropic: broadcast scalar to all three components
+        return (value, value, value)
+
+
 @autoinit
 class Material(TreeClass):
     """
     Represents an electromagnetic material with specific electrical and magnetic properties.
 
     This class stores the fundamental electromagnetic properties of a material for use
-    in electromagnetic simulations.
+    in electromagnetic simulations. Supports both isotropic and non-isotropic (anisotropic) materials.
+
+    Note:
+        All material properties are stored internally as 3-tuples (x, y, z components).
+        Scalar inputs are automatically broadcast to all three components.
 
     """
 
     #: The relative permittivity (dielectric constant) of the material, which describes how the electric field is affected by the material.
-    #: Higher values indicate greater electric polarization in response to an applied electric field. Defaults to 1.0.
-    permittivity: float = frozen_field(default=1.0)
+    #: Higher values indicate greater electric polarization in response to an applied electric field.
+    #: For isotropic materials, provide a scalar float. For non-isotropic materials, provide a tuple of 3 floats (εx, εy, εz).
+    #: Stored internally as a 3-tuple. Defaults to (1.0, 1.0, 1.0).
+    permittivity: tuple[float, float, float] = frozen_field(
+        default=(1.0, 1.0, 1.0),
+        on_setattr=[_normalize_material_property],
+    )
 
     #: The relative permeability of the material, which describes how the magnetic field is affected by the material.
     #: Higher values indicate greater magnetic response to an applied magnetic field.
-    #: Defaults to 1.0 (non-magnetic material).
-    permeability: float = frozen_field(default=1.0)
+    #: For isotropic materials, provide a scalar float. For non-isotropic materials, provide a tuple of 3 floats (μx, μy, μz).
+    #: Stored internally as a 3-tuple. Defaults to (1.0, 1.0, 1.0).
+    permeability: tuple[float, float, float] = frozen_field(
+        default=(1.0, 1.0, 1.0),
+        on_setattr=[_normalize_material_property],
+    )
 
     #: The electrical conductivity of the material in siemens per meter (S/m), which describes how easily electric current can flow through it.
     #: Higher values indicate materials that conduct electricity more easily.
-    #: Defaults to 0.0 (perfect insulator).
-    electric_conductivity: float = frozen_field(default=0.0)
+    #: For isotropic materials, provide a scalar float. For non-isotropic materials, provide a tuple of 3 floats (σx, σy, σz).
+    #: Stored internally as a 3-tuple. Defaults to (0.0, 0.0, 0.0).
+    electric_conductivity: tuple[float, float, float] = frozen_field(
+        default=(0.0, 0.0, 0.0),
+        on_setattr=[_normalize_material_property],
+    )
 
     #: The magnetic conductivity, or magnetic loss of the material.
     #: This is an artificial parameter for numerical applications and does not represent an actual physical unit,
[…]
