### 2025-10-01 babc82a
Custom stopping condition implementation, along with tests
