+++ tests/fdtd/test_condition.py
@@ -1,0 +1,356 @@
+import jax
+import jax.numpy as jnp
+import pytest
+
+from fdtdx.config import SimulationConfig
+from fdtdx.core.wavelength import WaveCharacter
+from fdtdx.fdtd.container import ArrayContainer
+from fdtdx.fdtd.initialization import place_objects
+
+# Import the module to test
+from fdtdx.fdtd.stop_conditions import DetectorConvergenceCondition, EnergyThresholdCondition, TimeStepCondition
+from fdtdx.objects.boundaries.boundary import BaseBoundaryState
+from fdtdx.objects.detectors.detector import DetectorState
+from fdtdx.objects.detectors.energy import EnergyDetector
+from fdtdx.objects.static_material.static import SimulationVolume
+
+
+class TestCondition:
+    @pytest.fixture
+    def setup_simulation_state(self):
+        """Set up a basic simulation state for testing."""
+        # Create dummy arrays with appropriate shapes
+        E = jnp.ones((3, 10, 10, 10))  # 3D electric field
+        H = jnp.ones((3, 10, 10, 10))  # 3D magnetic field
+        inv_permittivities = jnp.ones((10, 10, 10))
+        inv_permeabilities = jnp.ones((10, 10, 10))
+
+        # Create mock boundary and detector states
+        boundary_states = {"pml": BaseBoundaryState()}
+        detector_states = {"detector1": DetectorState()}
+
+        # Create array container
+        arrays = ArrayContainer(
+            E=E,
+            H=H,
+            inv_permittivities=inv_permittivities,
+            inv_permeabilities=inv_permeabilities,
+            boundary_states=boundary_states,
+            detector_states=detector_states,
+            recording_state=None,
+            electric_conductivity=None,
+            magnetic_conductivity=None,
+        )
+
+        # Create simulation state
+        time_step = jnp.array(0)
+        state = (time_step, arrays)
+
+        config = SimulationConfig(
+            time=100e-11,
+            resolution=1e-4,
+            courant_factor=0.99,
+        )
+        detector_name = "test_detector"
+        detector = EnergyDetector(name=detector_name)
+        volume = SimulationVolume(partial_real_shape=(1e-3, 1e-3, 1e-3))
+        position_constraints = []
+        position_constraints.extend(detector.same_position_and_size(volume))
+        key = jax.random.PRNGKey(0)
+        objects, arrays, _, config, _ = place_objects(
+            volume=volume,
+            config=config,
+            constraints=position_constraints,
+            key=key,
+        )
+
+        return {"state": state, "config": config, "objects": objects, "dn": detector_name}
+
+    def test_condition_function_signature(self, setup_simulation_state):
+        """Test that stopping condition function has the correct signature and returns expected types."""
+        state = setup_simulation_state["state"]
+        config = setup_simulation_state["config"]
+        objects = setup_simulation_state["objects"]
+        detector_name = setup_simulation_state["dn"]
+        arrays = state[1]
+        cw_source_period = 5e-11  # 20 GHz
+        wave_character = WaveCharacter(period=cw_source_period)
+        threshold = 1e-6
+        min_steps = 1572  # Approximately 5 periods at 20 GHz with dt ~ 1.906e-13 s.
+        # This value ensures min_steps > spp * (prev_periods+1),
+        # otherwise a ValueError is raised
+
+        arrays.detector_states[detector_name] = {"energy": jnp.zeros((config.time_steps_total, 1))}
+
+        # TimeStepCondition
+        time_step_cond = TimeStepCondition()
+        ts_result = time_step_cond(state, config, objects)
+
+        assert isinstance(ts_result, jax.Array)
+        assert ts_result.dtype == jnp.bool_
+        assert ts_result.shape == ()
+
+        # EnergyThresholdCondition
[…]
