### 2025-10-01 153cac0
fixed core fdtdx components to accept anisotropic materials, some cleanup
