+++ src/fdtdx/objects/sources/mode.py
@@ -41,9 +41,11 @@ class ModePlaneSource(TFSFPlaneSource):
         ):
             raise NotImplementedError()
 
-        inv_permittivity_slice = inv_permittivities[*self.grid_slice]
+        # inv_permittivities shape: (3, Nx, Ny, Nz) - slice with component dimension
+        inv_permittivity_slice = inv_permittivities[:, *self.grid_slice]
         if isinstance(inv_permeabilities, jax.Array) and inv_permeabilities.ndim > 0:
-            inv_permeability_slice = inv_permeabilities[*self.grid_slice]
+            # inv_permeabilities shape: (3, Nx, Ny, Nz) - slice with component dimension
+            inv_permeability_slice = inv_permeabilities[:, *self.grid_slice]
         else:
             inv_permeability_slice = inv_permeabilities
 
@@ -69,9 +71,11 @@ class ModePlaneSource(TFSFPlaneSource):
             [round(self.grid_shape[self.horizontal_axis]), round(self.grid_shape[self.vertical_axis])], dtype=jnp.int32
         )
 
-        inv_permittivity_slice = inv_permittivities[*self.grid_slice]
+        # inv_permittivities shape: (3, Nx, Ny, Nz) - slice with component dimension
+        inv_permittivity_slice = inv_permittivities[:, *self.grid_slice]
         if isinstance(inv_permeabilities, jax.Array) and inv_permeabilities.ndim > 0:
-            inv_permeability_slice = inv_permeabilities[*self.grid_slice]
+            # inv_permeabilities shape: (3, Nx, Ny, Nz) - slice with component dimension
+            inv_permeability_slice = inv_permeabilities[:, *self.grid_slice]
         else:
             inv_permeability_slice = inv_permeabilities
 
