### 2025-10-02 f05f3d4
fix init for pylance type errors that might confuse users
