+++ tests/fdtd/test_fdtd.py
@@ -7,7 +7,6 @@ from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 from fdtdx.fdtd.fdtd import checkpointed_fdtd, custom_fdtd_forward, reversible_fdtd
 from fdtdx.objects.object import SimulationObject
 
-
 class DummySimulationObject(SimulationObject):
     """Minimal mock implementation of a SimulationObject."""
 
@@ -21,7 +20,7 @@ class DummySimulationObject(SimulationObject):
 @pytest.fixture
 def dummy_arrays():
     field_shape = (3, 2, 2, 2)  # (components, nx, ny, nz)
-    mat_shape = (2, 2, 2)  # scalar per voxel
+    mat_shape = (3, 2, 2, 2)  # scalar per voxel
 
     return ArrayContainer(
         E=jnp.zeros(field_shape),
@@ -52,7 +51,7 @@ def dummy_config():
     return SimulationConfig(
         time=0.1,
         resolution=1.0,
-        backend="gpu",
+        backend="cpu",
         dtype=jnp.float32,
         courant_factor=0.99,
         gradient_config=None,
@@ -100,7 +99,7 @@ def test_zero_time(dummy_arrays, dummy_objects):
     config = SimulationConfig(
         time=0.0,
         resolution=1.0,
-        backend="gpu",
+        backend="cpu",
         dtype=jnp.float32,
         courant_factor=0.99,
         gradient_config=None,
