### 2025-10-08 ymahlau/fdtdx#159 @victorapm
JOSS review
Hi @ymahlau and collaborators,

just wanted to share a few comments regarding the JOSS review: https://github.com/openjournals/joss-reviews/issues/8912

1. Consider adding a `CONTRIBUTING.md` file in the top-level folder with instructions on external contributions to the package.
2. JOSS Paper:
   1. Statement of need: clarify who is the target audience for `FDTDX`, i.e., who would realistically use, extend, or learn from the software. Some examples: computational scientists, researchers, domain scientists...
   2. Implementation section, 3rd paragraph: consider rewriting statements involving "There exist", "Firstly", "Secondly"
   3. Limitations section: could you elaborate in a few words why simulating non-linear materials is a difficult task

Congratulations for the software!

I'll work on running tests next and will get back to you in case there are further questions/suggestions
