### 2025-10-09 4281a07
Add argmin and argmax to Unitful, fix pre-commit issue
