+++ src/fdtdx/core/fraction.py
@@ -1,20 +1,21 @@
 import math
 from typing import Any, NamedTuple, Union, overload
 
+
 class Fraction(NamedTuple):
     num: int
     denom: int
-    
+
     def __repr__(self) -> str:
         return str(self)
-    
+
     def __str__(self) -> str:
         if self.num == 0:
             return "0"
         if self.denom == 1:
             return str(self.num)
         return f"({self.num}/{self.denom})"
-    
+
     def __eq__(self, other: Any) -> bool:
         if not isinstance(other, int | float | Fraction):
             return False
@@ -30,35 +31,35 @@ class Fraction(NamedTuple):
         if isinstance(other, float):
             # Compare float values (be careful with floating point precision)
             return abs(self.value() - other) < 1e-15
-    
+
     def value(self) -> float:
         return self.num / self.denom
-        
+
     def reduced(self) -> "Fraction":
         if self.num == 0:
             return Fraction(0, 1)
-        
+
         common_divisor = math.gcd(abs(self.num), abs(self.denom))
         reduced_num = self.num // common_divisor
         reduced_denom = self.denom // common_divisor
-        
+
         # Ensure denominator is positive
         if reduced_denom < 0:
             reduced_num = -reduced_num
             reduced_denom = -reduced_denom
-        
+
         return Fraction(reduced_num, reduced_denom)
-    
+
     # Addition
     @overload
     def __add__(self, other: "Fraction") -> "Fraction": ...
-    
+
     @overload
     def __add__(self, other: int) -> "Fraction": ...  # type: ignore
-    
+
     @overload
     def __add__(self, other: float) -> float: ...
-    
+
     def __add__(self, other: Any) -> Union["Fraction", float]:  # type: ignore
         if isinstance(other, Fraction):
             new_num = self.num * other.denom + other.num * self.denom
@@ -69,29 +70,29 @@ class Fraction(NamedTuple):
         elif isinstance(other, float):
             return self.value() + other
         return NotImplemented
-    
+
     @overload
     def __radd__(self, other: "Fraction") -> "Fraction": ...
-    
+
     @overload
     def __radd__(self, other: int) -> "Fraction": ...  # type: ignore
-    
+
     @overload
     def __radd__(self, other: float) -> float: ...
-    
+
     def __radd__(self, other: Any) -> Union["Fraction", float]:
         return self.__add__(other)
-    
+
     # Multiplication
     @overload
     def __mul__(self, other: "Fraction") -> "Fraction": ...
-    
+
     @overload
     def __mul__(self, other: int) -> "Fraction": ...  # type: ignore
-    
+
     @overload
     def __mul__(self, other: float) -> float: ...
-    
+
     def __mul__(self, other: Any) -> Union["Fraction", float]:  # type: ignore
         if isinstance(other, Fraction):
             new_num = self.num * other.num
@@ -102,29 +103,29 @@ class Fraction(NamedTuple):
         elif isinstance(other, float):
             return self.value() * other
         return NotImplemented
-    
+
     @overload
     def __rmul__(self, other: "Fraction") -> "Fraction": ...
-    
+
     @overload
     def __rmul__(self, other: int) -> "Fraction": ...  # type: ignore
-    
+
     @overload
     def __rmul__(self, other: float) -> float: ...
-    
+
     def __rmul__(self, other: Any) -> Union["Fraction", float]:  # type: ignore
         return self.__mul__(other)
-    
+
     # Division
     @overload
     def __truediv__(self, other: "Fraction") -> "Fraction": ...
-    
+
     @overload
     def __truediv__(self, other: int) -> "Fraction": ...  # type: ignore
-    
+
     @overload
     def __truediv__(self, other: float) -> float: ...
-    
+
[…]
