### 2025-10-15 7e6e095
added sign function
