### 2025-10-15 795933b
updated paper.md
