### 2025-10-15 87f41b2
paper.md: minor change for better formatting
