### 2025-10-23 9a82687
Initial support for linear grading of alpha and exposing respective knobs through BoundaryConfig
