+++ src/fdtdx/objects/boundaries/initialization.py
@@ -90,6 +90,42 @@ class BoundaryConfig(TreeClass):
     #: Final kappa value at max z boundary. Default 1.5.
     kappa_end_maxz: float = frozen_field(default=1.5)
 
+    #: Initial alpha value at min x boundary. Default 1e-8.
+    alpha_start_minx: float = frozen_field(default=1e-8)
+
+    #: Final alpha value at min x boundary. Default 1e-8.
+    alpha_end_minx: float = frozen_field(default=1e-8)
+
+    #: Initial alpha value at max x boundary. Default 1e-8.
+    alpha_start_maxx: float = frozen_field(default=1e-8)
+
+    #: Final alpha value at max x boundary. Default 1e-8.
+    alpha_end_maxx: float = frozen_field(default=1e-8)
+
+    #: Initial alpha value at min y boundary. Default 1e-8.
+    alpha_start_miny: float = frozen_field(default=1e-8)
+
+    #: Final alpha value at min y boundary. Default 1e-8.
+    alpha_end_miny: float = frozen_field(default=1e-8)
+
+    #: Initial alpha value at max y boundary. Default 1e-8.
+    alpha_start_maxy: float = frozen_field(default=1e-8)
+
+    #: Final alpha value at max y boundary. Default 1e-8.
+    alpha_end_maxy: float = frozen_field(default=1e-8)
+
+    #: Initial alpha value at min z boundary. Default 1e-8.
+    alpha_start_minz: float = frozen_field(default=1e-8)
+
+    #: Final alpha value at min z boundary. Default 1e-8.
+    alpha_end_minz: float = frozen_field(default=1e-8)
+
+    #: Initial alpha value at max z boundary. Default 1e-8.
+    alpha_start_maxz: float = frozen_field(default=1e-8)
+
+    #: Final alpha value at max z boundary. Default 1e-8.
+    alpha_end_maxz: float = frozen_field(default=1e-8)
+
     def get_dict(self) -> dict[str, int]:
         """Gets a dictionary mapping boundary names to their grid thicknesses.
 
@@ -160,6 +196,44 @@ class BoundaryConfig(TreeClass):
         else:
             raise Exception(f"Unknown: {prop=}")
 
+    def get_alpha_dict(
+        self,
+        prop: Literal["alpha_start", "alpha_end"],
+    ) -> dict[str, float]:
+        """Gets a dictionary mapping boundary names to their kappa values.
+
+        Args:
+            prop (Literal["alpha_start", "alpha_end"]): Which alpha property to get,
+                either "alpha_start" or "alpha_end".
+
+        Returns:
+            dict[str, float]: Dictionary with keys 'min_x', 'max_x', 'min_y', 'max_y', 'min_z', 'max_z'
+                mapping to their respective alpha values.
+
+        Raises:
+            Exception: If prop is not "alpha_start" or "alpha_end"
+        """
+        if prop == "alpha_start":
+            return {
+                "min_x": self.alpha_start_minx,
+                "max_x": self.alpha_start_maxx,
+                "min_y": self.alpha_start_miny,
+                "max_y": self.alpha_start_maxy,
+                "min_z": self.alpha_start_minz,
+                "max_z": self.alpha_start_maxz,
+            }
+        elif prop == "alpha_end":
+            return {
+                "min_x": self.alpha_end_minx,
+                "max_x": self.alpha_end_maxx,
+                "min_y": self.alpha_end_miny,
+                "max_y": self.alpha_end_maxy,
+                "min_z": self.alpha_end_minz,
+                "max_z": self.alpha_end_maxz,
+            }
+        else:
+            raise Exception(f"Unknown: {prop=}")
+
[…]
