### 2025-10-23 d8d1162
Adjust Unitful comparison operator
