+++ src/fdtdx/units/unitful.py
@@ -32,6 +32,22 @@ like jnp.ones will return a Unitful instead of Array for scale optimization.
 """
 STATIC_OPTIM_STOP_FLAG = False
 
+"""
+Relative tolerance for floating point comparisons
+TODO: 
+    adjust RTOL_COMPARISON and ATOL_COMPARISON dynamically
+    based on the numeric precision (e.g., float16, float32, float64).
+    For now, these fixed values are sufficient for most use cases.
+
+| Precision Type | RTOL (Relative Tolerance) | ATOL (Absolute Tolerance) |
+| -------------- | ------------------------- | ------------------------- |
+| `float16`      | 1e-2 - 1e-3               | 1e-3 - 1e-4               |
+| `float32`      | 1e-4 - 1e-5               | 1e-6 - 1e-7               |
+| `float64`      | 1e-6 - 1e-8               | 1e-9 - 1e-12              | 
+"""
+RTOL_COMPARSION = 1e-5
+ATOL_COMPARSION = 1e-7
+
 
 @autoinit
 class Unit(TreeClass):
@@ -338,6 +354,15 @@ class Unitful(TreeClass):
         except Exception:
             return f"Unitful [{self.unit}]: {self.shape}"
 
+    def __bool__(self) -> bool:
+        if isinstance(self.val, (bool, np.bool_, jnp.bool_)):
+            return bool(self.val)
+        if isinstance(self.val, (np.ndarray, jnp.ndarray)):
+            if self.val.size != 1:
+                raise ValueError("Truth value of an array with multiple elements is ambiguous")
+            return bool(self.val.item())
+        return bool(self.val)
+
     def __repr__(self) -> str:
         return str(self)
 
@@ -365,22 +390,22 @@ class Unitful(TreeClass):
     def __rsub__(self, other: "Unitful | PhysicalArrayLike") -> "Unitful":
         return subtract(other, self)
 
-    def __lt__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
+    def __lt__(self, other: "Unitful | RealPhysicalArrayLike") -> "Unitful":
         return lt(self, other)
 
-    def __le__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
+    def __le__(self, other: "Unitful | RealPhysicalArrayLike") -> "Unitful":
         return le(self, other)
 
-    def __eq__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
+    def __eq__(self, other: "Unitful | RealPhysicalArrayLike") -> "Unitful":
         return eq(self, other)
 
-    def __ne__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:  # type: ignore
+    def __ne__(self, other: "Unitful | RealPhysicalArrayLike") -> "Unitful":  # type: ignore[override]  # allow non-bool return for array-style comparison
         return ne(self, other)
 
-    def __ge__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
+    def __ge__(self, other: "Unitful | RealPhysicalArrayLike") -> "Unitful":
         return ge(self, other)
 
-    def __gt__(self, other: "Unitful | RealPhysicalArrayLike") -> np.bool | np.ndarray | jax.Array | bool:
[…]
