+++ src/fdtdx/objects/sources/source.py
@@ -157,8 +157,7 @@ class HardConstantAmplitudePlanceSource(DirectionalPlaneSourceBase):
             return E
         delta_t = self._config.time_step_duration
         time_phase = (
-            2 * jnp.pi * time_step * delta_t / self.wave_character.get_period() 
-            + self.wave_character.phase_shift
+            2 * jnp.pi * time_step * delta_t / self.wave_character.get_period() + self.wave_character.phase_shift
         )
         magnitude = ff.real(self.amplitude * jnp.exp(-1j * time_phase))
         magnitude = magnitude * self.static_amplitude_factor
@@ -185,9 +184,8 @@ class HardConstantAmplitudePlanceSource(DirectionalPlaneSourceBase):
         if inverse:
             return H
         delta_t = self._config.time_step_duration
-        time_phase =    (
-            2 * jnp.pi * time_step * delta_t / self.wave_character.get_period() 
-            + self.wave_character.phase_shift
+        time_phase = (
+            2 * jnp.pi * time_step * delta_t / self.wave_character.get_period() + self.wave_character.phase_shift
         )
         magnitude = ff.real(self.amplitude * jnp.exp(-1j * time_phase))
         magnitude = magnitude * self.static_amplitude_factor
