### 2025-10-29 4997bdb
refactoring object placement code #167
