### 2025-10-30 bc11a4c
fix: typo "capabilitiy"
