### 2025-11-01 13ebbcc
add unit-trigonometirc methods
