+++ src/fdtdx/functional/numpy.py
@@ -1,5 +1,6 @@
 # ruff: noqa: F811
 
+import cmath
 import math
 from typing import Sequence
 
@@ -587,15 +588,36 @@ def imag(  # type: ignore
 
 ## sin #####################################
 @overload
-def sin(x: Unitful) -> Unitful:
-    new_val = jnp._orig_sin(x.val)  # type: ignore
+def sin(x: Unitful, *args, **kwargs) -> Unitful:
+    # check unit
+    if x.unit.dim != {}:
+        raise ValueError("Input to sin function must be dimensionless")
+
+    if (
+        isinstance(x.val, (bool, np.bool_))
+        or (isinstance(x.val, np.ndarray) and x.val.dtype == np.bool_)
+        or (isinstance(x.val, jax.Array) and x.val.dtype == jnp.bool_)
+    ):
+        raise ValueError("Function sin does not support boolean inputs.")
+
+    x_val_nonscale = x.val * (10**x.unit.scale)
+    if isinstance(x.val, jax.Array):
+        new_val = jnp._orig_sin(x_val_nonscale)  # type: ignore
+    elif isinstance(x.val, np.ndarray | np.number):
+        new_val = np.sin(x_val_nonscale, *args, **kwargs)
+    elif isinstance(x.val, complex):
+        new_val = cmath.sin(x_val_nonscale)
+    else:  # only int, float
+        new_val = math.sin(x_val_nonscale)
+
     # static computation
     new_static_arr = None
     if is_traced(new_val):
         x_arr = get_static_operand(x)
         if x_arr is not None:
-            new_static_arr = np.sin(x_arr)
-    return Unitful(val=new_val, unit=x.unit, static_arr=new_static_arr)
+            x_arr_nonscale = x_arr * (10**x.unit.scale)
+            new_static_arr = np.sin(x_arr_nonscale)
+    return Unitful(val=new_val, unit=EMPTY_UNIT, static_arr=new_static_arr)
 
 
 @overload
@@ -604,8 +626,13 @@ def sin(x: int | float) -> float:
 
 
 @overload
-def sin(x: np.ndarray | np.number) -> np.ndarray:
-    return np.sin(x)
+def sin(x: complex) -> complex:
+    return cmath.sin(x)
+
+
+@overload
+def sin(x: np.ndarray | np.number, *args, **kwargs) -> np.ndarray:
+    return np.sin(x, *args, **kwargs)
 
 
 @overload
@@ -614,23 +641,42 @@ def sin(x: jax.Array) -> jax.Array:
 
 
 @dispatch
-def sin(  # type: ignore
-    x,
-):
-    del x
+def sin(x, *args, **kwargs):  # type: ignore
+    del x, args, kwargs
     raise NotImplementedError()
 
 
 ## cos #####################################
 @overload
-def cos(x: Unitful) -> Unitful:
-    new_val = jnp._orig_cos(x.val)  # type: ignore
+def cos(x: Unitful, *args, **kwargs) -> Unitful:
+    # check unit
+    if x.unit.dim != {}:
+        raise ValueError("Input to sin function must be dimensionless")
+
+    if (
+        isinstance(x.val, (bool, np.bool_))
+        or (isinstance(x.val, np.ndarray) and x.val.dtype == np.bool_)
+        or (isinstance(x.val, jax.Array) and x.val.dtype == jnp.bool_)
+    ):
+        raise ValueError("Function cos does not support boolean inputs.")
+
+    x_val_nonscale = x.val * (10**x.unit.scale)
+
+    if isinstance(x.val, jax.Array):
+        new_val = jnp._orig_cos(x_val_nonscale)  # type: ignore
+    elif isinstance(x.val, np.ndarray | np.number):
[…]
