+++ tests/functional/test_numpy.py
@@ -1,13 +1,15 @@
+import cmath
 import math
 
 import jax
 import jax.numpy as jnp
+import numpy as np
 import pytest
 
 from fdtdx.core.fraction import Fraction
 from fdtdx.functional.numpy import roll, sqrt, transpose
 from fdtdx.units.typing import SI
-from fdtdx.units.unitful import Unit, Unitful
+from fdtdx.units.unitful import EMPTY_UNIT, Unit, Unitful
 
 
 def test_sqrt_unitful_even_integer_dimensions():
@@ -810,3 +812,304 @@ def test_imag_jax_array():
     assert not jnp.iscomplexobj(result)
     # Shape should be preserved: (3, 3)
     assert result.shape == (3, 3)
+
+
+def test_sin_unitful_with_Jax_array():
+    """Test sin with a Unitful whose val is a Jax array"""
+    rad_unit = EMPTY_UNIT
+    angles = Unitful(val=jnp.array([0.0, jnp.pi / 4]), unit=rad_unit)
+    result = jnp.sin(angles, dtype=jnp.float64)  # type: ignore
+
+    assert isinstance(result, Unitful)
+    assert isinstance(result.val, jnp.ndarray)
+    assert jnp.issubdtype(result.val.dtype, jnp.float32)
+
+    expected_vals = jnp.array([0.0, math.sin(math.pi / 4)])
+    assert jnp.allclose(result.val, expected_vals)
+
+
+def test_sin_unitful_with_numpy_array():
+    """Test sin with a Unitful whose val is a Numpy array (angle in radians)"""
+    rad_unit = EMPTY_UNIT
+    angles = Unitful(val=np.array([0.0, np.pi / 2]), unit=rad_unit)
+    result = jnp.sin(angles, dtype=np.float32)  # type: ignore
+
+    assert isinstance(result, Unitful)
+    assert isinstance(result.val, np.ndarray)
+    assert np.issubdtype(result.val.dtype, np.number)
+
+    expected_vals = np.array([0.0, math.sin(math.pi / 2)])
+    assert np.allclose(result.val, expected_vals)
+
+
+def test_sin_unitful_with_Python_Scalar():
+    """Test sin with a Unitful whose val is a float (angle in radians)"""
+    rad_unit = EMPTY_UNIT
+    angles = Unitful(val=math.pi / 2, unit=rad_unit)
+    result = jnp.sin(angles)  # type: ignore
+
+    assert isinstance(result, Unitful)
+    assert isinstance(result.val, float)
+
+    expected_vals = math.sin(math.pi / 2)
+    assert math.isclose(result.val, expected_vals)
+
+
+def test_sin_unitful_with_jax_array_physical_phase():
+    """Test sin with a Unitful whose val is a JAX array (angle in radians with physical units)"""
+    # --- Define units ---
+    m_unit = Unit(scale=0, dim={SI.m: 1})
+    s_unit = Unit(scale=0, dim={SI.s: 1})
+    GHz_unit = Unit(scale=9, dim={SI.s: -1})
+    ns_unit = Unit(scale=-9, dim={SI.s: 1})
+
+    m = Unitful(val=jnp.array([1.0]), unit=m_unit)
+    s = Unitful(val=jnp.array([1.0]), unit=s_unit)
+    GHz = Unitful(val=jnp.array([1.0]), unit=GHz_unit)
+    ns = Unitful(val=jnp.array([1.0]), unit=ns_unit)
+
+    # --- Physical constants ---
+    c = 3.0e8 * m / s
+    f = 1.0 * GHz
+    λ = c / f
+    k = 2 * jnp.pi / λ
+    ω = 2 * jnp.pi * f
+
+    x = 0.15 * m
+    t = 0.25 * ns
+
+    phase = k * x - ω * t
+
[…]
