### 2025-11-02 d10291c
refactoring place_objects and _resolve_object_constraints for #167
