### 2025-11-02 0c84bd4
Replacing jnp.exp for jnp.expm1, updating fields according to literature
