+++ src/fdtdx/objects/boundaries/perfectly_matched_layer.py
@@ -38,6 +38,12 @@ class PMLBoundaryState(BaseBoundaryState):
     #: Auxiliary field for Hz component
     psi_Hz: jax.Array
 
+    #: Derivatives of magnetic field H at last time step
+    dH_for_E: jax.Array
+
+    #: Derivatives of electric field E at last time step
+    dE_for_H: jax.Array
+
     #: Electric field scaling coefficient
     bE: jax.Array
 
@@ -127,11 +133,19 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
             dtype=dtype,
         )
 
-        bE = jnp.exp(-self._config.courant_number * (sigma_E / kappa + alpha))
-        bH = jnp.exp(-self._config.courant_number * (sigma_H / kappa + alpha))
+        theta_E = self._config.courant_number * (sigma_E / kappa + alpha)
+        theta_H = self._config.courant_number * (sigma_H / kappa + alpha)
 
-        cE = (bE - 1) * sigma_E / (sigma_E * kappa + kappa**2 * alpha)
-        cH = (bH - 1) * sigma_H / (sigma_H * kappa + kappa**2 * alpha)
+        bE = jnp.exp(-theta_E)
+        bH = jnp.exp(-theta_H)
+
+        den_E = sigma_E + (kappa * alpha)
+        den_H = sigma_H + (kappa * alpha)
+        frac_E = sigma_E / den_E
+        frac_H = sigma_H / den_H
+
+        cE = (frac_E / kappa) * jnp.expm1(-theta_E)
+        cH = (frac_H / kappa) * jnp.expm1(-theta_H)
 
         return dtype, bE, bH, cE, cH, kappa
 
@@ -141,6 +155,7 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
     ) -> PMLBoundaryState:
         dtype, bE, bH, cE, cH, kappa = self._get_dtype_update_coefficients()
         ext_shape = (3,) + self.grid_shape
+        d2_shape = (3, 2) + self.grid_shape
 
         boundary_state = PMLBoundaryState(
             psi_Ex=jnp.zeros(shape=ext_shape, dtype=dtype),
@@ -149,6 +164,8 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
             psi_Hx=jnp.zeros(shape=ext_shape, dtype=dtype),
             psi_Hy=jnp.zeros(shape=ext_shape, dtype=dtype),
             psi_Hz=jnp.zeros(shape=ext_shape, dtype=dtype),
+            dH_for_E=jnp.zeros(shape=d2_shape, dtype=dtype),
+            dE_for_H=jnp.zeros(shape=d2_shape, dtype=dtype),
             bE=bE.astype(dtype),
             bH=bH.astype(dtype),
             cE=cE.astype(dtype),
@@ -168,6 +185,8 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
             psi_Hx=state.psi_Hx * 0,
             psi_Hy=state.psi_Hy * 0,
             psi_Hz=state.psi_Hz * 0,
+            dH_for_E=state.dH_for_E * 0,
+            dE_for_H=state.dE_for_H * 0,
             bE=bE.astype(dtype),
             bH=bH.astype(dtype),
             cE=cE.astype(dtype),
[…]
