### 2025-11-03 2309071
Formatting, properly using each component of kappa
