+++ src/fdtdx/objects/boundaries/perfectly_matched_layer.py
@@ -236,23 +236,23 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
             * (boundary_state.cE[1, :, 1:, :] if self.axis == 1 else boundary_state.cE[1])
         )
 
-        # Directional partial derivatives for Ex: (+) d_y Hz, (-) d_z Hy
         dY_Hz = jnp.zeros_like(Hx).at[:, 1:, :].set(Hz[:, 1:, :] - Hz[:, :-1, :])
         dZ_Hy = jnp.zeros_like(Hx).at[:, :, 1:].set(Hy[:, :, 1:] - Hy[:, :, :-1])
-        
-        # For Ey: (+) d_z Hx, (-) d_x Hz
+
         dZ_Hx = jnp.zeros_like(Hy).at[:, :, 1:].set(Hx[:, :, 1:] - Hx[:, :, :-1])
         dX_Hz = jnp.zeros_like(Hy).at[1:, :, :].set(Hz[1:, :, :] - Hz[:-1, :, :])
-        
-        # For Ez: (+) d_x Hy, (-) d_y Hx
+
         dX_Hy = jnp.zeros_like(Hz).at[1:, :, :].set(Hy[1:, :, :] - Hy[:-1, :, :])
         dY_Hx = jnp.zeros_like(Hz).at[:, 1:, :].set(Hx[:, 1:, :] - Hx[:, :-1, :])
-        
-        dH_for_E = jnp.stack([
-            jnp.stack([dY_Hz, dZ_Hy], axis=0),
-            jnp.stack([dZ_Hx, dX_Hz], axis=0),
-            jnp.stack([dX_Hy, dY_Hx], axis=0),
-        ], axis=0)
+
+        dH_for_E = jnp.stack(
+            [
+                jnp.stack([dY_Hz, dZ_Hy], axis=0),
+                jnp.stack([dZ_Hx, dX_Hz], axis=0),
+                jnp.stack([dX_Hy, dY_Hx], axis=0),
+            ],
+            axis=0,
+        )
 
         boundary_state = boundary_state.at["psi_Ex"].set(psi_Ex)
         boundary_state = boundary_state.at["psi_Ey"].set(psi_Ey)
@@ -302,23 +302,23 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
             * (boundary_state.cH[1, :, :-1, :] if self.axis == 1 else boundary_state.cH[1])
         )
 
-        # Hx needs (+) d_y Ez, (-) d_z Ey
         dY_Ez = jnp.zeros_like(Ex).at[:, 1:, :].set(Ez[:, 1:, :] - Ez[:, :-1, :])
         dZ_Ey = jnp.zeros_like(Ex).at[:, :, 1:].set(Ey[:, :, 1:] - Ey[:, :, :-1])
-        
-        # Hy needs (+) d_z Ex, (-) d_x Ez
+
         dZ_Ex = jnp.zeros_like(Ey).at[:, :, 1:].set(Ex[:, :, 1:] - Ex[:, :, :-1])
         dX_Ez = jnp.zeros_like(Ey).at[1:, :, :].set(Ez[1:, :, :] - Ez[:-1, :, :])
-        
-        # Hz needs (+) d_x Ey, (-) d_y Ex
+
         dX_Ey = jnp.zeros_like(Ez).at[1:, :, :].set(Ey[1:, :, :] - Ey[:-1, :, :])
         dY_Ex = jnp.zeros_like(Ez).at[:, 1:, :].set(Ex[:, 1:, :] - Ex[:, :-1, :])
-        
-        dE_for_H = jnp.stack([
[…]
