### 2025-11-03 ef4b728
Properly dealing with other directions in kappa and alpha, guarding against NaNs
