### 2025-11-04 103b48a
adjusted grid coordinate to centre and updated test
