+++ src/fdtdx/fdtd/initialization.py
@@ -21,22 +21,20 @@ from fdtdx.objects.object import (
     SizeConstraint,
     SizeExtensionConstraint,
 )
-from fdtdx.objects.static_material.static import StaticMultiMaterialObject, UniformMaterialObject
+from fdtdx.objects.static_material.static import SimulationVolume, StaticMultiMaterialObject, UniformMaterialObject
 from fdtdx.typing import SliceTuple3D
 
 
 def place_objects(
-    objects: list["SimulationObject"],
-    config: "SimulationConfig",
-    constraints: Sequence[
-        ("PositionConstraint" | "SizeConstraint" | "SizeExtensionConstraint" | "GridCoordinateConstraint")
-    ],
-    key: "jax.Array",
+    objects: list[SimulationObject],
+    config: SimulationConfig,
+    constraints: Sequence[(PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint)],
+    key: jax.Array,
 ) -> tuple[
-    "ObjectContainer",
-    "ArrayContainer",
-    "ParameterContainer",
-    "SimulationConfig",
+    ObjectContainer,
+    ArrayContainer,
+    ParameterContainer,
+    SimulationConfig,
     dict[str, Any],
 ]:
     """Places simulation objects according to specified constraints and initializes containers.
@@ -74,9 +72,16 @@ def place_objects(
 
     # Step 3: Convert name → object for placement
     object_map = {obj.name: obj for obj in objects}
-    volume_obj = next((o for o in objects if o.name.lower() == "volume"), None)
-    if not volume_obj:
-        raise ValueError("No volume object found in objects list.")
+    volume_objects = [o for o in objects if isinstance(o, SimulationVolume)]
+    if not volume_objects:
+        raise ValueError("No SimulationVolume object found in the provided objects list.")
+    elif len(volume_objects) > 1:
+        raise ValueError(
+            f"Multiple SimulationVolume objects found ({[o.name for o in volume_objects]}). "
+            "There must be exactly one simulation volume."
+        )
+
+    volume_obj = volume_objects[0]
 
     # Step 4: Place objects on grid based on resolved slice tuples
     placed_objects = []
@@ -326,7 +331,7 @@ def _init_arrays(
     if config.gradient_config is not None and config.gradient_config.recorder is not None:
         input_shape_dtypes = {}
         for boundary in objects.pml_objects:
-            cur_shape = boundary.interface_grid_shape()
+            cur_shape = boundary.boundary_interface_grid_shape()
             extended_shape = (3, *cur_shape)
             input_shape_dtypes[f"{boundary.name}_E"] = jax.ShapeDtypeStruct(shape=extended_shape, dtype=config.dtype)
             input_shape_dtypes[f"{boundary.name}_H"] = jax.ShapeDtypeStruct(shape=extended_shape, dtype=config.dtype)
@@ -378,12 +383,10 @@ def _init_params(
 
 
 def _resolve_object_constraints(
-    objects: list["SimulationObject"],
-    constraints: Sequence[
-        "PositionConstraint" | "SizeConstraint" | "SizeExtensionConstraint" | "GridCoordinateConstraint"
-    ],
-    config: "SimulationConfig",
-) -> Tuple[dict[str, "SliceTuple3D" | None], dict[str, Optional[str]]]:
+    objects: list[SimulationObject],
+    constraints: Sequence[PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint],
+    config: SimulationConfig,
+) -> Tuple[dict[str, SliceTuple3D | None], dict[str, Optional[str]]]:
     """
[…]
