+++ tests/fdtd/test_initialization.py
@@ -1,256 +1,519 @@
-from unittest.mock import MagicMock, Mock, patch
+from unittest.mock import Mock, patch
 
 import jax
 import jax.numpy as jnp
 import pytest
 
 from fdtdx.config import SimulationConfig
-from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
-
-# Import the functions to test
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer, ParameterContainer
 from fdtdx.fdtd.initialization import (
     _init_arrays,
     _init_params,
-    _resolve_object_constraints,
-    apply_params,
     place_objects,
 )
-from fdtdx.objects.device.parameters.transform import ParameterType
 from fdtdx.objects.object import (
     GridCoordinateConstraint,
-    PositionConstraint,
     SimulationObject,
 )
-from fdtdx.objects.static_material.static import UniformMaterialObject
+from fdtdx.objects.static_material.static import SimulationVolume, UniformMaterialObject
 
 
-class TestInitialization:
-    @pytest.fixture
-    def mock_config(self):
-        """Create a mock SimulationConfig for testing."""
-        config = Mock(spec=SimulationConfig)
-        config.dtype = jnp.float32
-        config.backend = "cpu"
-        config.resolution = 1.0
-        config.time_steps_total = 100
-        config.gradient_config = None
-        return config
+@pytest.fixture
+def mock_volume():
+    v = Mock(spec=SimulationVolume)
+    v.name = "volume"
+    v._size = [100, 100, 100]
+    v.partial_grid_shape = [100, 100, 100]
+    v.partial_real_shape = [None, None, None]
+    v.place_on_grid.return_value = v
+    return v
 
-    @pytest.fixture
-    def mock_volume(self):
-        """Create a mock volume object for testing."""
-        volume = Mock(spec=SimulationObject)
-        volume.name = "volume"
-        volume.grid_shape = (10, 10, 10)
-        volume.partial_grid_shape = (10, 10, 10)
-        volume.partial_real_shape = (10.0, 10.0, 10.0)
-        volume.placement_order = 0
-        return volume
 
-    @pytest.fixture
-    def mock_object(self):
-        """Create a mock simulation object for testing."""
-        obj = Mock(spec=SimulationObject)
-        obj.name = "test_object"
-        obj.grid_shape = (5, 5, 5)
-        obj.partial_grid_shape = (5, 5, 5)
-        obj.partial_real_shape = (5.0, 5.0, 5.0)
-        obj.placement_order = 1
-        return obj
+@pytest.fixture
+def mock_object():
+    o = Mock(spec=SimulationObject)
+    o.name = "obj1"
+    o._size = [10, 10, 10]
+    o.partial_grid_shape = [10, 10, 10]
+    o.partial_real_shape = [None, None, None]
+    o.place_on_grid.return_value = o
+    return o
 
-    @pytest.fixture
-    def mock_objects(self):
-        """Create a mock ObjectContainer for testing."""
-        objects = Mock(spec=ObjectContainer)
-        objects.devices = []
-        return objects
 
-    @pytest.fixture
-    def mock_uniform_material_object(self):
-        """Create a mock UniformMaterialObject for testing."""
-        material = Mock()
-        material.permittivity = 2.0
-        material.permeability = 1.0
-        material.electric_conductivity = 0.0
-        material.magnetic_conductivity = 0.0
+@pytest.fixture
+def mock_config():
+    c = Mock(spec=SimulationConfig)
+    c.resolution = 1.0
+    c.grid_shape = (100, 100, 100)
+    c.dtype = jnp.float32
+    c.backend = "cpu"
+    c.gradient_config = None
[…]
