### 2025-11-06 9bfcd53
docs: added contributing and faq to docs
