+++ docs/source/faq.rst
@@ -1,0 +1,14 @@
+FAQ
+===================
+
+I have a question not listed anywhere
+-------------------------------------
+If you find yourself pondering about a question which is not answered in the documentation, feel free to open a discussion `here <https://github.com/ymahlau/fdtdx/discussions>`_ . We are regularly checking the discussion section and will answer any questions.
+
+I suspect there is a bug in the code 
+-------------------------------------
+If you find a bug, or even just suspect that there is one feel free to open an issue on github `here <https://github.com/ymahlau/fdtdx/issues>`_ .
+
+There is a feature missing in FDTDX that I really need
+-------------------------------------------------------
+If you find yourself missing a feature in FDTDX for your specific application, feel free to open an issue `here <https://github.com/ymahlau/fdtdx/issues>`_ . While we cannot promise that we will implement all requested features, we will be open to any discussion. Additionally, we are happy to receive community contributions as well if you feel comfortable implementing the feature yourself. We will also help you with any questions regarding the implementation and internals of FDTDX that may come up during the impelementation process. Please take a look at the `contributing </contributing.html>`_ guidelines for more details.
