### 2025-11-08 fa2cf0a
Breaking up intersecting PMLs into their own objects, generalizing alpha/kappa/sigma for intersections between axes
