+++ src/fdtdx/objects/boundaries/initialization.py
@@ -1,6 +1,7 @@
 from typing import Literal, Union
 
 from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field
+from fdtdx.config import SimulationConfig
 from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
 from fdtdx.objects.boundaries.periodic import PeriodicBoundary
 from fdtdx.objects.boundaries.utils import axis_direction_from_kind
@@ -320,6 +321,7 @@ class BoundaryConfig(TreeClass):
 def boundary_objects_from_config(
     config: BoundaryConfig,
     volume: SimulationVolume,
+    sim_config: SimulationConfig,
 ) -> tuple[dict[str, Union[PerfectlyMatchedLayer, PeriodicBoundary]], list[PositionConstraint]]:
     """Creates boundary objects from a boundary configuration.
 
@@ -336,7 +338,10 @@ def boundary_objects_from_config(
             - dict mapping boundary names ('min_x', 'max_x', etc) to boundary objects
             - list of PositionConstraint objects for placing the boundaries
     """
-    boundaries, constraints = {}, []
+    boundaries: dict[str, Union[PerfectlyMatchedLayer, PeriodicBoundary]] = {}
+    constraints: list[PositionConstraint] = []
+
+    # Face settings
     thickness_dict = config.get_dict()
     type_dict = config.get_type_dict()
     kappa_start_dict = config.get_kappa_dict("kappa_start")
@@ -344,45 +349,178 @@ def boundary_objects_from_config(
     alpha_start_dict = config.get_alpha_dict("alpha_start")
     alpha_end_dict = config.get_alpha_dict("alpha_end")
 
-    for kind, thickness in thickness_dict.items():
-        axis, direction = axis_direction_from_kind(kind)
-        boundary_type = type_dict[kind]
-        kappa_start, kappa_end = kappa_start_dict[kind], kappa_end_dict[kind]
-        alpha_start, alpha_end = alpha_start_dict[kind], alpha_end_dict[kind]
+    # Helper: per-axis index and face keys
+    axes = ("x", "y", "z")
+    face_min_keys = ("min_x", "min_y", "min_z")
+    face_max_keys = ("max_x", "max_y", "max_z")
 
-        grid_shape_list: list[int | None] = [None, None, None]
-        grid_shape_list[axis] = thickness if boundary_type == "pml" else 1
-        grid_shape: PartialGridShape3D = tuple(grid_shape_list)  # type: ignore
+    # Periodic faces are still added as-is (they don't overlap).
+    for key in (*face_min_keys, *face_max_keys):
+        if type_dict[key] == "periodic":
+            ax, direction = axis_direction_from_kind(key)
+            direction_int = -1 if direction == "-" else 1
+            grid_shape_list: list[int | None] = [None, None, None]
+            grid_shape_list[ax] = 1  # zero-thickness placeholder for periodic
+            grid_shape: PartialGridShape3D = tuple(grid_shape_list)  # type: ignore
+            other_axes = [0, 1, 2]
+            del other_axes[ax]
 
-        other_axes = [0, 1, 2]
-        del other_axes[axis]
-
-        if boundary_type == "pml":
-            cur_boundary = PerfectlyMatchedLayer(
-                axis=axis,
-                partial_grid_shape=grid_shape,
-                kappa_start=kappa_start,
-                kappa_end=kappa_end,
-                alpha_start=alpha_start,
-                alpha_end=alpha_end,
-                direction=direction,
-            )
-        else:  # periodic
-            cur_boundary = PeriodicBoundary(
-                axis=axis,
+            pb = PeriodicBoundary(
+                axis=ax,
                 partial_grid_shape=grid_shape,
[…]
