+++ src/fdtdx/objects/boundaries/perfectly_matched_layer.py
@@ -1,3 +1,5 @@
+from typing import Tuple
+
 import jax
 import jax.numpy as jnp
 from typing_extensions import override
@@ -69,6 +71,9 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
     axis orientation and both positive/negative directions.
     """
 
+    #: Direction of parameters alpha, kappa and sigma.
+    direction_params: Tuple[int, int, int] = frozen_field()
+
     #: Initial loss parameter for complex frequency shifting. Defaults to 1e-8.
     alpha_start: float = frozen_field(default=1.0e-8)
 
@@ -92,9 +97,20 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
         Returns:
             str: Description like "min_x" or "max_z" indicating position
         """
-        axis_str = "x" if self.axis == 0 else "y" if self.axis == 1 else "z"
-        direction_str = "min" if self.direction == "-" else "max"
-        return f"{direction_str}_{axis_str}"
+        labels = []
+        for i, s in enumerate(self.direction_params):
+            if s == 0:
+                continue
+            axis_str = "x" if i == 0 else "y" if i == 1 else "z"
+            direction_str = "min" if s < 0 else "max"
+            labels.append(f"{direction_str}_{axis_str}")
+        if not labels:
+            axis_str = "x" if self.axis == 0 else "y" if self.axis == 1 else "z"
+            direction_str = "min" if self.direction == "-" else "max"
+            return f"{direction_str}_{axis_str}"
+        if len(labels) == 1:
+            return labels[0]
+        return "pml_" + "_".join(labels)
 
     @property
     @override
@@ -108,30 +124,82 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
 
     def _get_dtype_update_coefficients(self):
         dtype = self._config.dtype
-        sigma_E, sigma_H = standard_sigma_from_direction_axis(
-            thickness=self.thickness,
-            direction=self.direction,
-            axis=self.axis,
-            dtype=dtype,
-        )
 
-        kappa = kappa_from_direction_axis(
-            kappa_start=self.kappa_start,
-            kappa_end=self.kappa_end,
-            thickness=self.thickness,
-            direction=self.direction,
-            axis=self.axis,
-            dtype=dtype,
-        )
+        active_axes = [i for i, v in enumerate(self.direction_params) if v != 0]
+        # One axis (original behavior; non-overlapping PMLs)
+        if len(active_axes) == 1:
+            sigma_E, sigma_H = standard_sigma_from_direction_axis(
+                thickness=self.grid_shape[active_axes[0]],
+                direction="+" if self.direction_params[active_axes[0]] == 1 else "-",
+                axis=active_axes[0],
+                dtype=dtype,
+            )
+            alpha = alpha_from_direction_axis(
+                alpha_start=self.alpha_start,
+                alpha_end=self.alpha_end,
+                thickness=self.grid_shape[active_axes[0]],
+                direction="+" if self.direction_params[active_axes[0]] == 1 else "-",
+                axis=active_axes[0],
+                dtype=dtype,
+            )
+            kappa = kappa_from_direction_axis(
[…]
