### 2025-11-11 f7dc7cc
resolved some params and functions
