+++ tests/test_placement.py
@@ -124,12 +124,8 @@ def test_placement():
     # place the objects
     key = jax.random.PRNGKey(42)
     objects, arrays, params, config, _ = fdtdx.place_objects(
-        volume=volume,
+        objects=[volume, bound_cfg, d_energy, d_poynting, d_field, d_mode, d_phasor, materials, device, plane_source, gauss_source, mode_source, box, polygon, sphere, cylinder],
         config=config,
         constraints=constraints,
         key=key,
     )
-
-
-if __name__ == "__main__":
-    test_placement()
