+++ tests/fdtd/test_initialization.py
@@ -1,519 +1,146 @@
-from unittest.mock import Mock, patch
-
-import jax
-import jax.numpy as jnp
 import pytest
+from unittest.mock import MagicMock, patch
+import jax.numpy as jnp
 
-from fdtdx.config import SimulationConfig
-from fdtdx.fdtd.container import ArrayContainer, ObjectContainer, ParameterContainer
-from fdtdx.fdtd.initialization import (
-    _init_arrays,
-    _init_params,
-    place_objects,
-)
-from fdtdx.objects.object import (
-    GridCoordinateConstraint,
-    SimulationObject,
-)
-from fdtdx.objects.static_material.static import SimulationVolume, UniformMaterialObject
+import fdtdx.fdtd.initialization as initialization
 
 
+# ---------------------------------------------------------------------
+# Fixtures and Mocks
+# ---------------------------------------------------------------------
 @pytest.fixture
-def mock_volume():
-    v = Mock(spec=SimulationVolume)
-    v.name = "volume"
-    v._size = [100, 100, 100]
-    v.partial_grid_shape = [100, 100, 100]
-    v.partial_real_shape = [None, None, None]
-    v.place_on_grid.return_value = v
-    return v
+def mock_config():
+    cfg = MagicMock()
+    cfg.resolution = 1.0
+    cfg.dtype = "float32"
+    cfg.backend = "cpu"
+    cfg.gradient_config = None
+    cfg.time_steps_total = 100
+    cfg.grid_shape = (10, 10, 10)
+    return cfg
 
 
 @pytest.fixture
 def mock_object():
-    o = Mock(spec=SimulationObject)
-    o.name = "obj1"
-    o._size = [10, 10, 10]
-    o.partial_grid_shape = [10, 10, 10]
-    o.partial_real_shape = [None, None, None]
-    o.place_on_grid.return_value = o
-    return o
-
-
-@pytest.fixture
-def mock_config():
-    c = Mock(spec=SimulationConfig)
-    c.resolution = 1.0
-    c.grid_shape = (100, 100, 100)
-    c.dtype = jnp.float32
-    c.backend = "cpu"
-    c.gradient_config = None
-    return c
-
-
-@pytest.fixture
-def random_key():
-    return jax.random.PRNGKey(0)
-
-
-class MockArray:
-    """Mock array that supports .at[].set() pattern"""
-
-    def __init__(self):
-        self.at = self
-        self.set = Mock(return_value=self)
-        self.add = Mock(return_value=self)
-
-    def __getitem__(self, key):
-        return self
-
-    def __setitem__(self, key, value):
-        pass
-
-
-class MockArrayContainer:
-    """Mock ArrayContainer that supports .at[field].set(value) pattern"""
-
-    def __init__(self):
-        self.at = Mock(return_value=self)
-        self.set = Mock(return_value=self)
-        self.inv_permittivities = MockArray()
-        self.inv_permeabilities = MockArray()
-        self.E = Mock()
-        self.H = Mock()
-        self.electric_conductivity = None
-        self.magnetic_conductivity = None
-        self.detector_states = {}
-        self.boundary_states = {}
-        self.recording_state = None
-
-    def __getitem__(self, key):
-        return getattr(self, key)
-
-    def __setitem__(self, key, value):
-        setattr(self, key, value)
-
-
-class TestInitializationRefactor:
-    # Existing tests for place_objects (keep them as they are)...
-    def test_place_objects_success(self, mock_volume, mock_object, mock_config, random_key):
-        """place_objects should call underlying init methods and return valid structures."""
-        constraint = Mock(spec=GridCoordinateConstraint)
-
-        with (
[…]
