+++ src/fdtdx/fdtd/initialization.py
@@ -24,6 +24,7 @@ from fdtdx.objects.object import (
 from fdtdx.objects.static_material.static import SimulationVolume, StaticMultiMaterialObject, UniformMaterialObject
 from fdtdx.typing import SliceTuple3D
 
+DEFAULT_MAX_ITER = 1000
 
 def place_objects(
     objects: list[SimulationObject],
@@ -57,8 +58,16 @@ def place_objects(
     Raises:
         ValueError: If constraint resolution fails for one or more objects.
     """
+    # Sanity check: Ensure all objects have unique names
+    object_names = [obj.name for obj in objects]
+    duplicates = {name for name in object_names if object_names.count(name) > 1}
+    if duplicates:
+        raise Exception(
+            f"Duplicate object names detected: {', '.join(sorted(duplicates))}. "
+            "Each object must have a unique name before resolving constraints into grid slices."
+        )
     # Step 1: Resolve constraints into grid slices
-    resolved_slices, errors = _resolve_object_constraints(
+    resolved_slices, errors = resolve_object_constraints(
         objects=objects,
         constraints=constraints,
         config=config,
@@ -382,13 +391,13 @@ def _init_params(
     return params
 
 
-def _resolve_object_constraints(
+def resolve_object_constraints(
     objects: list[SimulationObject],
     constraints: Sequence[PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint],
     config: SimulationConfig,
 ) -> Tuple[dict[str, SliceTuple3D | None], dict[str, Optional[str]]]:
     """
-    Cleaned version of `_resolve_object_constraints`.
+    Cleaned version of `resolve_object_constraints`.
 
     Resolves positioning and sizing constraints between simulation objects.
     Returns dictionaries for resolved slices and error messages.
@@ -457,10 +466,13 @@ def _resolve_static_shapes(objects, shape_dict, config, resolution, errors):
 
 def _apply_constraints_iteratively(
     object_map: dict[str, SimulationObject],
-    constraints: Sequence[PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint],
+    constraints: Sequence[
+        PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint
+    ],
     shape_dict: dict["SimulationObject", list[int | None]],
     slice_dict: dict["SimulationObject", list[list[int | None]]],
     errors: dict[str, Optional[str]],
+    max_iter: int = DEFAULT_MAX_ITER,
 ) -> None:
     """
     Iteratively apply all constraints until shapes and positions converge.
@@ -471,11 +483,10 @@ def _apply_constraints_iteratively(
         shape_dict: Dict tracking (x, y, z) shape values in grid units.
         slice_dict: Dict tracking grid slice start/end positions for each axis.
         errors: Dict collecting user-friendly error messages.
+        max_iter: Maximum number of iterations before aborting (default: DEFAULT_MAX_ITER).
     """
-    MAX_ITER = 1000
-    stable_iterations = 0
 
-    for iteration in range(MAX_ITER):
+    for iteration in range(max_iter):
         changed = False
 
         for c in constraints:
@@ -488,20 +499,30 @@ def _apply_constraints_iteratively(
                 errors[obj_name] = f"Unknown object '{obj_name}' in constraint."
                 continue
 
+            handled = False
             try:
                 if isinstance(c, PositionConstraint):
                     changed |= _apply_position_constraint(c, obj, other, shape_dict, slice_dict)
+                    handled = True
 
                 elif isinstance(c, SizeConstraint):
                     changed |= _apply_size_constraint(c, obj, other, shape_dict)
[…]
