### 2025-11-12 fd77d2c
fixing some errors
