+++ src/fdtdx/fdtd/initialization.py
@@ -61,10 +61,16 @@ def place_objects(
     # Sanity check: Ensure all objects have unique names
     object_names = [obj.name for obj in objects]
     duplicates = {name for name in object_names if object_names.count(name) > 1}
+    invalid_objects = [obj for obj in objects if not isinstance(obj, SimulationObject)]
     if duplicates:
         raise Exception(
             f"Duplicate object names detected: {', '.join(sorted(duplicates))}. "
             "Each object must have a unique name before resolving constraints into grid slices."
+        )
+    if invalid_objects:
+        raise ValueError(
+            f"Invalid object types detected: {', '.join(sorted(invalid_objects))}. "
+            "All objects must be instances or subclasses of SimulationObject."
         )
     # Step 1: Resolve constraints into grid slices
     resolved_slices, errors = resolve_object_constraints(
@@ -392,38 +398,47 @@ def _init_params(
 
 
 def resolve_object_constraints(
-    objects: list[SimulationObject],
-    constraints: Sequence[PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint],
-    config: SimulationConfig,
-) -> Tuple[dict[str, SliceTuple3D | None], dict[str, Optional[str]]]:
-    """
-    Cleaned version of `resolve_object_constraints`.
+        objects: list[SimulationObject],
+        constraints: Sequence[
+            PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint
+            ],
+        config: SimulationConfig,
+) -> tuple[dict, dict]:
+    """Resolve object constraints into grid slices and shapes."""
 
-    Resolves positioning and sizing constraints between simulation objects.
-    Returns dictionaries for resolved slices and error messages.
+    # Convert objects list to object_map dictionary
+    object_map = {}
+    for obj in objects:
+        object_map[obj.name] = obj
+        print(f"Added to object_map: {obj.name} -> {type(obj).__name__}")  # Debug
 
-    Args:
-        objects: List of simulation objects to resolve.
-        constraints: List of constraints defining object relations.
-        config: Simulation configuration object.
+    # Initialize shape_dict and slice_dict with object references as keys
+    shape_dict = {}
+    slice_dict = {}
+    for obj in objects:
+        shape_dict[obj] = [None, None, None]
+        slice_dict[obj] = [[None, None], [None, None], [None, None]]
 
-    Returns:
-        resolved: dict mapping object names to SliceTuple3D or None
-        errors: dict mapping object names to error messages (None if OK)
-    """
-    resolution = config.resolution
+    errors = {obj.name: None for obj in objects}
 
-    # Initialize structures
-    errors: dict[str, Optional[str]] = {obj.name: None for obj in objects}
+    # Debug: print constraints before applying
+    print(f"Number of constraints: {len(constraints)}")
+    for i, c in enumerate(constraints):
+        obj_field = getattr(c, "object", None)
+        other_field = getattr(c, "other_object", None)
+        print(f"Constraint {i}: object={type(obj_field)}, other_object={type(other_field)}")
 
-    all_objects = _collect_objects_from_constraints(objects, constraints)
-    shape_dict, slice_dict = _initialize_shape_dicts(all_objects)
+    # Apply constraints iteratively
+    _apply_constraints_iteratively(object_map, constraints, shape_dict, slice_dict, errors)
 
-    _resolve_static_shapes(all_objects, shape_dict, config, resolution, errors)
-    _apply_constraints_iteratively(all_objects, constraints, shape_dict, slice_dict, errors)
+    # Convert shape_dict and slice_dict from object references to object names
+    resolved_slices = {}
[…]
