### 2025-11-13 485125f
fix sigma_E and sigma_H scaling in update.py by factor of eta0
