### 2025-11-14 9be9039
Implement first three improvements requested in issue #174
