+++ src/fdtdx/units/unitful.py
@@ -1,6 +1,8 @@
 # ruff: noqa: F811
 
 from dataclasses import dataclass
+
+# from numbers import Number
 from typing import Any, Self
 
 import jax
@@ -550,7 +552,7 @@ def get_static_operand(
     return x_arr
 
 
-def can_perform_scatic_ops(x: StaticArrayLike | None):
+def can_perform_static_ops(x: StaticArrayLike | None):
     if x is None:
         return False
     if isinstance(x, NonPhysicalArrayLike):
@@ -642,6 +644,7 @@ def multiply(x: ArrayLike, y: Unitful) -> Unitful:
         unit_x = Unitful(val=x, unit=Unit(scale=0, dim={}))
         x = unit_x.val
         scale_offset = unit_x.unit.scale
+
     new_val = y.val * x
     new_scale = y.unit.scale + scale_offset
     new_static_arr = None
@@ -667,7 +670,27 @@ def multiply(x: int, y: int) -> int:
 
 
 @overload
+def multiply(x: int, y: float) -> float:
+    return x * y
+
+
+@overload
+def multiply(x: float, y: int) -> float:
+    return x * y
+
+
+@overload
 def multiply(x: float, y: float) -> float:
+    return x * y
+
+
+@overload
+def multiply(x: complex, y: float) -> complex:
+    return x * y
+
+
+@overload
+def multiply(x: float, y: complex) -> complex:
     return x * y
 
 
@@ -849,9 +872,17 @@ def add(x, y):  # type: ignore
 ## Matrix Multiplication ###################################
 @overload
 def matmul(x: Unitful, y: Unitful, **kwargs) -> Unitful:
-    # TODO: numpy values should stay numpy if possible
     x_align, y_align = align_scales(x, y)
-    new_val = jnp._orig_matmul(x_align.val, y_align.val)  # type: ignore
+
+    invalid_types = (bool, int, float, complex, np.bool_, np.number)
+    if isinstance(x.val, invalid_types) or isinstance(y.val, invalid_types):
+        raise TypeError(f"matmul received invalid types: {type(x).__name__}, {type(y).__name__}")
+    # handling of numpy arrays
+    elif isinstance(x.val, np.ndarray) and isinstance(y.val, np.ndarray):
+        new_val = np.matmul(x_align.val, y_align.val, **kwargs)
+    else:
+        new_val = jnp._orig_matmul(x_align.val, y_align.val, **kwargs)  # type: ignore
+
     unit_dict = dim_after_multiplication(x.unit.dim, y.unit.dim)
     new_scale = 2 * x_align.unit.scale
     # if static arrays exist, perform subtract with static arrs
@@ -859,13 +890,28 @@ def matmul(x: Unitful, y: Unitful, **kwargs) -> Unitful:
     if is_traced(new_val):
         x_arr = get_static_operand(x_align)
         y_arr = get_static_operand(y_align)
-        if can_perform_scatic_ops(x_arr) and can_perform_scatic_ops(y_arr):
-            new_static_arr = x_arr @ y_arr  # type: ignore
+        if can_perform_static_ops(x_arr) and can_perform_static_ops(y_arr):
+            new_static_arr = np.matmul(x_arr, y_arr, **kwargs)  # type: ignore
     return Unitful(val=new_val, unit=Unit(scale=new_scale, dim=unit_dict), static_arr=new_static_arr)
 
 
 @overload
 def matmul(x: jax.Array, y: jax.Array, **kwargs) -> jax.Array:
+    return jnp._orig_matmul(x, y, **kwargs)  # type: ignore
+
+
+@overload
+def matmul(x: np.ndarray, y: np.ndarray, **kwargs) -> np.ndarray:
[…]
