+++ src/fdtdx/objects/boundaries/initialization.py
@@ -90,6 +90,24 @@ class BoundaryConfig(TreeClass):
     #: Final kappa value at max z boundary. Default 1.5.
     kappa_end_maxz: float = frozen_field(default=1.5)
 
+    #: Polynomial order for kappa grading at min x boundary. Default 1.0.
+    kappa_order_minx: float = frozen_field(default=1.0)
+
+    #: Polynomial order for kappa grading at max x boundary. Default 1.0.
+    kappa_order_maxx: float = frozen_field(default=1.0)
+
+    #: Polynomial order for kappa grading at min y boundary. Default 1.0.
+    kappa_order_miny: float = frozen_field(default=1.0)
+
+    #: Polynomial order for kappa grading at max y boundary. Default 1.0.
+    kappa_order_maxy: float = frozen_field(default=1.0)
+
+    #: Polynomial order for kappa grading at min z boundary. Default 1.0.
+    kappa_order_minz: float = frozen_field(default=1.0)
+
+    #: Polynomial order for kappa grading at max z boundary. Default 1.0.
+    kappa_order_maxz: float = frozen_field(default=1.0)
+
     #: Initial alpha value at min x boundary. Default 1e-8.
     alpha_start_minx: float = frozen_field(default=1e-8)
 
@@ -125,6 +143,78 @@ class BoundaryConfig(TreeClass):
 
     #: Final alpha value at max z boundary. Default 1e-8.
     alpha_end_maxz: float = frozen_field(default=1e-8)
+
+    #: Polynomial order for alpha grading at min x boundary. Default 1.0.
+    alpha_order_minx: float = frozen_field(default=1.0)
+
+    #: Polynomial order for alpha grading at max x boundary. Default 1.0.
+    alpha_order_maxx: float = frozen_field(default=1.0)
+
+    #: Polynomial order for alpha grading at min y boundary. Default 1.0.
+    alpha_order_miny: float = frozen_field(default=1.0)
+
+    #: Polynomial order for alpha grading at max y boundary. Default 1.0.
+    alpha_order_maxy: float = frozen_field(default=1.0)
+
+    #: Polynomial order for alpha grading at min z boundary. Default 1.0.
+    alpha_order_minz: float = frozen_field(default=1.0)
+
+    #: Polynomial order for alpha grading at max z boundary. Default 1.0.
+    alpha_order_maxz: float = frozen_field(default=1.0)
+
+    #: Initial sigma value at min x boundary. Default 0.0.
+    sigma_start_minx: float = frozen_field(default=0.0)
+
+    #: Final sigma value at min x boundary. Default 1.0.
+    sigma_end_minx: float = frozen_field(default=1.0)
+
+    #: Initial sigma value at max x boundary. Default 0.0.
+    sigma_start_maxx: float = frozen_field(default=0.0)
+
+    #: Final sigma value at max x boundary. Default 1.0.
+    sigma_end_maxx: float = frozen_field(default=1.0)
+
+    #: Initial sigma value at min y boundary. Default 0.0.
+    sigma_start_miny: float = frozen_field(default=0.0)
+
+    #: Final sigma value at min y boundary. Default 1.0.
+    sigma_end_miny: float = frozen_field(default=1.0)
+
+    #: Initial sigma value at max y boundary. Default 0.0.
+    sigma_start_maxy: float = frozen_field(default=0.0)
+
+    #: Final sigma value at max y boundary. Default 1.0.
+    sigma_end_maxy: float = frozen_field(default=1.0)
+
+    #: Initial sigma value at min z boundary. Default 0.0.
+    sigma_start_minz: float = frozen_field(default=0.0)
+
+    #: Final sigma value at min z boundary. Default 1.0.
[…]
