### 2025-11-14 5718eb6
refactored object placement
