### 2025-11-16 ddf3dac
Improved example scripts
