### 2025-11-17 #8912 @andrewgiuliani
Hi @fruzsinaagocs, I've completed almost all of my checklist, but leaving one item unchecked because it appears one of the examples doesn't run on my machine.  The software installs fine and the second example works fine.

I don't want to hold up this review, because I am happy with this submission otherwise and the authors have quickly addressed all of my other concerns.  I will try to troubleshoot with the authors a bit longer, but I do not think that this should stop this work's acceptance in JOSS.

Thanks!
