### 2025-11-18 #8912 @ymahlau
Hi @victorapm, @andrewgiuliani and @fruzsinaagocs, first of all many thanks for the review and issues raised, I believe this improved the software greatly.

Regarding the remaining issue, I was able to reproduce the setup of @andrewgiuliani and it seems that the gradient computation through the simulation is just very slow on CPU. The reason for that is that the JAX framework is optimized for GPU/TPU execution and sometimes lacks efficiency on CPU. The script does seem to work on the setup of @andrewgiuliani, he just needs to run it a few minutes longer to see the first results.

If there are any other concerns I am happy to address them as well! Many thanks to all of you!
