### 2025-11-18 22c0e48
added new test cases for fdtd initialization
