+++ tests/fdtd/test_initialization.py
@@ -1,129 +1,1266 @@
+from unittest.mock import Mock, patch
+
+import jax
+import jax.numpy as jnp
 import pytest
 
-import fdtdx.fdtd.initialization as init
 from fdtdx.config import SimulationConfig
-from fdtdx.materials import Material
-from fdtdx.objects.static_material.static import SimulationVolume, UniformMaterialObject
+from fdtdx.fdtd.container import ObjectContainer
+from fdtdx.fdtd.initialization import (
+    _apply_constraints_iteratively,
+    _apply_grid_coordinate_constraint,
+    _apply_position_constraint,
+    _apply_real_coordinate_constraint,
+    _apply_size_constraint,
+    _apply_size_extension_constraint,
+    _check_objects_names_from_constraints,
+    _extend_to_inf_if_possible,
+    _handle_unresolved_objects,
+    _init_arrays,
+    _resolve_static_shapes,
+    _resolve_volume_name,
+    _update_grid_shapes_from_slices,
+    _update_grid_slices_from_shapes,
+    place_objects,
+    resolve_object_constraints,
+)
+from fdtdx.objects.object import (
+    GridCoordinateConstraint,
+    PositionConstraint,
+    RealCoordinateConstraint,
+    SimulationObject,
+    SizeConstraint,
+    SizeExtensionConstraint,
+)
+from fdtdx.objects.static_material.static import (
+    SimulationVolume,
+    StaticMultiMaterialObject,
+    UniformMaterialObject,
+)
 
 
-# ---------------------------------------------------------------------
-# Fixtures and Mocks
-# ---------------------------------------------------------------------
-@pytest.fixture
-def mock_config():
-    cfg = SimulationConfig(
-        time=100e-12,
-        resolution=100e-9,
+# Test for duplicate object names
+def test_resolve_object_constraints_duplicate_names():
+    """Test that duplicate object names raise an exception."""
+    config = Mock(spec=SimulationConfig)
+    config.resolution = 1.0
+
+    obj1 = Mock(spec=SimulationVolume)
+    obj1.name = "volume"
+    obj2 = Mock(spec=UniformMaterialObject)
+    obj2.name = "volume"  # Duplicate name
+
+    objects = [obj1, obj2]
+    constraints = []
+
+    with pytest.raises(Exception, match="Duplicate object names detected"):
+        resolve_object_constraints(objects, constraints, config)
+
+
+# Test for invalid object types
+def test_resolve_object_constraints_invalid_objects():
+    """Test that invalid object types raise an exception."""
+    config = Mock(spec=SimulationConfig)
+    config.resolution = 1.0
+
+    obj1 = Mock(spec=SimulationVolume)
+    obj1.name = "volume"
+
+    # Create an object without proper SimulationObject type
+    invalid_obj = Mock()
+    invalid_obj.name = "invalid"
+
+    objects = [obj1, invalid_obj]
+    constraints = []
+
+    with pytest.raises((ValueError, TypeError)):
+        resolve_object_constraints(objects, constraints, config)
+
+
+# Test for unknown object name in constraints
+def test_check_objects_names_unknown_object():
+    """Test that unknown object names in constraints raise an exception."""
+    constraint = Mock(spec=PositionConstraint)
+    constraint.object = "unknown_object"
+    constraint.other_object = None
+
+    constraints = [constraint]
+    object_names = ["volume", "obj1"]
+
+    with pytest.raises(ValueError, match="Unknown object name in constraint"):
+        _check_objects_names_from_constraints(constraints, object_names)
+
+
+# Test for unknown object name in other_object field
+def test_check_objects_names_unknown_other_object():
[…]
