### 2025-11-18 f13ce8c
delete redundant periodic field updates, refactor update.py
