+++ src/fdtdx/fdtd/update.py
@@ -1,9 +1,7 @@
 import jax
-import jax.numpy as jnp
 
 from fdtdx.config import SimulationConfig
-from fdtdx.constants import c as c0
-from fdtdx.constants import eps0, eta0
+from fdtdx.constants import eta0
 from fdtdx.core.physics.curl import curl_E, curl_H, interpolate_fields
 from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 from fdtdx.fdtd.misc import add_boundary_interfaces, collect_boundary_interfaces
@@ -52,119 +50,20 @@ def update_E(
         ArrayContainer: Updated ArrayContainer with new E field values
     """
 
-    # Handle periodic boundaries by copying values from opposite sides
-    if simulate_boundaries:
-        E = arrays.E
-        for boundary in objects.boundary_objects:
-            if isinstance(boundary, PeriodicBoundary):
-                E = E.at[..., boundary.boundary_slice].set(E[..., boundary.opposite_slice])
-        arrays = arrays.at["E"].set(E)
-
     inv_eps = arrays.inv_permittivities
     c = config.courant_number
     sigma_E = arrays.electric_conductivity
-
-    # Get periodic axes for curl operation
-    periodic_axes = get_periodic_axes(objects)
-    H_pad = arrays.H
-    for i, periodic in enumerate(periodic_axes):
-        pad_mode = "wrap" if periodic else "constant"
-        # Create padding tuple for current axis
-        if i == 0:
-            pad_width = ((0, 0), (1, 1), (0, 0), (0, 0))
-        elif i == 1:
-            pad_width = ((0, 0), (0, 0), (1, 1), (0, 0))
-        else:  # i == 2
-            pad_width = ((0, 0), (0, 0), (0, 0), (1, 1))
-        H_pad = jnp.pad(H_pad, pad_width, mode=pad_mode)
-
-    DyHz = (H_pad[2] - jnp.roll(H_pad[2], 1, axis=1))[1:-1, 1:-1, 1:-1]
-    DzHy = (H_pad[1] - jnp.roll(H_pad[1], 1, axis=2))[1:-1, 1:-1, 1:-1]
-    DzHx = (H_pad[0] - jnp.roll(H_pad[0], 1, axis=2))[1:-1, 1:-1, 1:-1]
-    DxHz = (H_pad[2] - jnp.roll(H_pad[2], 1, axis=0))[1:-1, 1:-1, 1:-1]
-    DxHy = (H_pad[1] - jnp.roll(H_pad[1], 1, axis=0))[1:-1, 1:-1, 1:-1]
-    DyHx = (H_pad[0] - jnp.roll(H_pad[0], 1, axis=1))[1:-1, 1:-1, 1:-1]
-
-    # Auxiliary fields
-    psi_Exy = arrays.psi_E[0, :, :, :]
-    psi_Exz = arrays.psi_E[1, :, :, :]
-    psi_Eyz = arrays.psi_E[2, :, :, :]
-    psi_Eyx = arrays.psi_E[3, :, :, :]
-    psi_Ezx = arrays.psi_E[4, :, :, :]
-    psi_Ezy = arrays.psi_E[5, :, :, :]
-
-    if simulate_boundaries:
-        # Get E-field PML coefficients
-        b_x = (
-            jnp.expm1(
-                -c
-                * config.resolution
-                / c0
-                / eps0
-                * (arrays.sigma[0, :, :, :] / arrays.kappa[0, :, :, :] + arrays.alpha[0, :, :, :])
-            )
-            + 1
-        )
-        b_y = (
-            jnp.expm1(
[…]
