### 2025-11-20 a199648
removed unnecessary mocks and kept real objects for tests
