### 2025-11-20 7adb85e
added new test cases for fdtd initialization (#192)
* added new test cases for fdtd initialization

* removed unnecessary mocks and kept real objects for tests
