+++ tests/fdtd/test_initialization.py
@@ -1,129 +1,665 @@
+"""
+Additional test cases for initialization.py to improve code coverage.
+These tests should be ADDED to the existing test_initialization.py file.
+DO NOT replace existing tests - these are supplementary tests.
+"""
+
+from unittest.mock import MagicMock, Mock, patch
+
+import jax
+import jax.numpy as jnp
 import pytest
 
-import fdtdx.fdtd.initialization as init
 from fdtdx.config import SimulationConfig
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
+from fdtdx.fdtd.initialization import (
+    apply_params,
+    resolve_object_constraints,
+)
 from fdtdx.materials import Material
-from fdtdx.objects.static_material.static import SimulationVolume, UniformMaterialObject
+from fdtdx.objects.device.parameters.transform import ParameterType
+from fdtdx.objects.object import (
+    GridCoordinateConstraint,
+    PositionConstraint,
+    RealCoordinateConstraint,
+    SizeConstraint,
+    SizeExtensionConstraint,
+)
+from fdtdx.objects.static_material.static import (
+    SimulationVolume,
+    UniformMaterialObject,
+)
 
 
-# ---------------------------------------------------------------------
-# Fixtures and Mocks
-# ---------------------------------------------------------------------
+# Fixtures for reusable test objects
 @pytest.fixture
-def mock_config():
-    cfg = SimulationConfig(
-        time=100e-12,
-        resolution=100e-9,
+def simple_config():
+    """Create a simple simulation configuration."""
+    return SimulationConfig(
+        resolution=1.0,
+        time=100e-15,  # 100 femtoseconds
     )
-    return cfg
-
-
-@pytest.fixture
-def mock_volume():
-    return SimulationVolume(partial_real_shape=(5e-6, 5e-6, 5e-6))
-
-
-@pytest.fixture
-def mock_object():
-    obj = UniformMaterialObject(
-        partial_real_shape=(1e-6, 1e-6, 1e-6),
-        material=Material(),
-    )
-    return obj
 
 
 @pytest.fixture
-def mock_object2():
+def simple_volume():
+    """Create a simple simulation volume."""
+    return SimulationVolume(
+        name="volume",
+        partial_grid_shape=(100, 100, 100),
+    )
+
+
+@pytest.fixture
+def simple_material():
+    """Create a simple material."""
+    return Material(
+        permittivity=2.0,
+        permeability=1.0,
+        electric_conductivity=0.0,
+        magnetic_conductivity=0.0,
+    )
+
+
+# Test cases for constraint resolution with real objects
+
+
+def test_resolve_constraints_with_duplicate_object_names(simple_config, simple_volume):
+    """Test that duplicate object names raise an exception."""
+    volume2 = SimulationVolume(
+        name="volume",  # Duplicate name
+        partial_grid_shape=(50, 50, 50),
+    )
+
+    objects = [simple_volume, volume2]
+    constraints = []
+
+    with pytest.raises(Exception, match="Duplicate object names"):
+        resolve_object_constraints(objects, constraints, simple_config)
+
+
+def test_resolve_constraints_unknown_object_in_constraint(simple_config, simple_volume):
+    """Test that unknown object names in constraints raise an exception."""
+    objects = [simple_volume]
+
+    # Reference an object that doesn't exist
+    constraint = GridCoordinateConstraint(object="nonexistent_object", axes=[0], sides=["-"], coordinates=[10])
+
+    constraints = [constraint]
+
+    with pytest.raises(ValueError, match="Unknown object name"):
+        resolve_object_constraints(objects, constraints, simple_config)
+
+
+def test_resolve_constraints_no_simulation_volume(simple_config, simple_material):
[…]
