### 2025-11-21 22a2fa7
Implement improvements from issue #174, added test code
