+++ src/fdtdx/units/unitful.py
@@ -685,12 +685,12 @@ def multiply(x: float, y: float) -> float:
 
 
 @overload
-def multiply(x: complex, y: float) -> complex:
+def multiply(x: complex, y: int | float) -> complex:
     return x * y
 
 
 @overload
-def multiply(x: float, y: complex) -> complex:
+def multiply(x: int | float, y: complex) -> complex:
     return x * y
 
 
@@ -700,12 +700,43 @@ def multiply(x: complex, y: complex) -> complex:
 
 
 @overload
+def multiply(x: int | float | complex | np.number, y: np.ndarray) -> np.ndarray:
+    return x * y
+
+
+@overload
+def multiply(x: np.ndarray, y: int | float | complex | np.number) -> np.ndarray:
+    return x * y
+
+
+@overload
+def multiply(x: int | float | complex | np.number, y: jax.Array) -> jax.Array:
+    return jnp.asarray(x) * y
+
+
+@overload
+def multiply(x: jax.Array, y: int | float | complex | np.number) -> jax.Array:
+    return x * jnp.asarray(y)
+
+
+@overload
 def multiply(x: jax.Array, y: jax.Array) -> jax.Array:
     return x * y
 
 
 @overload
-def multiply(x: ArrayLike, y: ArrayLike) -> ArrayLike:
+def multiply(x: jax.Array, y: np.ndarray) -> jax.Array:
+    return x * y
+
+
+@overload
+def multiply(x: np.ndarray, y: jax.Array) -> jax.Array:
+    x_jax = jnp.asarray(x)
+    return x_jax * y
+
+
+@overload
+def multiply(x: np.ndarray, y: np.ndarray) -> np.ndarray:
     return x * y
 
 
@@ -783,7 +814,27 @@ def divide(x1: int, x2: int) -> float:
 
 
 @overload
+def divide(x1: int, x2: float) -> float:
+    return x1 / x2
+
+
+@overload
+def divide(x1: float, x2: int) -> float:
+    return x1 / x2
+
+
+@overload
 def divide(x1: float, x2: float) -> float:
+    return x1 / x2
+
+
+@overload
+def divide(x1: int | float, x2: complex) -> complex:
+    return x1 / x2
+
+
+@overload
+def divide(x1: complex, x2: int | float) -> complex:
     return x1 / x2
 
 
@@ -793,12 +844,43 @@ def divide(x1: complex, x2: complex) -> complex:
 
 
 @overload
+def divide(x: int | float | complex | np.number, y: np.ndarray) -> np.ndarray:
+    return x / y
+
+
+@overload
+def divide(x: np.ndarray, y: int | float | complex | np.number) -> np.ndarray:
+    return x / y
+
+
+@overload
+def divide(x: int | float | complex | np.number, y: jax.Array) -> jax.Array:
+    return jnp.asarray(x) / y
+
+
+@overload
+def divide(x: jax.Array, y: int | float | complex | np.number) -> jax.Array:
+    return x / jnp.asarray(y)
+
+
+@overload
 def divide(x1: jax.Array, x2: jax.Array) -> jax.Array:
     return x1 / x2
 
 
 @overload
-def divide(x1: ArrayLike, x2: ArrayLike) -> ArrayLike:
+def divide(x1: jax.Array, x2: np.ndarray) -> jax.Array:
+    return x1 / x2
+
+
+@overload
+def divide(x1: np.ndarray, x2: jax.Array) -> jax.Array:
+    x1_jax = jnp.asarray(x1)
+    return x1_jax / x2
+
+
+@overload
+def divide(x1: np.ndarray, x2: np.ndarray) -> np.ndarray:
     return x1 / x2
 
 
@@ -844,7 +926,27 @@ def add(x: int, y: int) -> int:
 
 
 @overload
[…]
