+++ tests/units/test_unitful.py
@@ -26,6 +26,8 @@ from fdtdx.units.unitful import (
     multiply,
     ne,
     pow,
+    reshape,
+    squeeze,
     subtract,
 )
 
@@ -48,7 +50,7 @@ def test_multiply_unitful_unitful_different_dimensions():
 
     result = multiply(time_val, freq_val)
 
-    assert jnp.allclose(result.value(), 50.0)
+    assert math.isclose(result.value(), 50.0)  # type: ignore
     assert result.unit.dim == {}  # s^1 * s^-1 = dimensionless
 
 
@@ -59,34 +61,32 @@ def test_multiply_unitful_unitful_different_scales():
 
     result = multiply(time1, time2)
 
-    assert jnp.allclose(result.value(), 6.0e-3)
+    assert math.isclose(result.value(), 6.0e-3)  # type: ignore
     assert result.unit.dim == {SI.s: 2}
 
 
-# unitful-unitful with val python-python
 def test_multiply_unitful_unitful_python_python_type_and_dim():
     """Test multiplication of Unitful objects with value Python scalar and Python scalar"""
     a = 2 * s
     b = 3.5 * s
     res = multiply(a, b)
 
-    assert res.value() == 7.0
+    assert math.isclose(res.value(), 7.0)  # type: ignore
     assert res.unit.dim == {SI.s: 2}
     assert isinstance(res.val, float)
 
 
-# unitful-unitful with val numpy-numpy
 def test_multiply_unitful_unitful_numpy_numpy_type_and_dim():
     """Test multiplication of Unitful objects with value numpy Array and numpy Array"""
     a = s * np.array([1.0, 2.0, 3.0])
     b = s * np.array([4.0, 5.0, 6.0])
     res = multiply(a, b)
+
     assert isinstance(res.val, np.ndarray)
     assert jnp.allclose(res.value(), jnp.array([4.0, 10.0, 18.0]))
     assert res.unit.dim == {SI.s: 2}
 
 
-# unitful-unitful with val jax-jax
 def test_multiply_unitful_unitful_jax_jax_type_and_dim():
     """Test multiplication of Unitful objects with value Jax Array and Jax Array"""
     a = s * jnp.array([1.0, 2.0, 3.0])
@@ -111,8 +111,8 @@ def test_multiply_with_arrays():
     assert result.unit.dim == {}  # dimensionless
 
 
-def test_multiply_arraylike_arraylike():
-    """Test multiplication of two ArrayLike objects"""
+def test_multiply_overload_jax_jax():
+    """Test multiplication of two jax Arrays"""
     arr1 = jnp.array([1.0, 2.0, 3.0])
     arr2 = jnp.array([4.0, 5.0, 6.0])
 
@@ -120,6 +120,29 @@ def test_multiply_arraylike_arraylike():
 
     expected = jnp.array([4.0, 10.0, 18.0])
     assert jnp.allclose(result, expected)
+    assert isinstance(result, jax.Array)
+
+
+def test_multiply_overload_numpy_numpy():
+    """Test multiplication of two numpy arrays"""
+    arr1 = np.array([1.0, 2.0, 3.0])
+    arr2 = np.array([4.0, 5.0, 6.0])
+
+    result = multiply(arr1, arr2)
+
+    expected = np.array([4.0, 10.0, 18.0])
+    assert np.allclose(result, expected)
+    assert isinstance(result, np.ndarray)
+
+
+def test_multiply_python_scalars_promotion():
+    """Test multiplication with Python Scalar int, float and complex"""
[…]
