### 2025-11-21 e1e1068
adding tests for place_objects
