### 2025-11-22 5aec5e9
fix bug in unary_fn
