### 2025-11-23 35fe3cb
adding tests for place_objects (#195)
* adding tests for place_objects

* minor correction
