### 2025-11-24 5ef7716
refactored plot_setup by adding a helper function plot_setup_from_side
