+++ src/fdtdx/utils/plot_setup.py
@@ -1,179 +1,128 @@
-from pathlib import Path
-from typing import Any, Sequence
+from typing import Optional
 
 import matplotlib.pyplot as plt
-from matplotlib.figure import Figure
-from matplotlib.patches import Patch, Rectangle
+from matplotlib.patches import Rectangle
 
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.container import ObjectContainer
-from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
-from fdtdx.objects.boundaries.periodic import PeriodicBoundary
-from fdtdx.objects.object import SimulationObject
+
+
+def plot_setup_from_side(
+    config: SimulationConfig,
+    objects: ObjectContainer,
+    resolution: float,
+    ax,
+    viewing_side: str,
+    exclude_large_object_ratio: Optional[float] = None,
+):
+    """
+    Plot a 2D projection of the simulation setup from a chosen axis.
+
+    viewing_side: one of 'x', 'y', 'z' indicating which projection to plot.
+    exclude_large_object_ratio: if given, objects that cover more than this
+        fraction of the image area will be excluded.
+    """
+
+    volume = objects.volume
+
+    # Map viewing_side to coordinate pairs
+    if viewing_side == "x":
+        plane = (1, 2)
+        xlabel, ylabel = "y (µm)", "z (µm)"
+        title = "YZ plane"
+    elif viewing_side == "y":
+        plane = (0, 2)
+        xlabel, ylabel = "x (µm)", "z (µm)"
+        title = "XZ plane"
+    elif viewing_side == "z":
+        plane = (0, 1)
+        xlabel, ylabel = "x (µm)", "y (µm)"
+        title = "XY plane"
+    else:
+        raise ValueError("viewing_side must be 'x', 'y', or 'z'")
+
+    # Total area for filtering
+    total_w = volume.grid_shape[plane[0]] * resolution
+    total_h = volume.grid_shape[plane[1]] * resolution
+    total_area = total_w * total_h
+
+    for obj in objects:
+        # Determine bounds in this projection
+        x0 = obj.grid_bounds[plane[0]][0] * resolution
+        x1 = obj.grid_bounds[plane[0]][1] * resolution
+        y0 = obj.grid_bounds[plane[1]][0] * resolution
+        y1 = obj.grid_bounds[plane[1]][1] * resolution
+
+        w = x1 - x0
+        h = y1 - y0
+        area = w * h
+
+        # Skip large objects
+        if exclude_large_object_ratio is not None:
+            if area / total_area > exclude_large_object_ratio:
+                continue
+
+        rect = Rectangle((x0, y0), w, h, facecolor=obj.color, edgecolor="black", alpha=0.6)
+        ax.add_patch(rect)
+
+    ax.set_xlabel(xlabel)
+    ax.set_ylabel(ylabel)
+    ax.set_title(title)
+    ax.set_xlim([0, total_w])
+    ax.set_ylim([0, total_h])
+    ax.set_aspect("equal")
+    ax.grid(True)
 
 
 def plot_setup(
     config: SimulationConfig,
     objects: ObjectContainer,
-    exclude_object_list: list[SimulationObject] | None = None,
-    filename: str | Path | None = None,
-    axs: Sequence[Any] | None = None,
-    plot_legend: bool = True,
-    exclude_xy_plane_object_list: list[SimulationObject] | None = None,
-    exclude_yz_plane_object_list: list[SimulationObject] | None = None,
-    exclude_xz_plane_object_list: list[SimulationObject] | None = None,
-) -> Figure:
-    """Creates a visualization of the simulation setup showing objects in XY, XZ and YZ planes.
-
[…]
