### 2025-11-25 758eea4
bug fixes for permittivity array management, added anisotropic example
