### 2025-11-26 f1b58ee
added more documentation and improved cases
