+++ src/fdtdx/utils/plot_setup.py
@@ -1,123 +1,272 @@
-from typing import Optional
+from pathlib import Path
+from typing import Any, Literal
 
 import matplotlib.pyplot as plt
-from matplotlib.patches import Rectangle
+from matplotlib.figure import Figure
+from matplotlib.patches import Patch, Rectangle
 
 from fdtdx.config import SimulationConfig
 from fdtdx.fdtd.container import ObjectContainer
+from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
+from fdtdx.objects.boundaries.periodic import PeriodicBoundary
+from fdtdx.objects.object import SimulationObject
 
 
 def plot_setup_from_side(
     config: SimulationConfig,
     objects: ObjectContainer,
-    resolution: float,
-    ax,
-    viewing_side: str,
-    exclude_large_object_ratio: Optional[float] = None,
-):
-    """
-    Plot a 2D projection of the simulation setup from a chosen axis.
+    viewing_side: Literal["x", "y", "z"],
+    exclude_object_list: list[SimulationObject] | None = None,
+    filename: str | Path | None = None,
+    ax: Any | None = None,
+    plot_legend: bool = True,
+    exclude_xy_plane_object_list: list[SimulationObject] | None = None,
+    exclude_yz_plane_object_list: list[SimulationObject] | None = None,
+    exclude_xz_plane_object_list: list[SimulationObject] | None = None,
+    exclude_large_object_ratio: float | None = None,
+) -> Figure:
+    """Creates a visualization of the simulation setup from a single viewing side.
 
-    viewing_side: one of 'x', 'y', 'z' indicating which projection to plot.
-    exclude_large_object_ratio: if given, objects that cover more than this
-        fraction of the image area will be excluded.
-    """
+    Generates a single subplot showing a cross-section of the simulation volume and the objects
+    within it from the specified viewing side. Objects are drawn as colored rectangles with
+    optional legends.
 
+    Args:
+        config (SimulationConfig): Configuration object containing simulation parameters like resolution
+        objects (ObjectContainer): Container holding all simulation objects to be plotted
+        viewing_side (Literal['x', 'y', 'z']): Which plane to view ('x' for YZ, 'y' for XZ, 'z' for XY)
+        exclude_object_list (list[SimulationObject] | None, optional): List of objects to exclude from all plots
+        filename (str | Path | None, optional): If provided, saves the plot to this file instead of displaying
+        ax (Any | None, optional): Optional matplotlib axis to plot on. If None, creates new figure
+        plot_legend (bool, optional): Whether to add a legend showing object names/types
+        exclude_xy_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from XY plane plot
+        exclude_yz_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from YZ plane plot
+        exclude_xz_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from XZ plane plot
+        exclude_large_object_ratio (float | None, optional): If provided, excludes objects that cover more than
+            this ratio of the image (e.g., 1.0 excludes objects covering 100% of the image)
+
+    Returns:
+        Figure: The generated figure object
+
+    Note:
+        The plots show object positions in micrometers, converting from simulation units.
+        PML objects are automatically excluded from their respective boundary planes.
+    """
+    # default to empty lists
+    exclude_object_list = exclude_object_list or []
+    exclude_xy_plane_object_list = exclude_xy_plane_object_list or []
[…]
