### 2025-11-26 5334f27
added tests for plot_setup
