+++ tests/utils/test_plot_setup.py
@@ -1,0 +1,250 @@
+from pathlib import Path
+
+import matplotlib
+import pytest
+
+matplotlib.use("Agg")  # Use non-interactive backend for testing
+import jax
+import matplotlib.pyplot as plt
+
+import fdtdx
+from fdtdx import SimulationVolume
+from fdtdx.config import SimulationConfig
+from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
+from fdtdx.objects.object import GridCoordinateConstraint, SimulationObject
+from fdtdx.utils.plot_setup import plot_setup, plot_setup_from_side
+
+# Ensure test output directory exists
+TEST_OUTPUT_DIR = Path("tests/generated")
+TEST_OUTPUT_DIR.mkdir(parents=True, exist_ok=True)
+
+
+@pytest.fixture
+def simulation_setup():
+    """Fixture that creates a simple simulation setup with various objects."""
+    # Create a basic configuration
+    config = SimulationConfig(
+        resolution=20e-9,  # 20 nm resolution
+        time=100e-15,  # 100 fs total simulation time
+    )
+
+    # Create simulation volume (10 x 10 x 10 µm = 10e-6 meters)
+    volume = SimulationVolume(
+        partial_real_shape=(10e-6, 10e-6, 10e-6),  # 10 µm in each direction
+        name="simulation_volume",
+    )
+
+    # Add a large background object (covers most of the volume)
+    class LargeObject(SimulationObject):
+        def __init__(self):
+            super().__init__(
+                partial_real_shape=(480 * 20e-9, 480 * 20e-9, 480 * 20e-9),  # 480 grid points * resolution
+                partial_grid_shape=(10, 10, 10),
+                name="large_background",
+            )
+
+    # Add a smaller centered object
+    class CenterObject(SimulationObject):
+        def __init__(self):
+            super().__init__(
+                partial_real_shape=(100 * 20e-9, 100 * 20e-9, 100 * 20e-9),  # 100 grid points * resolution
+                partial_grid_shape=(200, 200, 200),
+                name="center_object",
+            )
+
+    # Add a thin vertical object
+    class VerticalObject(SimulationObject):
+        def __init__(self):
+            super().__init__(
+                partial_real_shape=(50 * 20e-9, 50 * 20e-9, 200 * 20e-9),  # grid points * resolution
+                partial_grid_shape=(350, 350, 150),
+                name="vertical_object",
+            )
+
+    # Add a horizontal slab
+    class HorizontalSlab(SimulationObject):
+        def __init__(self):
+            super().__init__(
+                partial_real_shape=(200 * 20e-9, 200 * 20e-9, 30 * 20e-9),  # grid points * resolution
+                partial_grid_shape=(100, 100, 400),
+                name="horizontal_slab",
+            )
+
+    # Add PML boundaries (they don't have color attribute, so won't be plotted)
+    pml_thickness = 20
+
+    # X boundaries
+    PerfectlyMatchedLayer(partial_grid_shape=(pml_thickness, 500, 500), axis=0, direction="low", name="pml_x_low")
+
+    PerfectlyMatchedLayer(partial_grid_shape=(pml_thickness, 500, 500), axis=0, direction="high", name="pml_x_high")
+
+    # Y boundaries
+    PerfectlyMatchedLayer(partial_grid_shape=(500, pml_thickness, 500), axis=1, direction="low", name="pml_y_low")
+
+    PerfectlyMatchedLayer(partial_grid_shape=(500, pml_thickness, 500), axis=1, direction="high", name="pml_y_high")
+
+    # Z boundaries
[…]
