### 2025-11-27 e85fcde
refactored the test and improved coverage
