### 2025-11-27 8f069b2
added colors in the objects
