+++ src/fdtdx/objects/object.py
@@ -798,6 +798,7 @@ class OrderableObject(SimulationObject):
                     partial_grid_shape=(10, 10, 10),
                     name="large_background",
                 )
+                self.color = "#e0e0e0"
 
         # Add a smaller centered object
         class CenterObject(SimulationObject):
@@ -807,6 +808,7 @@ class OrderableObject(SimulationObject):
                     partial_grid_shape=(200, 200, 200),
                     name="center_object",
                 )
+                self.color = "#ff5722"
 
         # Add a thin vertical object
         class VerticalObject(SimulationObject):
@@ -816,6 +818,7 @@ class OrderableObject(SimulationObject):
                     partial_grid_shape=(350, 350, 150),
                     name="vertical_object",
                 )
+                self.color = "#2196f3"
 
         # Add a horizontal slab
         class HorizontalSlab(SimulationObject):
@@ -825,3 +828,4 @@ class OrderableObject(SimulationObject):
                     partial_grid_shape=(100, 100, 400),
                     name="horizontal_slab",
                 )
+                self.color = "#4caf50"
