### 2025-11-28 2e321c5
refactored plot_setup (#197)
* refactored plot_setup by adding a helper function plot_setup_from_side

* added more documentation and improved cases

* added tests for plot_setup

* refactored the test and improved coverage

* added colors in the objects
