+++ src/fdtdx/utils/plot_setup.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Any, Sequence
+from typing import Any, Literal
 
 import matplotlib.pyplot as plt
 from matplotlib.figure import Figure
@@ -12,33 +12,38 @@ from fdtdx.objects.boundaries.periodic import PeriodicBoundary
 from fdtdx.objects.object import SimulationObject
 
 
-def plot_setup(
+def plot_setup_from_side(
     config: SimulationConfig,
     objects: ObjectContainer,
+    viewing_side: Literal["x", "y", "z"],
     exclude_object_list: list[SimulationObject] | None = None,
     filename: str | Path | None = None,
-    axs: Sequence[Any] | None = None,
+    ax: Any | None = None,
     plot_legend: bool = True,
     exclude_xy_plane_object_list: list[SimulationObject] | None = None,
     exclude_yz_plane_object_list: list[SimulationObject] | None = None,
     exclude_xz_plane_object_list: list[SimulationObject] | None = None,
+    exclude_large_object_ratio: float | None = None,
 ) -> Figure:
-    """Creates a visualization of the simulation setup showing objects in XY, XZ and YZ planes.
+    """Creates a visualization of the simulation setup from a single viewing side.
 
-    Generates three subplots showing cross-sections of the simulation volume and the objects
-    within it. Objects are drawn as colored rectangles with optional legends. The visualization
-    helps verify the correct positioning and sizing of objects in the simulation setup.
+    Generates a single subplot showing a cross-section of the simulation volume and the objects
+    within it from the specified viewing side. Objects are drawn as colored rectangles with
+    optional legends.
 
     Args:
         config (SimulationConfig): Configuration object containing simulation parameters like resolution
         objects (ObjectContainer): Container holding all simulation objects to be plotted
+        viewing_side (Literal['x', 'y', 'z']): Which plane to view ('x' for YZ, 'y' for XZ, 'z' for XY)
         exclude_object_list (list[SimulationObject] | None, optional): List of objects to exclude from all plots
         filename (str | Path | None, optional): If provided, saves the plot to this file instead of displaying
-        axs (Sequence[Any] | None, optional): Optional matplotlib axes to plot on. If None, creates new figure
+        ax (Any | None, optional): Optional matplotlib axis to plot on. If None, creates new figure
         plot_legend (bool, optional): Whether to add a legend showing object names/types
         exclude_xy_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from XY plane plot
         exclude_yz_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from YZ plane plot
         exclude_xz_plane_object_list (list[SimulationObject] | None, optional): Objects to exclude from XZ plane plot
+        exclude_large_object_ratio (float | None, optional): If provided, excludes objects that cover more than
+            this ratio of the image (e.g., 1.0 excludes objects covering 100% of the image)
 
     Returns:
         Figure: The generated figure object
@@ -63,20 +68,60 @@ def plot_setup(
             exclude_xz_plane_object_list.append(o)
         elif o.axis == 2:
             exclude_xy_plane_object_list.append(o)
+
     # add volume to exclude list
     volume = objects.volume
     exclude_object_list.append(volume)
 
     object_list = [o for o in objects.objects if o not in exclude_object_list]
-    if axs is None:
-        fig, axs = plt.subplots(1, 3, figsize=(15, 5))
+
+    if ax is None:
[…]
