+++ tests/utils/test_plot_setup.py
@@ -1,0 +1,259 @@
+from pathlib import Path
+
+import matplotlib
+import pytest
+
+matplotlib.use("Agg")  # Use non-interactive backend for testing
+import jax
+import matplotlib.pyplot as plt
+
+import fdtdx
+from fdtdx import SimulationVolume
+from fdtdx.config import SimulationConfig
+from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
+from fdtdx.objects.object import GridCoordinateConstraint, OrderableObject
+from fdtdx.utils.plot_setup import plot_setup, plot_setup_from_side
+
+# Ensure test output directory exists
+TEST_OUTPUT_DIR = Path("tests/generated")
+TEST_OUTPUT_DIR.mkdir(parents=True, exist_ok=True)
+
+
+@pytest.fixture
+def simulation_setup():
+    """Fixture that creates a simple simulation setup with various objects."""
+    # Create a basic configuration
+    config = SimulationConfig(
+        resolution=20e-9,  # 20 nm resolution
+        time=100e-15,  # 100 fs total simulation time
+    )
+
+    # Create simulation volume (10 x 10 x 10 µm = 10e-6 meters)
+    volume = SimulationVolume(
+        partial_real_shape=(10e-6, 10e-6, 10e-6),  # 10 µm in each direction
+        name="simulation_volume",
+    )
+
+    # Add PML boundaries (they don't have color attribute, so won't be plotted)
+    pml_thickness = 20
+
+    # X boundaries
+    PerfectlyMatchedLayer(partial_grid_shape=(pml_thickness, 500, 500), axis=0, direction="low", name="pml_x_low")
+
+    PerfectlyMatchedLayer(partial_grid_shape=(pml_thickness, 500, 500), axis=0, direction="high", name="pml_x_high")
+
+    # Y boundaries
+    PerfectlyMatchedLayer(partial_grid_shape=(500, pml_thickness, 500), axis=1, direction="low", name="pml_y_low")
+
+    PerfectlyMatchedLayer(partial_grid_shape=(500, pml_thickness, 500), axis=1, direction="high", name="pml_y_high")
+
+    # Z boundaries
+    PerfectlyMatchedLayer(partial_grid_shape=(500, 500, pml_thickness), axis=2, direction="low", name="pml_z_low")
+
+    PerfectlyMatchedLayer(partial_grid_shape=(500, 500, pml_thickness), axis=2, direction="high", name="pml_z_high")
+
+    # Create all objects
+    large_obj = OrderableObject.EmptySimulationObject.LargeObject()
+    center_obj = OrderableObject.EmptySimulationObject.CenterObject()
+    vertical_obj = OrderableObject.EmptySimulationObject.VerticalObject()
+    horizontal_obj = OrderableObject.EmptySimulationObject.HorizontalSlab()
+
+    # Create object list (without PML - they don't have colors and won't be plotted anyway)
+    object_list = [
+        volume,
+        large_obj,
+        center_obj,
+        vertical_obj,
+        horizontal_obj,
+    ]
+
+    # Create constraints for positioning - use GridCoordinateConstraint for absolute positioning
+    # Each object needs constraints for all 6 sides (x-, x+, y-, y+, z-, z+)
+    constraints = [
+        # large_background: from (10, 10, 10) to (490, 490, 490)
+        GridCoordinateConstraint(object="large_background", axes=(0,), sides=("-",), coordinates=(10,)),
+        GridCoordinateConstraint(object="large_background", axes=(0,), sides=("+",), coordinates=(490,)),
+        GridCoordinateConstraint(object="large_background", axes=(1,), sides=("-",), coordinates=(10,)),
[…]
