+++ src/fdtdx/fdtd/initialization.py
@@ -22,22 +22,15 @@ from fdtdx.objects.object import (
     SizeConstraint,
     SizeExtensionConstraint,
 )
-from fdtdx.objects.static_material.static import StaticMultiMaterialObject, UniformMaterialObject
-from fdtdx.typing import SliceTuple3D
+from fdtdx.objects.static_material.static import SimulationVolume, StaticMultiMaterialObject, UniformMaterialObject
+
+DEFAULT_MAX_ITER = 1000
 
 
 def place_objects(
-    volume: SimulationObject,
+    object_list: list[SimulationObject],
     config: SimulationConfig,
-    constraints: Sequence[
-        (
-            PositionConstraint
-            | SizeConstraint
-            | SizeExtensionConstraint
-            | GridCoordinateConstraint
-            | RealCoordinateConstraint
-        )
-    ],
+    constraints: Sequence[(PositionConstraint | SizeConstraint | SizeExtensionConstraint | GridCoordinateConstraint)],
     key: jax.Array,
 ) -> tuple[
     ObjectContainer,
@@ -49,75 +42,90 @@ def place_objects(
     """Places simulation objects according to specified constraints and initializes containers.
 
     Args:
-        volume (SimulationObject): The volume object defining the simulation boundaries
-        config (SimulationConfig): The simulation configuration
-        constraints (Sequence[PositionConstraint| SizeConstraint| SizeExtensionConstraint| GridCoordinateConstraint| RealCoordinateConstraint]):
-            Sequence of positioning and sizing constraints for objects
-        key (jax.Array): JAX random key for initialization
+        objects (list[SimulationObject]): List of all simulation objects, including the simulation volume.
+        config (SimulationConfig): Simulation configuration.
+        constraints (Sequence[Constraint]): List of positioning/sizing constraints referencing object names.
+        key (jax.Array): JAX random key for initialization.
 
     Returns:
-        tuple[ObjectContainer, ArrayContainer, ParameterContainer, SimulationConfig, dict[str, Any]]: A tuple containing
+        tuple[ObjectContainer, ArrayContainer, ParameterContainer, SimulationConfig, dict[str, Any]]:
+        A tuple containing:
             - ObjectContainer with placed simulation objects
             - ArrayContainer with initialized field arrays
             - ParameterContainer with device parameters
             - Updated SimulationConfig
             - Dictionary with additional initialization info
+
+    Raises:
+        ValueError: If constraint resolution fails for one or more objects.
     """
-    slice_tuple_dict = _resolve_object_constraints(
-        volume=volume,
+
+    # Step 1: Resolve constraints into grid slices
+    resolved_slices, errors = resolve_object_constraints(
+        objects=object_list,
         constraints=constraints,
         config=config,
     )
-    obj_list = list(slice_tuple_dict.keys())
 
-    # place objects on computed grid positions
+    # Step 2: Aggregate errors and raise if needed
+    failed = {name: msg for name, msg in errors.items() if msg}
+    if failed:
+        formatted = "\n".join(f"  - {name}: {msg}" for name, msg in failed.items())
+        raise ValueError(f"Failed to resolve object constraints:\n{formatted}")
+
+    # Step 3: Convert name → object for placement
+    object_map = {obj.name: obj for obj in object_list}
+    volume_name = _resolve_volume_name(object_map)
+    volume_obj = object_map[volume_name]
+
+    # Step 4: Place objects on grid based on resolved slice tuples
     placed_objects = []
-    for o in obj_list:
-        if o == volume:
+    for name, slice_tuple in resolved_slices.items():
+        if name == volume_obj.name:
             continue
+        obj = object_map[name]
         key, subkey = jax.random.split(key)
         placed_objects.append(
-            o.place_on_grid(
[…]

[…]
