### 2025-11-29 c88ea04
PML boundary layer refactor (#185)
