+++ src/fdtdx/objects/boundaries/perfectly_matched_layer.py
@@ -2,60 +2,14 @@ import jax
 import jax.numpy as jnp
 from typing_extensions import override
 
+from fdtdx.constants import c, eps0, eta0
 from fdtdx.core.jax.pytrees import autoinit, frozen_field
 from fdtdx.core.plotting.colors import DARK_GREY
-from fdtdx.objects.boundaries.boundary import BaseBoundary, BaseBoundaryState
-from fdtdx.objects.boundaries.utils import (
-    alpha_from_direction_axis,
-    kappa_from_direction_axis,
-    standard_sigma_from_direction_axis,
-)
+from fdtdx.objects.boundaries.boundary import BaseBoundary
 
 
 @autoinit
-class PMLBoundaryState(BaseBoundaryState):
-    """State container for PML boundary conditions.
-
-    Stores the auxiliary field variables and coefficients needed to implement
-    the convolutional perfectly matched layer (CPML) boundary conditions.
-    """
-
-    #: Auxiliary field for Ex component
-    psi_Ex: jax.Array
-
-    #: Auxiliary field for Ey component
-    psi_Ey: jax.Array
-
-    #: Auxiliary field for Ez component
-    psi_Ez: jax.Array
-
-    #:  Auxiliary field for Hx component
-    psi_Hx: jax.Array
-
-    #: Auxiliary field for Hy component
-    psi_Hy: jax.Array
-
-    #: Auxiliary field for Hz component
-    psi_Hz: jax.Array
-
-    #: Electric field scaling coefficient
-    bE: jax.Array
-
-    #: Magnetic field scaling coefficient
-    bH: jax.Array
-
-    #: Electric field update coefficient
-    cE: jax.Array
-
-    #: Magnetic field update coefficient
-    cH: jax.Array
-
-    #: PML stretching coefficient
-    kappa: jax.Array
-
-
-@autoinit
-class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
+class PerfectlyMatchedLayer(BaseBoundary):
     """Implements a Convolutional Perfectly Matched Layer (CPML) boundary condition.
 
     The CPML absorbs outgoing electromagnetic waves with minimal reflection by using
@@ -63,20 +17,80 @@ class PerfectlyMatchedLayer(BaseBoundary[PMLBoundaryState]):
     axis orientation and both positive/negative directions.
     """
 
-    #: Initial loss parameter for complex frequency shifting. Defaults to 1e-8.
-    alpha_start: float = frozen_field(default=1.0e-8)
+    #: Initial loss parameter for complex frequency shifting. Defaults to 0.01 * 2 * jnp.pi * c / wavelength * eps0 if not provided.
+    alpha_start: float | None = frozen_field(default=None)
 
-    #: Final loss parameter for complex frequency shifting. Defaults to 1e-8.
-    alpha_end: float = frozen_field(default=1.0e-8)
+    #: Final loss parameter for complex frequency shifting. Defaults to 0.0 if not provided.
+    alpha_end: float | None = frozen_field(default=None)
 
-    #: Initial kappa stretching coefficient. Defaults to 1.0.
-    kappa_start: float = frozen_field(default=1.0)
+    #: Polynomial order for alpha grading. Defaults to 1.0 if not provided.
+    alpha_order: float | None = frozen_field(default=None)
 
-    #: Final kappa stretching coefficient. Defaults to 1.5.
-    kappa_end: float = frozen_field(default=1.5)
+    #: Initial kappa stretching coefficient. Defaults to 0.0 if not provided.
+    kappa_start: float | None = frozen_field(default=None)
+
+    #: Final kappa stretching coefficient. Defaults to 0.0 if not provided.
[…]
