+++ tests/fdtd/test_condition.py
@@ -9,7 +9,6 @@ from fdtdx.fdtd.initialization import place_objects
 
 # Import the module to test
 from fdtdx.fdtd.stop_conditions import DetectorConvergenceCondition, EnergyThresholdCondition, TimeStepCondition
-from fdtdx.objects.boundaries.boundary import BaseBoundaryState
 from fdtdx.objects.detectors.detector import DetectorState
 from fdtdx.objects.detectors.energy import EnergyDetector
 from fdtdx.objects.static_material.static import SimulationVolume
@@ -22,20 +21,28 @@ class TestCondition:
         # Create dummy arrays with appropriate shapes
         E = jnp.ones((3, 10, 10, 10))  # 3D electric field
         H = jnp.ones((3, 10, 10, 10))  # 3D magnetic field
+        psi_E = jnp.zeros((6, 10, 10, 10))  # 3D auxiliary electric field
+        psi_H = jnp.zeros((6, 10, 10, 10))  # 3D auxiliary magnetic field
+        alpha = jnp.zeros((3, 10, 10, 10))  # 3D alpha array
+        kappa = jnp.ones((3, 10, 10, 10))  # 3D kappa array
+        sigma = jnp.zeros((3, 10, 10, 10))  # 3D sigma array
         inv_permittivities = jnp.ones((10, 10, 10))
         inv_permeabilities = jnp.ones((10, 10, 10))
 
-        # Create mock boundary and detector states
-        boundary_states = {"pml": BaseBoundaryState()}
+        # Create mock detector states
         detector_states = {"detector1": DetectorState()}
 
         # Create array container
         arrays = ArrayContainer(
             E=E,
             H=H,
+            psi_E=psi_E,
+            psi_H=psi_H,
+            alpha=alpha,
+            kappa=kappa,
+            sigma=sigma,
             inv_permittivities=inv_permittivities,
             inv_permeabilities=inv_permeabilities,
-            boundary_states=boundary_states,
             detector_states=detector_states,
             recording_state=None,
             electric_conductivity=None,
@@ -311,7 +318,7 @@ class TestCondition:
         state_converged = (jnp.array(min_steps + 2), arrays)
         cond_fun = cond_fun.setup(state_converged, config, objects)
         # TODO(ymahlau): this test case did not work anymore for some reason
-        # assert not cond_fun(state_converged, config, objects)
+        # assert cond_fun(state_converged, config, objects)
 
         # Test at end_step -> should stop regardless of convergence
         state_at_end = (jnp.array(config.time_steps_total), arrays)

[…]
