### 2025-12-01 0dbb9e1
added plot_material and its helper function plot_material_from_side
