+++ src/fdtdx/utils/plot_setup.py
@@ -2,14 +2,216 @@ from pathlib import Path
 from typing import Any, Literal
 
 import matplotlib.pyplot as plt
+import numpy as np
 from matplotlib.figure import Figure
 from matplotlib.patches import Patch, Rectangle
 
 from fdtdx.config import SimulationConfig
-from fdtdx.fdtd.container import ObjectContainer
+from fdtdx.fdtd.container import ArrayContainer, ObjectContainer
 from fdtdx.objects.boundaries.perfectly_matched_layer import PerfectlyMatchedLayer
 from fdtdx.objects.boundaries.periodic import PeriodicBoundary
 from fdtdx.objects.object import SimulationObject
+
+MaterialType = Literal["permittivity", "permeability"]
+
+
+def plot_material_from_side(
+    config: SimulationConfig,
+    arrays: ArrayContainer,
+    viewing_side: Literal["x", "y", "z"],
+    filename: str | Path | None = None,
+    ax: Any | None = None,
+    plot_legend: bool = True,
+    position: float = 0.0,
+    type: MaterialType = "permittivity",
+) -> Figure:
+    """Creates a visualization of material distribution from a single viewing side.
+
+    Generates a single subplot showing a 2D slice of the material distribution
+    (permittivity or permeability) through the 3D simulation volume at a specified position.
+
+    Args:
+        config (SimulationConfig): Configuration object containing simulation parameters like resolution
+        arrays (ArrayContainer): Container holding the material arrays (permittivity, permeability)
+        viewing_side (Literal['x', 'y', 'z']): Which plane to view ('x' for YZ, 'y' for XZ, 'z' for XY)
+        filename (str | Path | None, optional): If provided, saves the plot to this file instead of displaying
+        ax (Any | None, optional): Optional matplotlib axis to plot on. If None, creates new figure
+        plot_legend (bool, optional): Whether to add a colorbar legend
+        position (float, optional): Position of the slice in meters. Zero means at center,
+            1e-6 would mean center+1µm
+        type (MaterialType, optional): Type of material to plot, either "permittivity" or "permeability"
+
+    Returns:
+        Figure: The generated figure object
+
+    Note:
+        The plots show material values in a 2D cross-section, with positions in micrometers.
+    """
+    if ax is None:
+        fig, ax = plt.subplots(1, 1, figsize=(6, 5))
+    else:
+        fig = None
+
+    # Get the appropriate material array
+    if type == "permittivity":
+        # Invert to get actual permittivity from inverse permittivity
+        material_array = 1.0 / arrays.inv_permittivities
+    else:  # type == "permeability"
+        # Invert to get actual permeability from inverse permeability
+        # Note: inv_permeabilities can be a float or array
+        if isinstance(arrays.inv_permeabilities, float):
+            material_array = np.full_like(arrays.inv_permittivities, 1.0 / arrays.inv_permeabilities)
+        else:
+            material_array = 1.0 / arrays.inv_permeabilities
+
+    resolution = config.resolution / 1.0e-6  # Convert to µm
+
+    # Calculate slice index from position
+    slice_offset = round(position / config.resolution)
+
+    # Get array shape
+    array_shape = material_array.shape
+
+    # Determine slice parameters based on viewing side
+    if viewing_side == "z":
+        # XY plane - slice through Z axis
+        center_idx = array_shape[2] // 2
+        slice_idx = center_idx + slice_offset
+        slice_idx = max(0, min(slice_idx, array_shape[2] - 1))
+        material_slice = material_array[:, :, slice_idx]
[…]
