+++ tests/units/test_unitful.py
@@ -451,16 +451,17 @@ def test_le_untiful_jitted():
 
 
 def test_le_jitted_static():
+    """Test less than or equal within a JIT-compiled function when input as unitful with static arrays"""
     length_unit = Unit(scale=-2, dim={SI.m: 1})
     lengths1 = Unitful(
-        val=jnp.array([12.0, 5.5, 18.3, 30.1, 22.7]),
+        val=jnp.array([12.0, 5.5, 18.3, 30.1]),
         unit=length_unit,
-        static_arr=np.array([12.0, 5.5, 18.3, 30.1, 22.7]),
+        static_arr=np.array([12.0, 5.5, 18.3, 30.1]),
     )
     lengths2 = Unitful(
-        val=jnp.array([12.0, 14.5, 1.3, 3012.1, 22.7]),
+        val=jnp.array([12.0, 14.5, 1.3, 3012.1]),
         unit=length_unit,
-        static_arr=np.array([12.0, 14.5, 100.3, 3012.1, 22.7]),
+        static_arr=np.array([12.0, 14.5, 100.3, 3012.1]),
     )
 
     def fn(x: Unitful, y: Unitful) -> Unitful:
@@ -478,7 +479,7 @@ def test_le_jitted_static():
     assert isinstance(result, Unitful)
     assert isinstance(result.val, jax.Array)
     assert isinstance(result.static_arr, np.ndarray)
-    assert jnp.allclose(result.val, jnp.array([True, True, False, True, True]))
+    assert jnp.allclose(result.val, jnp.array([True, True, False, True]))
 
 
 def test_le_different_units_raises_error():
@@ -2364,7 +2365,7 @@ def test_argmax_unitful_with_StaticScalar():
     assert isinstance(result_np0darray, Unitful)
     assert isinstance(result_jax0darray, Unitful)
 
-    assert isinstance(result_float.val, jax.Array)
+    assert isinstance(result_float.val, np.number)
     assert isinstance(result_np0darray.val, np.integer)
     assert isinstance(result_jax0darray.val, jax.Array)
 
