### 2025-12-01 e46d4bf
Unit Functions Improvements requested in issue #174 (#183)
* Implement first three improvements requested in issue #174

* Implement improvements from issue #174, added test code

* fix bug in unary_fn

* adjusted test code
