+++ src/fdtdx/units/unitful.py
@@ -1,6 +1,8 @@
 # ruff: noqa: F811
 
 from dataclasses import dataclass
+
+# from numbers import Number
 from typing import Any, Self
 
 import jax
@@ -550,7 +552,7 @@ def get_static_operand(
     return x_arr
 
 
-def can_perform_scatic_ops(x: StaticArrayLike | None):
+def can_perform_static_ops(x: StaticArrayLike | None):
     if x is None:
         return False
     if isinstance(x, NonPhysicalArrayLike):
@@ -642,6 +644,7 @@ def multiply(x: ArrayLike, y: Unitful) -> Unitful:
         unit_x = Unitful(val=x, unit=Unit(scale=0, dim={}))
         x = unit_x.val
         scale_offset = unit_x.unit.scale
+
     new_val = y.val * x
     new_scale = y.unit.scale + scale_offset
     new_static_arr = None
@@ -667,13 +670,48 @@ def multiply(x: int, y: int) -> int:
 
 
 @overload
+def multiply(x: int, y: float) -> float:
+    return x * y
+
+
+@overload
+def multiply(x: float, y: int) -> float:
+    return x * y
+
+
+@overload
 def multiply(x: float, y: float) -> float:
     return x * y
 
 
 @overload
-def multiply(x: complex, y: complex) -> complex:
+def multiply(x: complex, y: int | float) -> complex:
     return x * y
+
+
+@overload
+def multiply(x: int | float | complex, y: complex) -> complex:
+    return x * y
+
+
+@overload
+def multiply(x: int | float | complex | np.number, y: np.ndarray) -> np.ndarray:
+    return x * y
+
+
+@overload
+def multiply(x: np.ndarray, y: int | float | complex | np.number) -> np.ndarray:
+    return x * y
+
+
+@overload
+def multiply(x: int | float | complex | np.number, y: jax.Array) -> jax.Array:
+    return jnp.asarray(x) * y
+
+
+@overload
+def multiply(x: jax.Array, y: int | float | complex | np.number) -> jax.Array:
+    return x * jnp.asarray(y)
 
 
 @overload
@@ -682,7 +720,18 @@ def multiply(x: jax.Array, y: jax.Array) -> jax.Array:
 
 
 @overload
-def multiply(x: ArrayLike, y: ArrayLike) -> ArrayLike:
+def multiply(x: jax.Array, y: np.ndarray) -> jax.Array:
+    return x * y
+
+
+@overload
+def multiply(x: np.ndarray, y: jax.Array) -> jax.Array:
+    x_jax = jnp.asarray(x)
+    return x_jax * y
+
+
+@overload
+def multiply(x: np.ndarray, y: np.ndarray) -> np.ndarray:
     return x * y
 
 
@@ -760,13 +809,48 @@ def divide(x1: int, x2: int) -> float:
 
 
 @overload
+def divide(x1: int, x2: float) -> float:
+    return x1 / x2
+
+
+@overload
+def divide(x1: float, x2: int) -> float:
+    return x1 / x2
+
+
+@overload
 def divide(x1: float, x2: float) -> float:
     return x1 / x2
 
 
 @overload
-def divide(x1: complex, x2: complex) -> complex:
+def divide(x1: int | float, x2: complex) -> complex:
     return x1 / x2
+
+
+@overload
+def divide(x1: complex, x2: int | float | complex) -> complex:
+    return x1 / x2
+
+
+@overload
+def divide(x: int | float | complex | np.number, y: np.ndarray) -> np.ndarray:
+    return x / y
+
+
+@overload
+def divide(x: np.ndarray, y: int | float | complex | np.number) -> np.ndarray:
[…]
