+++ tests/units/test_unitful.py
@@ -1,3 +1,6 @@
+import cmath
+import math
+
 import jax
 import jax.numpy as jnp
 import numpy as np
@@ -22,6 +25,9 @@ from fdtdx.units.unitful import (
     matmul,
     multiply,
     ne,
+    pow,
+    reshape,
+    squeeze,
     subtract,
 )
 
@@ -44,7 +50,7 @@ def test_multiply_unitful_unitful_different_dimensions():
 
     result = multiply(time_val, freq_val)
 
-    assert jnp.allclose(result.value(), 50.0)
+    assert math.isclose(result.value(), 50.0)  # type: ignore
     assert result.unit.dim == {}  # s^1 * s^-1 = dimensionless
 
 
@@ -55,36 +61,41 @@ def test_multiply_unitful_unitful_different_scales():
 
     result = multiply(time1, time2)
 
-    assert jnp.allclose(result.value(), 6.0e-3)
+    assert math.isclose(result.value(), 6.0e-3)  # type: ignore
     assert result.unit.dim == {SI.s: 2}
 
 
-def test_multiply_arraylike_unitful():
-    """Test multiplication of ArrayLike with Unitful"""
-    scalar = 5.0
-    time_val = 3 * s
+def test_multiply_unitful_unitful_python_python_type_and_dim():
+    """Test multiplication of Unitful objects with value Python scalar and Python scalar"""
+    a = 2 * s
+    b = 3.5 * s
+    res = multiply(a, b)
 
-    result: Unitful = jnp.multiply(scalar, time_val)  # type: ignore
-
-    assert jnp.allclose(result.value(), 15.0)
-    assert result.unit.dim == {SI.s: 1}
-
-
-def test_multiply_arraylike_arraylike():
-    """Test multiplication of two ArrayLike objects"""
-    arr1 = jnp.array([1.0, 2.0, 3.0])
-    arr2 = jnp.array([4.0, 5.0, 6.0])
-
-    result = multiply(arr1, arr2)
-
-    expected = jnp.array([4.0, 10.0, 18.0])
-    assert jnp.allclose(result, expected)
+    assert math.isclose(res.value(), 7.0)  # type: ignore
+    assert res.unit.dim == {SI.s: 2}
+    assert isinstance(res.val, float)
 
 
-def test_multiply_not_implemented():
-    """Test that multiply raises NotFoundLookupError for unsupported types"""
-    with pytest.raises(plum.NotFoundLookupError):
-        multiply("string", "string")  # type: ignore
+def test_multiply_unitful_unitful_numpy_numpy_type_and_dim():
+    """Test multiplication of Unitful objects with value numpy Array and numpy Array"""
+    a = s * np.array([1.0, 2.0, 3.0])
+    b = s * np.array([4.0, 5.0, 6.0])
+    res = multiply(a, b)
+
+    assert isinstance(res.val, np.ndarray)
+    assert jnp.allclose(res.value(), jnp.array([4.0, 10.0, 18.0]))
+    assert res.unit.dim == {SI.s: 2}
+
+
+def test_multiply_unitful_unitful_jax_jax_type_and_dim():
+    """Test multiplication of Unitful objects with value Jax Array and Jax Array"""
+    a = s * jnp.array([1.0, 2.0, 3.0])
+    b = s * jnp.array([4.0, 5.0, 6.0])
+    res = multiply(a, b)
+
+    assert isinstance(res.val, jax.Array)
+    assert jnp.allclose(res.value(), jnp.array([4.0, 10.0, 18.0]))
+    assert res.unit.dim == {SI.s: 2}
 
 
[…]
