### 2025-12-02 #8912 @ymahlau
@fruzsinaagocs what would be the next steps since the last issue of the review has been resolved?
