### 2025-12-02 882b514
added tests for plot_material
