### 2025-12-02 35c0a54
added plot_material and its helper function plot_material_from_side (#202)
* added plot_material and its helper function plot_material_from_side

* added tests for plot_material
