+++ tests/utils/test_plot_material.py
@@ -1,0 +1,597 @@
+from pathlib import Path
+
+import matplotlib
+import pytest
+
+matplotlib.use("Agg")  # Use non-interactive backend for testing
+import jax
+import matplotlib.pyplot as plt
+
+import fdtdx
+from fdtdx import Cylinder, SimulationVolume, Sphere
+from fdtdx.config import SimulationConfig
+from fdtdx.objects.object import GridCoordinateConstraint
+from fdtdx.objects.static_material.static import Material, UniformMaterialObject
+from fdtdx.utils.plot_material import plot_material, plot_material_from_side
+
+# Ensure test output directory exists
+TEST_OUTPUT_DIR = Path("tests/generated")
+TEST_OUTPUT_DIR.mkdir(parents=True, exist_ok=True)
+
+
+@pytest.fixture
+def simple_material_setup():
+    """A simple setup with basic material objects for testing."""
+    # Create a basic configuration
+    config = SimulationConfig(
+        resolution=50e-9,  # 50 nm resolution
+        time=100e-15,  # 100 fs total simulation time
+    )
+
+    # Create simulation volume (2 x 2 x 2 µm = 2e-6 meters)
+    volume = SimulationVolume(
+        partial_real_shape=(2e-6, 2e-6, 2e-6),
+        name="simulation_volume",
+    )
+
+    # Create a simple material object
+    material = Material(permittivity=2.25, permeability=1.0)
+
+    test_object = UniformMaterialObject(
+        name="test_object",
+        material=material,
+    )
+
+    # Create object list
+    object_list = [
+        volume,
+        test_object,
+    ]
+
+    # Create constraints for positioning
+    constraints = [
+        # Volume - fill entire space
+        GridCoordinateConstraint(object="simulation_volume", axes=(0,), sides=("-",), coordinates=(0,)),
+        GridCoordinateConstraint(object="simulation_volume", axes=(0,), sides=("+",), coordinates=(40,)),
+        GridCoordinateConstraint(object="simulation_volume", axes=(1,), sides=("-",), coordinates=(0,)),
+        GridCoordinateConstraint(object="simulation_volume", axes=(1,), sides=("+",), coordinates=(40,)),
+        GridCoordinateConstraint(object="simulation_volume", axes=(2,), sides=("-",), coordinates=(0,)),
+        GridCoordinateConstraint(object="simulation_volume", axes=(2,), sides=("+",), coordinates=(40,)),
+        # Test object: centered cube
+        GridCoordinateConstraint(object="test_object", axes=(0,), sides=("-",), coordinates=(15,)),
+        GridCoordinateConstraint(object="test_object", axes=(0,), sides=("+",), coordinates=(25,)),
+        GridCoordinateConstraint(object="test_object", axes=(1,), sides=("-",), coordinates=(15,)),
+        GridCoordinateConstraint(object="test_object", axes=(1,), sides=("+",), coordinates=(25,)),
+        GridCoordinateConstraint(object="test_object", axes=(2,), sides=("-",), coordinates=(15,)),
+        GridCoordinateConstraint(object="test_object", axes=(2,), sides=("+",), coordinates=(25,)),
+    ]
+
+    # Use place_objects to create the initialized container
+    key = jax.random.PRNGKey(0)
+    objects, arrays, params, config, _ = fdtdx.place_objects(
+        object_list=object_list,
+        config=config,
+        constraints=constraints,
+        key=key,
+    )
+
+    return config, arrays
+
+
+@pytest.fixture
+def cylinder_setup():
+    """Setup with a cylinder object."""
+    config = SimulationConfig(
+        resolution=30e-9,
+        time=100e-15,
+    )
+
+    volume = SimulationVolume(
[…]
