### 2025-12-02 9f6d0ef
version bump 0.5.0
