+++ tests/core/physics/test_metrics.py
@@ -34,16 +34,20 @@ def test_compute_energy_basic_anisotropic():
     # Create simple uniform fields
     E = jnp.ones((3, 4, 4, 4))
     H = jnp.ones((3, 4, 4, 4))
-    inv_permittivity = jnp.stack([
-        jnp.full((4, 4, 4), 1.0),
-        jnp.full((4, 4, 4), 0.5),
-        jnp.full((4, 4, 4), 1.0),
-    ])
-    inv_permeability = jnp.stack([
-        jnp.full((4, 4, 4), 1.0),
-        jnp.full((4, 4, 4), 1.0),
-        jnp.full((4, 4, 4), 1.0),
-    ])
+    inv_permittivity = jnp.stack(
+        [
+            jnp.full((4, 4, 4), 1.0),
+            jnp.full((4, 4, 4), 0.5),
+            jnp.full((4, 4, 4), 1.0),
+        ]
+    )
+    inv_permeability = jnp.stack(
+        [
+            jnp.full((4, 4, 4), 1.0),
+            jnp.full((4, 4, 4), 1.0),
+            jnp.full((4, 4, 4), 1.0),
+        ]
+    )
 
     energy = compute_energy(E, H, inv_permittivity, inv_permeability)
 
@@ -80,12 +84,14 @@ def test_compute_energy_complex_fields_anisotropic():
     E = jnp.array([[[1 + 1j]], [[2 + 0j]], [[0 + 2j]]])
     H = jnp.array([[[1 - 1j]], [[1 + 1j]], [[2 + 0j]]])
     # Anisotropic permittivity
-    inv_permittivity = jnp.stack([
-        jnp.full((1, 1), 1/2.0),
-        jnp.full((1, 1), 1/1.5),
-        jnp.full((1, 1), 1/2.0),
-    ])
-    inv_permeability = 1/2.0
+    inv_permittivity = jnp.stack(
+        [
+            jnp.full((1, 1), 1 / 2.0),
+            jnp.full((1, 1), 1 / 1.5),
+            jnp.full((1, 1), 1 / 2.0),
+        ]
+    )
+    inv_permeability = 1 / 2.0
 
     energy = compute_energy(E, H, inv_permittivity, inv_permeability)
 
@@ -122,11 +128,13 @@ def test_compute_energy_spatially_varying_anisotropic_materials():
     H = jnp.ones((3, 2, 2, 2))
 
     # Create spatially varying anisotropic permittivity and isotropic permeability
-    inv_permittivity = jnp.stack([
-        jnp.array([[[1.0, 2.0], [3.0, 4.0]], [[5.0, 6.0], [7.0, 8.0]]]),
-        jnp.array([[[1.5, 2.5], [3.5, 4.5]], [[5.5, 6.5], [7.5, 8.5]]]),
-        jnp.array([[[2.0, 3.0], [4.0, 5.0]], [[6.0, 7.0], [8.0, 9.0]]]),
-    ])
+    inv_permittivity = jnp.stack(
+        [
+            jnp.array([[[1.0, 2.0], [3.0, 4.0]], [[5.0, 6.0], [7.0, 8.0]]]),
+            jnp.array([[[1.5, 2.5], [3.5, 4.5]], [[5.5, 6.5], [7.5, 8.5]]]),
+            jnp.array([[[2.0, 3.0], [4.0, 5.0]], [[6.0, 7.0], [8.0, 9.0]]]),
+        ]
+    )
     inv_permeability = jnp.array([[[0.5, 1.0], [1.5, 2.0]], [[2.5, 3.0], [3.5, 4.0]]])
 
     energy = compute_energy(E, H, inv_permittivity, inv_permeability)
[…]
