### 2025-12-06 5bd5fbc
added documentation for placeholder attributes
