### 2025-12-07 3c35bf3
descriptions directly above the attribute definitions
