+++ src/fdtdx/objects/sources/linear_polarization.py
@@ -13,15 +13,13 @@ from fdtdx.objects.sources.tfsf import TFSFPlaneSource
 
 @autoinit
 class LinearlyPolarizedPlaneSource(TFSFPlaneSource, ABC):
-    """ "
-    Args:
-        fixed_E_polarization_vector (np.array): the electric polarization vector
-        fixed_H_polarization_vector (np.array): the magnetic polarization vector
-        normalize_by_energy (bool): whether to normalize the polarization vector
-    """
-
+    #: the electric polarization vector
     fixed_E_polarization_vector: tuple[float, float, float] | None = frozen_field(default=None)
+
+    #: the magnetic polarization vector
     fixed_H_polarization_vector: tuple[float, float, float] | None = frozen_field(default=None)
+
+    #: whether to normalize the polarization vector
     normalize_by_energy: bool = frozen_field(default=True)
 
     def get_EH_variation(
@@ -137,13 +135,10 @@ class LinearlyPolarizedPlaneSource(TFSFPlaneSource, ABC):
 
 @autoinit
 class GaussianPlaneSource(LinearlyPolarizedPlaneSource):
-    """ "
-    Args:
-        radius (float): the radius of the gaussian source
-        std (float): the standard deviation of the gaussian source
-    """
-
+    #: the radius of the gaussian source
     radius: float = frozen_field()
+
+    #:  the standard deviation of the gaussian source
     std: float = frozen_field(default=1 / 3)  # relative to radius
 
     @staticmethod
@@ -189,11 +184,7 @@ class GaussianPlaneSource(LinearlyPolarizedPlaneSource):
 
 @autoinit
 class UniformPlaneSource(LinearlyPolarizedPlaneSource):
-    """ "
-    Args:
-        amplitude (float): the amplitude of the uniform source
-    """
-
+    #: the amplitude of the uniform source
     amplitude: float = frozen_field(default=1.0)
 
     def _get_amplitude_raw(
