### 2025-12-08 f9ed508
added documentation for placeholder attributes (#206)
* added documentation for placeholder attributes

* descriptions directly above the attribute definitions
