+++ src/fdtdx/objects/sources/tfsf.py
@@ -17,10 +17,19 @@ class TFSFPlaneSource(DirectionalPlaneSourceBase, ABC):
     positive offset of 0.25 in the yee grid in the axis of propagation.
     """
 
+    #: the azimuth angle
     azimuth_angle: float = frozen_field(default=0.0)
+
+    #: the elevation angle
     elevation_angle: float = frozen_field(default=0.0)
+
+    #: the max angle random offset
     max_angle_random_offset: float = frozen_field(default=0.0)
+
+    #: the max vertical offset
     max_vertical_offset: float = frozen_field(default=0.0)
+
+    #: the max horizontal offset
     max_horizontal_offset: float = frozen_field(default=0.0)
 
     _E: jax.Array = private_field()
