### 2025-12-08 1724bc1
Raise the minimum JAX version to 0.7
