+++ pyproject.toml
@@ -11,7 +11,7 @@ requires-python = ">=3.11"
 dependencies = [
     "loguru>=0.7.3",
     "pytreeclass>=0.11",
-    "jax>=0.4.36",
+    "jax>=0.7.0",
     "moviepy>=2.1.1",
     "optax>=0.2.4",
     "equinox>=0.11.10",
@@ -67,10 +67,13 @@ dev = [
     "myst_nb>=1.3.0",
 ]
 cuda12 = [
-    "jax[cuda12]>=0.4.36",
+    "jax[cuda12]>=0.7.0",
 ]
 rocm = [
-    "jax[rocm]>=0.4.36",
+    "jax==0.7.1",
+    "jax-rocm7-plugin==0.7.1",
+    "jax-rocm7-pjrt==0.7.1",
+    # pip install https://github.com/ROCm/jax/releases/download/rocm-jax-v0.7.1/jaxlib-0.7.1-cp312-cp312-manylinux_2_27_x86_64.manylinux_2_28_x86_64.whl
 ]
 
 [tool.pyright]
