### 2025-12-08 533067c
added plot_field_slice and its helper function plot_field_slice_component
