+++ src/fdtdx/utils/plot_field_slice.py
@@ -1,0 +1,170 @@
+from pathlib import Path
+from typing import Any
+
+import jax.numpy as jnp
+import matplotlib.pyplot as plt
+from matplotlib.figure import Figure
+
+
+def plot_field_slice_component(
+        field: jnp.ndarray,
+        component_name: str,
+        ax: Any,
+        plot_legend: bool = True,
+) -> None:
+    """Plots a single component of the electromagnetic field.
+
+    Args:
+        field (jnp.ndarray): 2D array of shape (w, h) containing field values
+        component_name (str): Name of the component (e.g., 'Ex', 'Hy')
+        ax (Any): Matplotlib axis to plot on
+        plot_legend (bool, optional): Whether to add a colorbar legend
+
+    Raises:
+        ValueError: If field is not 2D or contains invalid values
+    """
+    # Sanity checks
+    if field.ndim != 2:
+        raise ValueError(f"Field must be 2D, got shape {field.shape}")
+
+    if jnp.any(jnp.isnan(field)):
+        raise ValueError(f"{component_name} contains NaN values")
+
+    if jnp.any(jnp.isinf(field)):
+        raise ValueError(f"{component_name} contains infinite values")
+
+    # Plot the field component
+    im = ax.imshow(
+        field.T,  # Transpose for correct orientation
+        origin="lower",
+        aspect="equal",
+        cmap="RdBu_r",  # Red-blue colormap, centered at zero
+        interpolation="nearest",
+    )
+
+    # Set labels and title
+    ax.set_xlabel("x (grid points)")
+    ax.set_ylabel("y (grid points)")
+    ax.set_title(component_name)
+    ax.grid(True, alpha=0.3)
+
+    # Add colorbar if legend is requested
+    if plot_legend:
+        cbar = plt.colorbar(im, ax=ax)
+        cbar.set_label("Field value")
+
+
+def plot_field_slice(
+        E: jnp.ndarray,
+        H: jnp.ndarray,
+        filename: str | Path | None = None,
+        axs: Any | None = None,
+        plot_legend: bool = True,
+) -> Figure:
+    """Creates a visualization of electromagnetic field components.
+
+    Generates a 2x3 subplot showing all six components of the electromagnetic field
+    (Ex, Ey, Ez, Hx, Hy, Hz) in a single figure.
+
+    Args:
+        E (jnp.ndarray): Electric field array of shape (3, nx, ny, nz) or (3, w, h)
+        H (jnp.ndarray): Magnetic field array of shape (3, nx, ny, nz) or (3, w, h)
+        filename (str | Path | None, optional): If provided, saves the plot to this file
+        axs (Any | None, optional): Optional matplotlib axes to plot on. If None, creates new figure
+        plot_legend (bool, optional): Whether to add colorbar legends
+
+    Returns:
+        Figure: The generated figure object
+
+    Raises:
+        ValueError: If arrays have incorrect shapes or contain invalid values
+
+    Note:
+        For 4D inputs (3, nx, ny, nz), exactly one of nx, ny, nz must be 1.
+        The function will automatically squeeze out the singleton dimension.
+    """
+    # Sanity checks for E field
+    if E.shape[0] != 3:
+        raise ValueError(f"E field first dimension must be 3, got shape {E.shape}")
+
+    if E.ndim not in (3, 4):
+        raise ValueError(f"E field must be 3D (3, w, h) or 4D (3, nx, ny, nz), got shape {E.shape}")
+
+    # Sanity checks for H field
+    if H.shape[0] != 3:
+        raise ValueError(f"H field first dimension must be 3, got shape {H.shape}")
+
+    if H.ndim not in (3, 4):
[…]
