### 2025-12-08 8cd1626
added unit tests for plot_field_slice
