+++ tests/utils/test_plot_field_slice.py
@@ -1,0 +1,458 @@
+import tempfile
+from pathlib import Path
+
+import jax.numpy as jnp
+import matplotlib.pyplot as plt
+import pytest
+
+from fdtdx.utils.plot_field_slice import plot_field_slice, plot_field_slice_component
+
+
+class TestPlotFieldSliceComponent:
+    """Tests for plot_field_slice_component function."""
+
+    def test_valid_2d_field(self):
+        """Test plotting a valid 2D field component."""
+        field = jnp.array([[1.0, 2.0], [3.0, 4.0]])
+        fig, ax = plt.subplots()
+
+        # Should not raise any exception
+        plot_field_slice_component(field, "Ex", ax, plot_legend=True)
+        plt.close(fig)
+
+    def test_valid_2d_field_no_legend(self):
+        """Test plotting without colorbar legend."""
+        field = jnp.array([[1.0, 2.0], [3.0, 4.0]])
+        fig, ax = plt.subplots()
+
+        plot_field_slice_component(field, "Hy", ax, plot_legend=False)
+        plt.close(fig)
+
+    def test_negative_values(self):
+        """Test that negative values are handled correctly."""
+        field = jnp.array([[-1.0, 2.0], [3.0, -4.0]])
+        fig, ax = plt.subplots()
+
+        plot_field_slice_component(field, "Ez", ax)
+        plt.close(fig)
+
+    def test_zero_field(self):
+        """Test plotting a field with all zeros."""
+        field = jnp.zeros((5, 5))
+        fig, ax = plt.subplots()
+
+        plot_field_slice_component(field, "Ex", ax)
+        plt.close(fig)
+
+    def test_non_2d_field_raises_error(self):
+        """Test that non-2D field raises ValueError."""
+        field_1d = jnp.array([1.0, 2.0, 3.0])
+        fig, ax = plt.subplots()
+
+        with pytest.raises(ValueError, match="Field must be 2D"):
+            plot_field_slice_component(field_1d, "Ex", ax)
+        plt.close(fig)
+
+    def test_3d_field_raises_error(self):
+        """Test that 3D field raises ValueError."""
+        field_3d = jnp.ones((3, 4, 5))
+        fig, ax = plt.subplots()
+
+        with pytest.raises(ValueError, match="Field must be 2D"):
+            plot_field_slice_component(field_3d, "Ex", ax)
+        plt.close(fig)
+
+    def test_nan_values_raise_error(self):
+        """Test that NaN values raise ValueError."""
+        field = jnp.array([[1.0, jnp.nan], [3.0, 4.0]])
+        fig, ax = plt.subplots()
+
+        with pytest.raises(ValueError, match="contains NaN values"):
+            plot_field_slice_component(field, "Ex", ax)
+        plt.close(fig)
+
+    def test_inf_values_raise_error(self):
+        """Test that infinite values raise ValueError."""
+        field = jnp.array([[1.0, jnp.inf], [3.0, 4.0]])
+        fig, ax = plt.subplots()
+
+        with pytest.raises(ValueError, match="contains infinite values"):
+            plot_field_slice_component(field, "Ex", ax)
+        plt.close(fig)
+
+    def test_neg_inf_values_raise_error(self):
+        """Test that negative infinite values raise ValueError."""
+        field = jnp.array([[1.0, -jnp.inf], [3.0, 4.0]])
+        fig, ax = plt.subplots()
+
+        with pytest.raises(ValueError, match="contains infinite values"):
+            plot_field_slice_component(field, "Ex", ax)
+        plt.close(fig)
+
+
[…]
