### 2025-12-10 6bb68f9
updated changelog by adding the recent PR changes
