### 2025-12-12 2984535
updated changelog by adding the recent PR changes to unpublished
