### 2025-12-12 0637f9b
added plot_field_slice and its helper function plot_field_slice_component (#209)
* added plot_field_slice and its helper function plot_field_slice_component

* added unit tests for plot_field_slice

* updated changelog by adding the recent PR changes

* updated changelog by adding the recent PR changes to unpublished
