+++ changelog.md
@@ -1,6 +1,8 @@
 
 # Unpublished
 ## Added
+- added new description for the placeholder attributes (@renaissancenerd)
+- added plot_field_slice function which creates 2x3 images (with colorbar) for given E and H fields (@renaissancenerd)
 
 ## Changed
 
@@ -13,8 +15,7 @@
 - dependabot / codecov integration for better monitoring of repository
 - A lot of test cases (@renaissancenerd)
 - Json import / export of simulation objects. This will be necessary and helpful for the ongoing GUI development.
-- added plot_material function. This function plots the material distribution at a specific slice in the simulation.
-
+- added plot_material function. This function plots the material distribution at a specific slice in the simulation. (@renaissancenerd)
 
 ## Changed
 - moved docs to readthedocs using sphinx instead of mkdocs. This looks much nicer now.
