+++ src/fdtdx/fdtd/fdtd.py
@@ -63,13 +63,11 @@ def reversible_fdtd(
     def reversible_fdtd_base(
         arr: ArrayContainer,
     ) -> SimulationState:
-        def cond(s):
-            return config.time_steps_total > s[0]
 
         state = (jnp.asarray(0, dtype=jnp.int32), arr)
         state = eqxi.while_loop(
             max_steps=config.time_steps_total,
-            cond_fun=cond,
+            cond_fun=lambda s: config.time_steps_total > s[0],
             body_fun=partial(
                 forward,
                 config=config,
