### 2025-12-15 1dc6915
precommit format, correct modes.py input validation
