+++ src/fdtdx/core/grid.py
@@ -23,6 +23,7 @@ def calculate_spatial_offsets_yee() -> tuple[jax.Array, jax.Array]:
     )
     return offset_E, offset_H
 
+
 def calculate_time_offset_yee(
     center: jax.Array,
     wave_vector: jax.Array,
@@ -85,42 +86,46 @@ def calculate_time_offset_yee(
         # adjust speed for material and calculate time offset
         # inv_eps_eff = |p_x|^2 * inv_eps_x + |p_y|^2 * inv_eps_y + |p_z|^2 * inv_eps_z
         if inv_permittivities.ndim == 4:
-            #Remove when anisotropic case is verified
+            # Remove when anisotropic case is verified
             if inv_permittivities.shape[0] == 3:
-                is_isotropic = jnp.allclose(inv_permittivities[0], inv_permittivities[1]) & jnp.allclose(inv_permittivities[1], inv_permittivities[2])
+                is_isotropic = jnp.allclose(inv_permittivities[0], inv_permittivities[1]) & jnp.allclose(
+                    inv_permittivities[1], inv_permittivities[2]
+                )
+
                 def _raise_if_anisotropic(is_iso):
                     if not is_iso:
-                        raise NotImplementedError("Gaussian or planewave sources within anisotropic materials are not supported yet.")
+                        raise NotImplementedError(
+                            "Gaussian or planewave sources within anisotropic materials are not supported yet."
+                        )
+
                 jax.debug.callback(_raise_if_anisotropic, is_isotropic)
 
             if e_polarization is None:
-                raise ValueError(
-                    "e_polarization is required for anisotropic materials (4D permittivity array)"
-                )
+                raise ValueError("e_polarization is required for anisotropic materials (4D permittivity array)")
             e_pol_squared = e_polarization**2
-            inv_perm_eff = jnp.sum(
-                e_pol_squared[:, None, None, None] * inv_permittivities, axis=0
-            )
+            inv_perm_eff = jnp.sum(e_pol_squared[:, None, None, None] * inv_permittivities, axis=0)
         else:
             inv_perm_eff = inv_permittivities
 
         if isinstance(inv_permeabilities, jax.Array) and inv_permeabilities.ndim == 4:
-            #Remove when anisotropic case is verified
+            # Remove when anisotropic case is verified
             if inv_permeabilities.shape[0] == 3:
-                is_isotropic = jnp.allclose(inv_permeabilities[0], inv_permeabilities[1]) & jnp.allclose(inv_permeabilities[1], inv_permeabilities[2])
+                is_isotropic = jnp.allclose(inv_permeabilities[0], inv_permeabilities[1]) & jnp.allclose(
+                    inv_permeabilities[1], inv_permeabilities[2]
+                )
+
                 def _raise_if_anisotropic(is_iso):
                     if not is_iso:
-                        raise NotImplementedError("Gaussian or planewave sources within anisotropic materials are not supported yet.")
+                        raise NotImplementedError(
+                            "Gaussian or planewave sources within anisotropic materials are not supported yet."
+                        )
+
                 jax.debug.callback(_raise_if_anisotropic, is_isotropic)
 
             if h_polarization is None:
-                raise ValueError(
-                    "h_polarization is required for anisotropic materials (4D permeability array)"
-                )
+                raise ValueError("h_polarization is required for anisotropic materials (4D permeability array)")
             h_pol_squared = h_polarization**2
-            inv_perm_eff_perm = jnp.sum(
-                h_pol_squared[:, None, None, None] * inv_permeabilities, axis=0
-            )
+            inv_perm_eff_perm = jnp.sum(h_pol_squared[:, None, None, None] * inv_permeabilities, axis=0)
         else:
             inv_perm_eff_perm = inv_permeabilities
 
