+++ src/fdtdx/core/physics/modes.py
@@ -124,10 +124,18 @@ def compute_mode(
             Tuple of E, H field and the effective index as complex-valued jax arrays.
     """
     # Input validation
-    if not (inv_permittivities.ndim == 4 and (inv_permittivities.shape[0] == 1 or inv_permittivities.shape[0] == 3)):
+    if (
+        not (inv_permittivities.ndim == 4 and (inv_permittivities.shape[0] == 1 or inv_permittivities.shape[0] == 3))
+        or sum(dim == 1 for dim in inv_permittivities.shape[1:]) != 1
+    ):
         raise Exception(f"Invalid shape of inv_permittivities: {inv_permittivities.shape}")
     if isinstance(inv_permeabilities, jax.Array) and inv_permeabilities.ndim > 0:
-        if not (inv_permeabilities.ndim == 4 and (inv_permeabilities.shape[0] == 1 or inv_permeabilities.shape[0] == 3)):
+        if (
+            not (
+                inv_permeabilities.ndim == 4 and (inv_permeabilities.shape[0] == 1 or inv_permeabilities.shape[0] == 3)
+            )
+            or sum(dim == 1 for dim in inv_permeabilities.shape[1:]) != 1
+        ):
             raise Exception(f"Invalid shape of inv_permeabilities: {inv_permeabilities.shape}")
         # raise Exception("Mode solver currently does not support metallic materials")
 
@@ -262,7 +270,7 @@ def tidy3d_mode_computation_wrapper(
         track_freq="central",
         group_index_step=False,
     )
-    idx = [0,1,2] if permittivity_cross_section.shape[0] == 3 else [0,0,0]
+    idx = [0, 1, 2] if permittivity_cross_section.shape[0] == 3 else [0, 0, 0]
     od = np.zeros_like(permittivity_cross_section[1, :, :])
     eps_cross = [
         permittivity_cross_section[idx[0], :, :],
@@ -290,7 +298,7 @@ def tidy3d_mode_computation_wrapper(
                 permeability_cross_section,
             ]
         elif isinstance(permeability_cross_section, np.ndarray):
-            idx = [0,1,2] if permeability_cross_section.shape[0] == 3 else [0,0,0]
+            idx = [0, 1, 2] if permeability_cross_section.shape[0] == 3 else [0, 0, 0]
             mu_cross = [
                 permeability_cross_section[idx[0], :, :],
                 od,
