### 2025-12-17 91b2cd9
Add rotation for anisotropic materials
