### 2025-12-17 92c24c8
fix minor issues in phasor.py
