### 2025-12-18 9cf3d37
Implemented anisotropic materials (#204)
* added non isotropic permetivity, permeability, and conductivity

* fixed core fdtdx components to accept anisotropic materials, some cleanup

* fix init for pylance type errors that might confuse users

* bug fixes for permittivity array management, added anisotropic example

* bug fix revealed by width_sweep_analysis

* run precommit checks

* PR review correction

* precommit format, correct modes.py input validation

* Add rotation for anisotropic materials
