+++ tests/core/physics/test_modes.py
@@ -182,7 +182,7 @@ class TestComputeMode:
 
         # Test inputs
         frequency = 2e14  # 200 THz
-        inv_permittivities = jnp.ones((1, 5, 5)) * 0.25  # eps = 4.0
+        inv_permittivities = jnp.ones((3, 1, 5, 5)) * 0.25  # eps = 4.0
         inv_permeabilities = 1.0  # mu = 1.0
         resolution = 1e-8  # 10 nm
 
@@ -205,7 +205,7 @@ class TestComputeMode:
         """Test that invalid permittivities shape raises exception."""
         frequency = 2e14
         # 3D array but not squeezable to 2D
-        inv_permittivities = jnp.ones((2, 2, 2))
+        inv_permittivities = jnp.ones((3, 2, 2, 2))
         inv_permeabilities = 1.0
         resolution = 1e-8
 
@@ -233,15 +233,255 @@ class TestComputeMode:
                 jnp.ones((3, 3, 1, 4), dtype=jnp.complex64),
             )
 
-            # Test propagation along axis 0 (shape: 1, 3, 4)
-            inv_permittivities = jnp.ones((1, 3, 4))
+            # Test propagation along axis 0 (shape: 3, 1, 3, 4)
+            inv_permittivities = jnp.ones((3, 1, 3, 4))
             E, H, neff = compute_mode(2e14, inv_permittivities, 1.0, 1e-8, "+")
             assert E.shape == (3, 3, 1, 4)
 
-            # Test propagation along axis 1 (shape: 3, 1, 4)
-            inv_permittivities = jnp.ones((3, 1, 4))
+            # Test propagation along axis 1 (shape: 3, 3, 1, 4)
+            inv_permittivities = jnp.ones((3, 3, 1, 4))
             E, H, neff = compute_mode(2e14, inv_permittivities, 1.0, 1e-8, "+")
             assert E.shape == (3, 3, 1, 4)  # Note: axis handling in the function
+
+
+class TestAnisotropicModeComputation:
+    """Test anisotropic material handling in compute_mode."""
+
+    @patch("fdtdx.core.physics.modes.tidy3d_mode_computation_wrapper")
+    @patch("fdtdx.core.physics.modes.normalize_by_poynting_flux")
+    def test_anisotropic_permittivity_rotation_propagation_axis_0(self, mock_normalize, mock_tidy3d_wrapper):
+        """Test that anisotropic permittivity is rotated correctly for propagation along axis 0.
+
+        When propagation is along x (axis 0), tidy3d (x,y,z) maps to physical (y,z,x).
+        So physical permittivity (eps_x, eps_y, eps_z) should become tidy3d (eps_y, eps_z, eps_x).
+        """
+        mock_mode = ModeTupleType(
+            neff=1.5 + 0.1j,
+            Ex=np.ones((5, 6), dtype=np.complex64),
+            Ey=np.ones((5, 6), dtype=np.complex64),
+            Ez=np.ones((5, 6), dtype=np.complex64),
+            Hx=np.ones((5, 6), dtype=np.complex64),
+            Hy=np.ones((5, 6), dtype=np.complex64),
+            Hz=np.ones((5, 6), dtype=np.complex64),
+        )
+        mock_tidy3d_wrapper.return_value = [mock_mode]
+        mock_normalize.return_value = (
+            jnp.ones((3, 1, 5, 6), dtype=jnp.complex64),
+            jnp.ones((3, 1, 5, 6), dtype=jnp.complex64),
+        )
+
+        # Anisotropic permittivity with distinct values: eps_x=2, eps_y=3, eps_z=4
+        # Shape: (3, 1, 5, 6) - propagation along axis 0
[…]
