### 2025-12-19 f2725d7
Fix where overloads to preserve unit dims and assert on mismatch
