### 2025-12-19 5ce9164
add horizontal, vertical, point symmetry. Add 3D versions.
