+++ src/fdtdx/objects/device/parameters/symmetries.py
@@ -8,13 +8,15 @@ from fdtdx.objects.device.parameters.transform import SameShapeTypeParameterTran
 @autoinit
 class DiagonalSymmetry2D(SameShapeTypeParameterTransform):
     """
-    Enforce symmetries by effectively havling the parameter space. The symmetry is transposing by rotating the image
-    and taking the mean of original and transpose.
+    Enforce diagonal symmetry by effectively halving the parameter space.
+    The symmetry is achieved by transposing the image and taking the mean
+    of the original and transpose.
 
+    This creates a design that is symmetric across one of the two diagonals.
     """
 
-    #: if true, the symmetry axes is from (x_min, y_min) to (x_max, y_max).
-    #: If false, the other diagonal is used.
+    #: If true, the symmetry axis is from (x_min, y_min) to (x_max, y_max).
+    #: If false, the other diagonal (from (x_min, y_max) to (x_max, y_min)) is used.
     min_min_to_max_max: bool = frozen_field()
 
     _all_arrays_2d: bool = frozen_private_field(default=True)
@@ -40,4 +42,262 @@ class DiagonalSymmetry2D(SameShapeTypeParameterTransform):
 
             # expand dims again
             result[k] = jnp.expand_dims(cur_mean, vertical_axis)
+        return result
+
+
+@autoinit
+class HorizontalSymmetry2D(SameShapeTypeParameterTransform):
+    """
+    Enforce horizontal (x-axis) mirror symmetry.
+
+    This creates a design that is symmetric across a vertical line through
+    the center, i.e., the left half mirrors the right half.
+    The symmetry is enforced by averaging the array with its horizontally
+    flipped version.
+    """
+
+    _all_arrays_2d: bool = frozen_private_field(default=True)
+
+    def __call__(
+        self,
+        params: dict[str, jax.Array],
+        **kwargs,
+    ) -> dict[str, jax.Array]:
+        del kwargs
+        result = {}
+        for k, v in params.items():
+            # convert to 2d
+            vertical_axis = v.shape.index(1)
+            v_2d = v.squeeze(vertical_axis)
+
+            # enforce symmetry: flip along x-axis (axis 0)
+            flipped = v_2d[::-1, :]
+            cur_mean = (v_2d + flipped) / 2
+
+            # expand dims again
+            result[k] = jnp.expand_dims(cur_mean, vertical_axis)
+        return result
+
+
+@autoinit
+class VerticalSymmetry2D(SameShapeTypeParameterTransform):
+    """
+    Enforce vertical (y-axis) mirror symmetry.
+
+    This creates a design that is symmetric across a horizontal line through
+    the center, i.e., the top half mirrors the bottom half.
+    The symmetry is enforced by averaging the array with its vertically
+    flipped version.
+    """
+
+    _all_arrays_2d: bool = frozen_private_field(default=True)
+
+    def __call__(
+        self,
+        params: dict[str, jax.Array],
+        **kwargs,
+    ) -> dict[str, jax.Array]:
+        del kwargs
+        result = {}
+        for k, v in params.items():
+            # convert to 2d
+            vertical_axis = v.shape.index(1)
+            v_2d = v.squeeze(vertical_axis)
+
+            # enforce symmetry: flip along y-axis (axis 1)
+            flipped = v_2d[:, ::-1]
+            cur_mean = (v_2d + flipped) / 2
+
+            # expand dims again
+            result[k] = jnp.expand_dims(cur_mean, vertical_axis)
+        return result
+
+
+@autoinit
+class PointSymmetry2D(SameShapeTypeParameterTransform):
+    """
+    Enforce 180-degree rotational (point) symmetry.
+
[…]
