+++ tests/test_symmetries.py
@@ -1,0 +1,455 @@
+"""Tests for symmetry parameter transforms."""
+
+import jax.numpy as jnp
+import pytest
+
+from fdtdx import (
+    DiagonalSymmetry2D,
+    DiagonalSymmetry3D,
+    HorizontalSymmetry2D,
+    HorizontalSymmetry3D,
+    PointSymmetry2D,
+    PointSymmetry3D,
+    VerticalSymmetry2D,
+    VerticalSymmetry3D,
+)
+
+
+# =============================================================================
+# 2D Symmetry Tests
+# =============================================================================
+
+
+class TestHorizontalSymmetry2D:
+    """Tests for HorizontalSymmetry2D (x-axis mirror)."""
+
+    def test_symmetry_property(self):
+        """Test that output satisfies horizontal symmetry."""
+        # Create 2D array with singleton middle dimension (required by 2D transforms)
+        test_arr = jnp.array([
+            [[1, 2, 3, 4]],
+            [[5, 6, 7, 8]],
+            [[9, 10, 11, 12]],
+            [[13, 14, 15, 16]],
+        ], dtype=jnp.float32)
+
+        sym = HorizontalSymmetry2D()
+        result = sym({"test": test_arr})["test"]
+        squeezed = result.squeeze(1)
+
+        # Check symmetry: arr == arr[::-1, :]
+        assert jnp.allclose(squeezed, squeezed[::-1, :])
+
+    def test_point_reflection(self):
+        """Test that a single point is correctly mirrored."""
+        test_arr = jnp.zeros((5, 1, 5), dtype=jnp.float32)
+        test_arr = test_arr.at[1, 0, 1].set(1.0)
+
+        sym = HorizontalSymmetry2D()
+        result = sym({"test": test_arr})["test"]
+
+        # Point at [1, 0, 1] should mirror to [3, 0, 1]
+        assert result[1, 0, 1] == 0.5
+        assert result[3, 0, 1] == 0.5
+
+    def test_shape_preserved(self):
+        """Test that output shape matches input shape."""
+        test_arr = jnp.ones((4, 1, 6), dtype=jnp.float32)
+        sym = HorizontalSymmetry2D()
+        result = sym({"test": test_arr})["test"]
+        assert result.shape == test_arr.shape
+
+    def test_multiple_arrays(self):
+        """Test that transform works on multiple arrays in dict."""
+        params = {
+            "a": jnp.ones((4, 1, 4), dtype=jnp.float32),
+            "b": jnp.ones((6, 1, 6), dtype=jnp.float32),
+        }
+        sym = HorizontalSymmetry2D()
+        result = sym(params)
+        assert "a" in result and "b" in result
+        assert result["a"].shape == (4, 1, 4)
+        assert result["b"].shape == (6, 1, 6)
+
+
+class TestVerticalSymmetry2D:
+    """Tests for VerticalSymmetry2D (y-axis mirror)."""
+
+    def test_symmetry_property(self):
+        """Test that output satisfies vertical symmetry."""
+        test_arr = jnp.array([
+            [[1, 2, 3, 4]],
+            [[5, 6, 7, 8]],
+            [[9, 10, 11, 12]],
+            [[13, 14, 15, 16]],
+        ], dtype=jnp.float32)
+
+        sym = VerticalSymmetry2D()
+        result = sym({"test": test_arr})["test"]
+        squeezed = result.squeeze(1)
+
+        # Check symmetry: arr == arr[:, ::-1]
+        assert jnp.allclose(squeezed, squeezed[:, ::-1])
+
+    def test_point_reflection(self):
[…]
