### 2025-12-20 a393cbf
Implement more geometry symmetries and add script to code.zip (#213)
* add horizontal, vertical, point symmetry. Add 3D versions.

* add running script to code.zip

* format, fix init
