### 2025-12-20 3aa7812
improved unitful where (#210)
* improved unit where

* fine-grained unit where

* Fix where overloads to preserve unit dims and assert on mismatch
