+++ tests/functional/test_numpy.py
@@ -6,8 +6,10 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 
+import fdtdx.units.unitful as unitful
 from fdtdx.core.fraction import Fraction
-from fdtdx.functional.numpy import roll, sqrt, transpose
+from fdtdx.core.jax.utils import is_currently_compiling
+from fdtdx.functional.numpy import roll, sqrt, transpose, where
 from fdtdx.units.typing import SI
 from fdtdx.units.unitful import EMPTY_UNIT, Unit, Unitful
 
@@ -1113,3 +1115,184 @@ def test_arctanh_unitful_with_jax_array():
     assert isinstance(result.val, jax.Array)
     expected_vals = jnp.array([math.atanh(0.0), math.atanh(0.5)])
     assert jnp.allclose(result.val, expected_vals, atol=1e-7)
+
+
+def test_where_unitful_numpy_backend():
+    """Test where with three Unitfuls using NumPy arrays"""
+    m = Unit(scale=0, dim={SI.m: 1})
+
+    c = Unitful(val=np.array([True, False, True]), unit=EMPTY_UNIT)
+    x = Unitful(val=np.array([1.0, 2.0, 3.0]), unit=m)
+    y = Unitful(val=np.array([10.0, 20.0, 30.0]), unit=m)
+
+    result = where(c, x, y)
+
+    assert isinstance(result, Unitful)
+    assert isinstance(result.val, np.ndarray)
+
+    expected = np.where(c.value(), x.value(), y.value())
+    assert np.allclose(result.value(), expected)
+
+
+def test_where_unitful_jax_backend():
+    """Test where with three Unitfuls using JAX arrays"""
+    m = Unit(scale=0, dim={SI.m: 1})
+    mm = Unit(scale=-3, dim={SI.m: 1})
+
+    c = Unitful(val=jnp.array([True, False, True]), unit=EMPTY_UNIT)
+    x = Unitful(val=jnp.array([1.0, 2.0, 3.0]), unit=m)
+    y = Unitful(val=jnp.array([10.0, 20.0, 30.0]), unit=mm)
+
+    result = jnp.where(c, x, y)  # type: ignore
+
+    assert isinstance(result, Unitful)
+    assert isinstance(result.val, jax.Array)
+
+    expected = np.where(np.array(c.value()), np.array(x.value()), np.array(y.value()))
+
+    assert np.allclose(np.asarray(result.value()), expected)
+
+
+def test_where_numpy_and_unitful():
+    """Test where with NumPy condition/x and Unitful y"""
+    condition = np.array([True, False, True])
+    x = np.array([1.0, 2.0, 3.0])
+    y = Unitful(val=np.array([10.0, 20.0, 30.0]), unit=EMPTY_UNIT)
+
+    result = where(condition, x, y)
+
+    expected = np.where(condition, x, y.value())
+
+    assert isinstance(result, Unitful)
+    assert np.allclose(np.asarray(result.value()), expected)
+
+
+def test_where_jax_and_unitful():
+    """Test where with JAX condition/x and Unitful y"""
+    condition = jnp.array([True, False, True])
+    x = jnp.array([1.0, 2.0, 3.0])
+    y = Unitful(val=np.array([10.0, 20.0, 30.0]), unit=EMPTY_UNIT)
+
+    result = where(condition, x, y)
+
+    assert isinstance(result, Unitful)
+    assert isinstance(result.val, jax.Array)
+
+    expected = np.where(np.array(condition), np.array(x), y.value())
+    assert np.allclose(np.asarray(result.value()), expected)
+
+
+def test_where_pure_numpy():
+    """Test where with pure numpy arrays"""
[…]
