### 2025-12-23 #8912 @ymahlau
Hi @fruzsinaagocs, I think all the necessary steps for finalizing the publication are done. Do you think we can finalize the publication this year or early next year? We are planning to submit a paper which uses FDTDX in January and would like to have a good reference for the framework in the paper :)

A merry christmas to y'all and a happy new year if we do not hear from each other until then!
