### 2025-12-31 5f22968
final paper updates
