+++ paper.bib
@@ -67,13 +67,6 @@ howpublished = {https://github.com/flaport/fdtd},
 year = {2024}
 }
 
-@misc{luminescent,
-author = {Paul Shen},
-title = {Luminescent AI},
-howpublished = {https://github.com/paulxshen/Luminescent.jl},
-year = {2024},
-}
-
 @misc{khronos,
 author = {Alec Hammond},
 title = {Khronos},
@@ -89,15 +82,20 @@ keywords = {EM FDTD textbook},
 publisher = {Artech House},
 title = {Computational electrodynamics: the finite-difference time-domain method},
 year = {2005},
-isbn = {9781580538329}
+isbn = {9781580538329},
+doi = {10.1016/B978-012170960-0/50046-3}
 }
 
 
 @article{mahlau2025multi,
-title={Multi-Agent Reinforcement Learning for Inverse Design in Photonic Integrated Circuits},
-author={Mahlau, Yannik and Schier, Maximilian and Reinders, Christoph and Schubert, Frederik and B{\"u}gling, Marco and Rosenhahn, Bodo},
-journal={arXiv preprint arXiv:2506.18627},
-year={2025}
+    title={Multi-Agent Reinforcement Learning for Inverse Design in Photonic Integrated Circuits},
+    author={Mahlau, Yannik and Schier, Maximilian and Reinders, Christoph and Schubert, Frederik and B{\"{u}}gling, Marco and Rosenhahn, Bodo},
+    journal={Reinforcement Learning Journal},
+    volume={6},
+    pages={1794--1815},
+    year={2025},
+    url={https://rlj.cs.umass.edu/2025/papers/Paper192.html},
+    doi = {10.48550/arXiv.2506.18627}
 }
 
 @inproceedings{mahlau2025flexible,
@@ -161,5 +159,13 @@ number = {5},
 pages = {334-339},
 year = {2000},
 doi={10.1002/1098-2760(20001205)27:5<334::AID-MOP14>3.0.CO;2-A}
+}
+
+@book{pontryagin2018mathematical,
+  title={Mathematical theory of optimal processes},
+  author={Pontryagin, Lev Semenovich},
+  year={2018},
+  doi={10.1201/9780203749319},
+  publisher={Routledge}
 }
 
