+++ paper.md
@@ -32,23 +32,23 @@ The ability to precisely manipulate the propagation of light is an increasingly 
 Aside from analytical solutions for simple geometries, the majority of practical problems require robust numerical methods such as the finite-difference time-domain (FDTD) method for discretizing Maxwell's equations in both space and time.
 We present FDTDX, an efficient implementation of the FDTD method with GPU acceleration through the JAX framework.
 It provides a simple user interface for specifying a simulation scene as well as a suite of tools for inverse design.
-Using the time reversibility of maxwell's equations, gradients for optimizing the geometry of a design can be computated efficiently within FDTDX.
+Using the time reversibility of Maxwell's equations, gradients for optimizing the geometry of a design can be computated efficiently within FDTDX.
 
 # Statement of Need
 
-FDTDX implements the FDTD algorithm, which aims to simulate Maxwell's equations $\frac{\partial H}{\partial t} = - \frac{1}{\mu} \nabla \times E$ and $\frac{\partial E}{\partial t} = \frac{1}{\epsilon} \nabla \times H$, where $E$ and $H$ are the electric and magnetic field components.
+FDTDX implements the FDTD algorithm, which aims to simulate Maxwell's equations $\frac{\partial H}{\partial t} = - \frac{1}{\mu} \nabla \times E$ and $\frac{\partial E}{\partial t} = \frac{1}{\epsilon} \nabla \times H$, where $E$ and $H$ are the electric and magnetic fields.
 This algorithm has been used in a number of research applications, for example in the field of photonic integrated circuits [@augenstein2020inverse], optical computing [@mahlau2025multi] or quantum computing [@larsen2025integrated].
 
 The FDTD algorithm has been well known for a long time and a number of open-source packages already implement it.
-However, most previous packages implement the algorithm only for CPU, which misses out on massive speedups through GPU acceleration.
+However, most previous packages implement the algorithm only for CPU, thereby missing out on significant speedups through GPU acceleration.
 Additionally, the implementation of the FDTD algorithm in JAX allows for automatic differentiation using a specialized algorithm based on the time reversibility of Maxwell's equations [@schubertmahlau2025quantized].
-In contrast to the adjoint method, our custom gradient algorithm can calculate a gradient in the time-domain without the need to save electric and magnetic field after every time step.
+In contrast to the adjoint method [@pontryagin2018mathematical], our custom gradient algorithm can calculate a gradient in the time-domain without the need to save the electric and magnetic field after every time step.
 This enables memory efficient inverse design, i.e. topological optimization of optical components using gradient descent.
 The simulation capabilities of FDTDX are useful for physicists from all kinds of backgrounds, while the inverse design capabilities are targeted more towards computational scientists in particular.
 
 A non-exhaustive list of FDTD implementations includes the popular Meep [@meep], which was developed almost 20 years ago for execution on CPU and is still widely used today.
-Other frameworks for only CPU include OpenEMS [@openEMS], fdtd [@fdtd_laporte] and Ceviche [@ceviche].
-Existing open-source packages that support execution on GPU are Khronos [@khronos] and FDTD-Z [@fdtdz], but both packages are not maintained.
+Other frameworks for CPU only include OpenEMS [@openEMS], fdtd [@fdtd_laporte] and Ceviche [@ceviche].
+Existing open-source packages that support execution on GPU are Khronos [@khronos] and FDTD-Z [@fdtdz], but neither of these packages are maintained.
 Additionally, various commercial implementations of FDTD exist.
 Notably, Tidy3D [@tidy3d] is an extremely fast commercial software due to its GPU acceleration.
[…]
