### 2025-12-31 #8912 @ymahlau
@fruzsinaagocs thanks for your review! I have incorporated all of the requested changes.

Regarding the  "Multi-agent reinforcement learning..." paper, it has already been published in the Reinforcement Learning Journal, which I have updated as a reference in the paper. Unfortunately, the RLJ does not seem to have a DOI. Therefore, I have added the DOI to the Arxiv-preprint (which is identical). Do you think this is good or should I omit the DOI since it is technically not the same reference?
