### 2025-12-31 #8912 @fruzsinaagocs
> [@fruzsinaagocs](https://github.com/fruzsinaagocs) thanks for your review! I have incorporated all of the requested changes.
>
> Regarding the "Multi-agent reinforcement learning..." paper, it has already been published in the Reinforcement Learning Journal, which I have updated as a reference in the paper. Unfortunately, the RLJ does not seem to have a DOI. Therefore, I have added the DOI to the Arxiv-preprint (which is identical). Do you think this is good or should I omit the DOI since it is technically not the same reference?

I see that this journal assigns a DOI for basically the entire volume, which includes papers published that year. There is no DOI for the 2025 volume yet, so I think having the arxiv DOI is OK as long as the version matches the published one. If not, can you please update the arxiv version so it does?
