### 2025-12-31 #8912 @ymahlau
> I see that this journal assigns a DOI for basically the entire volume, which includes papers published that year. There is no DOI for the 2025 volume yet, so I think having the arxiv DOI is OK as long as the version matches the published one. If not, can you please update the arxiv version so it does?

Ah I did not see that there was a DOI for the previous year. The journal and arxiv version of the paper are already identical.
