### 2026-01-06 9f45164
updated PhasorDetector for multiple frequencies
