### 2026-01-06 #8912 @ymahlau
> Just one more thing: the zenodo archive's description has a small typo "...github reposiory here".

Thanks for catching this typo! I have fixed it now.
