### 2026-01-08 05efd6b
refactored GaussianPulseProfile by adding fdtdx.WaveCharacter
