+++ src/fdtdx/objects/sources/profile.py
@@ -5,6 +5,7 @@ import jax
 import jax.numpy as jnp
 
 from fdtdx.core.jax.pytrees import TreeClass, autoinit, frozen_field
+from fdtdx.core.wavelength import WaveCharacter
 
 
 @autoinit
@@ -59,13 +60,18 @@ class SingleFrequencyProfile(TemporalProfile):
 
 @autoinit
 class GaussianPulseProfile(TemporalProfile):
-    """Gaussian pulse temporal profile with carrier wave."""
+    """Gaussian pulse temporal profile with carrier wave.
 
-    #: Width of the carrier wave
-    spectral_width: float = frozen_field()  # Width of the Gaussian envelope in frequency domain
+    The pulse envelope is characterized by its spectral width, which determines
+    the temporal width of the pulse. The carrier wave can be specified using
+    any of the WaveCharacter parameters (wavelength, frequency, or period).
+    """
 
-    #: Center frequency of the carrier wave
-    center_frequency: float = frozen_field()  # Center frequency of the pulse
+    #: Spectral width of the Gaussian envelope (can specify via wavelength, frequency, or period)
+    spectral_width: WaveCharacter = frozen_field()
+
+    #: Center frequency/wavelength of the carrier wave
+    center_wave: WaveCharacter = frozen_field()
 
     def get_amplitude(
         self,
@@ -74,15 +80,20 @@ class GaussianPulseProfile(TemporalProfile):
         phase_shift: float = 0.0,
     ) -> jax.Array:
         del period
+
+        # Get frequency values from WaveCharacter objects
+        spectral_width_hz = self.spectral_width.get_frequency()
+        center_frequency_hz = self.center_wave.get_frequency()
+
         # Calculate envelope parameters
-        sigma_t = 1.0 / (2 * jnp.pi * self.spectral_width)
+        sigma_t = 1.0 / (2 * jnp.pi * spectral_width_hz)
         t0 = 6 * sigma_t  # Offset peak to avoid discontinuity at t=0
 
         # Gaussian envelope
         envelope = jnp.exp(-((time - t0) ** 2) / (2 * sigma_t**2))
 
-        # Carrier wave
-        carrier_phase = 2 * jnp.pi * self.center_frequency * time + phase_shift
+        # Carrier wave (including phase shift from center_wave)
+        carrier_phase = 2 * jnp.pi * center_frequency_hz * time + phase_shift + self.center_wave.phase_shift
         carrier = jnp.real(jnp.exp(-1j * carrier_phase))
 
         return envelope * carrier
