### 2026-01-09 ad01b4c
refactored Colors by adding class and xkcd color format
