+++ src/fdtdx/core/plotting/colors.py
@@ -1,40 +1,414 @@
-"""Color constants for visualization and plotting.
+"""Color module for fdtdx framework.
 
-This module provides a collection of predefined RGB color tuples normalized to the range [0,1].
-Colors are organized into categories: primary/bright colors, grayscale, and earth tones.
-These colors are designed to provide a consistent and visually appealing palette for plotting
-and visualization tasks throughout the FDTDX framework.
+This module provides a Color class for representing and converting between different
+color formats, along with a comprehensive collection of named colors from the XKCD
+color survey. The Color class is a subclass of fdtdx.TreeClass and supports:
 
-Each color is defined as a tuple of (red, green, blue) values normalized to [0,1].
-The normalization is done by dividing 8-bit RGB values (0-255) by 255.
+- Multiple input formats: RGB tuples (normalized or 8-bit), hex strings, named colors
+- Multiple output formats: normalized RGB, 8-bit RGB, hex, matplotlib-compatible
+- Type checking and validation
+- Integration with the fdtdx framework
 
-Categories:
-    - Bright and primary colors: Vibrant colors for emphasis and contrast
-    - Grayscale colors: Various shades of gray for backgrounds and subtle elements
-    - Earth tones: Natural, warm colors for material representations
+The module includes all XKCD colors as predefined constants for consistent visualization.
 
 Example:
-    >>> import matplotlib.pyplot as plt
-    >>> plt.plot([0, 1], [0, 1], color=GREEN)  # Plot a line in vibrant green
-    >>> plt.fill_between([0, 1], [0, 1], color=LIGHT_BLUE, alpha=0.3)  # Fill with pale blue
+    >>> from fdtdx import colors
+    >>> # Create colors from different formats
+    >>> c1 = colors.Color.from_rgb(255, 0, 0)  # 8-bit RGB
+    >>> c2 = colors.Color.from_hex("#00FF00")
+    >>> c3 = colors.XKCD_MINT_GREEN
+    >>>
+    >>> # Convert to different formats
+    >>> c1.to_hex()  # "#FF0000"
+    >>> c1.to_rgb_normalized()  # (1.0, 0.0, 0.0)
+    >>> c1.to_rgb_255()  # (255, 0, 0)
 """
 
+from typing import Tuple
+
+from fdtdx.core.jax.pytrees import TreeClass, autoinit
+
+
+@autoinit
+class Color(TreeClass):
+    """Color representation with multiple format support.
+
+    This class represents a color and provides methods to convert between
+    different color formats. Internally, colors are stored as normalized
+    RGB values in the range [0, 1].
+
+    Attributes:
[…]
