### 2026-01-09 8c57454
updated the sign function
