+++ src/fdtdx/units/unitful.py
@@ -1611,25 +1611,21 @@ def argmin(x, *args, **kwargs):  # type: ignore
 
 
 ## sign #######################################
-from typing import overload as _typing_overload  # ensure name not shadowed
 
 @overload
 def sign(x: Unitful) -> Unitful:
     # Drop units and return integer {-1,0,1}
-    # Handle NumPy-backed vs JAX-backed values separately, just like other ops.
     if isinstance(x.val, np.ndarray | np.number):
         # NumPy path
         y = np.sign(x.val)
-        # Ensure integer dtype (ndarray or scalar)
         if isinstance(y, np.ndarray):
             y_int: np.ndarray | np.number = y.astype(np.int_)
         else:
-            # numpy scalar -> numpy integer scalar
             y_int = np.int_(y)
         return Unitful(val=y_int, unit=EMPTY_UNIT, static_arr=None)
     else:
         # JAX path (includes tracers)
-        y = jax.lax.sign(x.val)  # avoids recursion and works under jit
+        y = jax.lax.sign(x.val)
         y_int = jax.lax.convert_element_type(y, jnp.int_)  # type: ignore
         if not isinstance(y_int, jax.Array):
             raise Exception(f"This is an internal error: sign produced {type(y_int)}")
@@ -1648,31 +1644,27 @@ def sign(x: Unitful) -> Unitful:
 
 @overload
 def sign(x: jax.Array) -> jax.Array:
-    # JAX original op via lax, then cast to integer dtype
-    y = jax.lax.sign(x)
-    return jax.lax.convert_element_type(y, jnp.int_)  # type: ignore
+    # JAX original op via lax, preserve output type
+    return jax.lax.sign(x)
 
 
 @overload
 def sign(x: np.ndarray) -> np.ndarray:
-    # NumPy sign then cast to numpy integer
-    y = np.sign(x)
-    return y.astype(np.int_)
+    # NumPy sign, preserve output type
+    return np.sign(x)
 
 
 @overload
 def sign(x: int) -> int:
-    # python integer → python integer in {-1, 0, 1}
     return -1 if x < 0 else (1 if x > 0 else 0)
+
+
+@overload
+def sign(x: float) -> float:
+    return -1.0 if x < 0 else (1.0 if x > 0 else 0.0)
 
 
 @dispatch
 def sign(x):  # type: ignore
     del x
     raise NotImplementedError()
-
-
-# --- make jnp.sign(...) use these overloads, preserving original ---
-if not hasattr(jnp, "_orig_sign"):
-    jnp._orig_sign = jnp.sign  # type: ignore
-jnp.sign = sign  # type: ignore
