### 2026-01-09 7e4c516
first commit for full anisotropy
