+++ src/fdtdx/core/physics/modes.py
@@ -1,6 +1,6 @@
 from collections import namedtuple
 from types import SimpleNamespace
-from typing import List, Literal
+from typing import List, Literal, cast
 
 import jax
 import jax.numpy as jnp
@@ -8,6 +8,7 @@ import numpy as np
 import tidy3d
 from tidy3d.components.mode.solver import compute_modes as _compute_modes
 
+from fdtdx.core.misc import expand_to_3x3
 from fdtdx.core.physics.metrics import normalize_by_poynting_flux
 
 ModeTupleType = namedtuple("Mode", ["neff", "Ex", "Ey", "Ez", "Hx", "Hy", "Hz"])
@@ -125,14 +126,14 @@ def compute_mode(
     """
     # Input validation
     if (
-        not (inv_permittivities.ndim == 4 and (inv_permittivities.shape[0] == 1 or inv_permittivities.shape[0] == 3))
+        not (inv_permittivities.ndim == 4 and (inv_permittivities.shape[0] == 1 or inv_permittivities.shape[0] == 3 or inv_permittivities.shape[0] == 9))
         or sum(dim == 1 for dim in inv_permittivities.shape[1:]) != 1
     ):
         raise Exception(f"Invalid shape of inv_permittivities: {inv_permittivities.shape}")
     if isinstance(inv_permeabilities, jax.Array) and inv_permeabilities.ndim > 0:
         if (
             not (
-                inv_permeabilities.ndim == 4 and (inv_permeabilities.shape[0] == 1 or inv_permeabilities.shape[0] == 3)
+                inv_permeabilities.ndim == 4 and (inv_permeabilities.shape[0] == 1 or inv_permeabilities.shape[0] == 3 or inv_permeabilities.shape[0] == 9)
             )
             or sum(dim == 1 for dim in inv_permeabilities.shape[1:]) != 1
         ):
@@ -176,10 +177,17 @@ def compute_mode(
         return mode_E, mode_H, neff
 
     # compute input to tidy3d Mode solver
-    permittivities = 1 / inv_permittivities
+    if isinstance(inv_permittivities, jax.Array) and inv_permittivities.ndim > 0 and inv_permittivities.shape[0] == 9:
+        eps = cast(jax.Array, expand_to_3x3(inv_permittivities))
+        # Invert the 3x3 matrix
+        perm = (2, 3, 4, 0, 1)      # (3, 3, nx, ny, nz) -> (nx, ny, nz, 3, 3)
+        inv_perm = (3, 4, 0, 1, 2)  # (nx, ny, nz, 3, 3) -> (3, 3, nx, ny, nz)
+        permittivities = jnp.linalg.inv(eps.transpose(perm)).transpose(inv_perm).reshape(9, *inv_permittivities.shape[1:])
+    else:
+        permittivities = 1 / inv_permittivities
     other_axes = [a for a in range(1, 4) if permittivities.shape[a] != 1]
     propagation_axis = permittivities.shape[1:].index(1)
-    coords = [np.arange(permittivities.shape[dim] + 1) * resolution / 1e-6 for dim in other_axes]
+    coords = [jnp.arange(permittivities.shape[dim] + 1) * resolution / 1e-6 for dim in other_axes]
     permittivity_squeezed = jnp.take(
         permittivities,
         indices=0,
@@ -201,8 +209,8 @@ def compute_mode(
         # propagation along z: tidy3d (x,y,z) → physical (x,y,z)
         perm_idx = [0, 1, 2]
 
-    # Only apply rotation if anisotropic (3 components)
-    if permittivity_squeezed.shape[0] == 3:
+    # Only apply rotation if anisotropic (3 or 9 components)
[…]
