+++ tests/units/test_unitful.py
@@ -23,6 +23,7 @@ from fdtdx.units.unitful import (
     multiply,
     ne,
     subtract,
+    sign
 )
 
 
@@ -1877,3 +1878,176 @@ def test_argmin_with_negative_nan_inf():
     # jnp.argmax([-1, nan, 5, inf]) → 3
     assert result1.val == jnp.array(1)
     assert result2.val == jnp.array(3)
+
+
+def test_sign_unitful_jax_array_drops_unit_and_preserves_float_dtype():
+    """Unitful(JAX float array) -> Unitful, unit dropped, JAX float dtype preserved."""
+    speed_unit = Unit(scale=3, dim={SI.m: 1, SI.s: -1})  # km/s
+    speeds = Unitful(val=jnp.array([-2.5, 0.0, 3.7, -1.2, 4.8], dtype=jnp.float32), unit=speed_unit)
+
+    result = jnp.sign(speeds)  # type: ignore
+
+    assert isinstance(result, Unitful)
+    assert result.unit.dim == {}
+    assert isinstance(result.val, jax.Array)
+    assert np.issubdtype(result.val.dtype, np.floating)
+    assert jnp.all(result.value() == jnp.array([-1.0, 0.0, 1.0, -1.0, 1.0], dtype=result.val.dtype))
+
+
+def test_sign_overload_jax_array_preserves_dtype():
+    """JAX array -> JAX array, dtype preserved (float->float)."""
+    x = jnp.array([-3.0, -0.0, 0.0, 2.2], dtype=jnp.float32)
+    y = sign(x)
+
+    assert isinstance(y, jax.Array)
+    assert np.issubdtype(y.dtype, np.floating)
+    assert jnp.all(y == jnp.array([-1.0, 0.0, 0.0, 1.0], dtype=y.dtype))
+
+
+def test_sign_overload_numpy_float_array_preserves_float_dtype():
+    """NumPy float array -> NumPy float array."""
+    x = np.array([-4.5, 0.0, 7.1], dtype=np.float64)
+    y = sign(x)
+
+    assert isinstance(y, np.ndarray)
+    assert np.issubdtype(y.dtype, np.floating)
+    assert np.array_equal(y, np.array([-1.0, 0.0, 1.0], dtype=y.dtype))
+
+
+def test_sign_overload_numpy_int_array_preserves_int_dtype():
+    """NumPy int array -> NumPy int array."""
+    x = np.array([-4, 0, 7], dtype=np.int32)
+    y = sign(x)
+
+    assert isinstance(y, np.ndarray)
+    assert np.issubdtype(y.dtype, np.integer)
+    assert np.array_equal(y, np.array([-1, 0, 1], dtype=y.dtype))
+
+
+def test_sign_overload_numpy_scalar():
+    """NumPy scalar -> NumPy scalar."""
+    x = np.float64(-0.3)
+    y = sign(x)
+
+    assert isinstance(y, np.number)
+    assert np.issubdtype(y.dtype, np.floating)
+    assert y == np.float64(-1.0)
+
+
+def test_sign_python_int_and_float():
+    """Python scalars go through dedicated overloads."""
+    assert sign(-7) == -1
+    assert sign(0) == 0
+    assert sign(9) == 1
+
+    assert sign(-7.2) == -1.0
+    assert sign(0.0) == 0.0
+    assert sign(9.9) == 1.0
+
+
+def test_sign_unitful_python_float_value():
+    """Unitful with python float val should be handled via sign(x.val) dispatch."""
+    unit = Unit(scale=0, dim={SI.s: 1})
+    u = Unitful(val=-0.3, unit=unit)
+
+    y = sign(u)
+
+    assert isinstance(y, Unitful)
+    assert y.unit.dim == {}
[…]
