### 2026-01-10 a1b83ee
changes in color definition
