+++ tests/units/test_unitful.py
@@ -17,6 +17,7 @@ from fdtdx.units.unitful import (
     Unit,
     Unitful,
     add,
+    cond,
     eq,
     ge,
     gt,
@@ -2494,3 +2495,315 @@ def test_argmin_with_negative_nan_inf():
     # jnp.argmax([-1, nan, 5, inf]) → 3
     assert result1.val == jnp.array(1)
     assert result2.val == jnp.array(3)
+
+
+def test_cond_unitful_numpy_operand_python_branch_true():
+    """All operand.val are numpy -> should use Python cond and only execute true branch."""
+    time = s * np.array([1.0, 2.0, 3.0])
+    operand = [time]
+
+    called = {"t": 0, "f": 0}
+
+    def true_fun(ops):
+        called["t"] += 1
+        x = ops[0]
+        return Unitful(val=np.squeeze(x.value()) + 100.0, unit=x.unit)
+
+    def false_fun(ops):
+        called["f"] += 1
+        x = ops[0]
+        return squeeze(x)
+
+    out = cond(True, true_fun, false_fun, operand)
+
+    assert called["t"] == 1
+    assert called["f"] == 0
+
+    assert isinstance(out, Unitful)
+    assert isinstance(out.val, np.ndarray)
+    assert out.val.shape == (3,)
+    assert np.allclose(out.value(), np.array([101.0, 102.0, 103.0]))
+    assert out.unit.dim == time.unit.dim
+
+
+def test_cond_unitful_numpy_operand_python_branch_false():
+    """All operand.val are numpy -> should use Python cond and only execute false branch."""
+    time = s * np.array([[1.0], [2.0], [3.0]])
+    operand = [time]
+
+    called = {"t": 0, "f": 0}
+
+    def true_fun(ops):
+        called["t"] += 1
+        x = ops[0]
+        return Unitful(val=np.squeeze(x.value()) + 999.0, unit=x.unit)
+
+    def false_fun(ops):
+        called["f"] += 1
+        x = ops[0]
+        return squeeze(x)
+
+    out = cond(np.bool_(False), true_fun, false_fun, operand)
+
+    assert called["t"] == 0
+    assert called["f"] == 1
+
+    assert isinstance(out, Unitful)
+    assert isinstance(out.val, np.ndarray)
+    assert out.val.shape == (3,)
+    assert np.allclose(out.value(), np.array([1.0, 2.0, 3.0]))
+    assert out.unit.dim == time.unit.dim
+
+
+def test_cond_unitful_jax_operand_uses_lax_cond():
+    """Operand contains jax.Array -> should use JAX cond and return jax.Array."""
+    time = s * jnp.array([[1.0], [2.0], [3.0]])
+    operand = [time]
+
+    def true_fun(ops):
+        x = ops[0]
+        return squeeze(x)
+
+    def false_fun(ops):
+        x = ops[0]
+        return Unitful(val=squeeze(x.value()) + 100.0, unit=x.unit)
+
+    out = cond(jnp.array(True), true_fun, false_fun, operand)
+
+    assert isinstance(out, Unitful)
+    assert isinstance(out.val, jax.Array)
+    assert out.val.shape == (3,)
+    assert jnp.allclose(out.value(), jnp.array([1.0, 2.0, 3.0]))
+    assert out.unit.dim == time.unit.dim
+
+
+def test_cond_unitful_jit_with_jax_operand_runs():
+    """cond should work under jax.jit when operand is JAX (pred is tracer)."""
+    time = s * jnp.array([[1.0], [2.0], [3.0]])
+    operand = [time]
+
+    def true_fun(ops):
+        x = ops[0]
+        return squeeze(x)
+
[…]
