### 2026-01-10 b26f9f8
changes in color definition in various files
