+++ src/fdtdx/objects/device/device.py
@@ -8,7 +8,7 @@ from fdtdx.config import SimulationConfig
 from fdtdx.core.jax.pytrees import autoinit, field, frozen_field, frozen_private_field
 from fdtdx.core.jax.utils import check_specs
 from fdtdx.core.misc import expand_matrix, is_float_divisible
-from fdtdx.core.plotting.colors import PINK
+from fdtdx.core.plotting.colors import XKCD_LIGHT_PINK
 from fdtdx.materials import Material
 from fdtdx.objects.device.parameters.transform import ParameterTransformation
 from fdtdx.objects.object import OrderableObject
@@ -37,8 +37,8 @@ class Device(OrderableObject, ABC):
     #: A Sequence of parameter transformation to be applied to the parameters when mapping them to simulation materials.
     param_transforms: Sequence[ParameterTransformation] = field()
 
-    #: Color of the object when plotted. Defaults to Pink.
-    color: tuple[float, float, float] | None = frozen_field(default=PINK)
+    #: Color of the object when plotted. Defaults to XKCD_LIGHT_PINK.
+    color: tuple[float, float, float] | None = frozen_field(default=XKCD_LIGHT_PINK)
 
     #: Size of the material voxels used within the device in metrical units (meter). Note that this is independent of the simulation voxel size.
     #: Defaults to undefined shape. For all three axes, either the voxel grid or real shape needs to be defined.
