### 2026-01-12 e9e6766
added post-init check for spectral width and a unit test
