### 2026-01-12 dcd43f3
full anisotropy unit tests
