+++ tests/test_anisotropic_materials.py
@@ -31,7 +31,7 @@ class TestMaterialClass:
         assert mat.electric_conductivity == (0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
         assert mat.magnetic_conductivity == (0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0)
 
-    def test_non_isotropic_material_tuple_input(self):
+    def test_diagonally_anisotropic_material_tuple_input(self):
         """Test that tuple inputs are stored correctly."""
         mat = Material(
             permittivity=(2.0, 2.5, 3.0),
@@ -44,14 +44,45 @@ class TestMaterialClass:
         assert mat.electric_conductivity == (0.1, 0.0, 0.0, 0.0, 0.2, 0.0, 0.0, 0.0, 0.3)
         assert mat.magnetic_conductivity == (0.01, 0.0, 0.0, 0.0, 0.02, 0.0, 0.0, 0.0, 0.03)
 
-    def test_mixed_isotropic_and_non_isotropic(self):
+    def test_fully_anisotropic_material_tuple_input(self):
+        """Test that tuple inputs are stored correctly."""
+        mat = Material(
+            permittivity=(2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0),
+            permeability=(1.0, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9),
+            electric_conductivity=(0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9),
+            magnetic_conductivity=(0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09),
+        )
+        assert mat.permittivity == (2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0)
+        assert mat.permeability == (1.0, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9)
+        assert mat.electric_conductivity == (0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9)
+        assert mat.magnetic_conductivity == (0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09)
+
+    def test_mixed_isotropic_and_diagonally_anisotropic(self):
         """Test material with some isotropic and some anisotropic properties."""
         mat = Material(
-            permittivity=(2.0, 2.5, 3.0),  # anisotropic
-            permeability=1.0,  # isotropic
+            permittivity=(2.0, 2.5, 3.0),
+            permeability=1.0,
         )
         assert mat.permittivity == (2.0, 0.0, 0.0, 0.0, 2.5, 0.0, 0.0, 0.0, 3.0)
         assert mat.permeability == (1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0)
+
+    def test_mixed_isotropic_and_fully_anisotropic(self):
+        """Test material with some isotropic and some fully anisotropic properties."""
+        mat = Material(
+            permittivity=(2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0),
+            permeability=1.0,
+        )
+        assert mat.permittivity == (2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0)
+        assert mat.permeability == (1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0)
+
[…]
