### 2026-01-13 0970fb8
updates JOSS paper references
