### 2026-01-13 9a20a94
Update changelog with issue reference for bug fix
