### 2026-01-13 d004b76
fix for the detector plotting bug (#223)
* fix for the detector plotting bug

* updated changelog.md

* updated changelog.md

* Update changelog with issue reference for bug fix

---------

Co-authored-by: Yannik Mahlau <59509701+ymahlau@users.noreply.github.com>
