### 2026-01-13 d1fe21e
Fix formatting of title in paper.bib
