+++ paper.bib
@@ -1,5 +1,5 @@
 @article{kaneyeeNumericalSolutionInitial1966,
-title = {Numerical Solution of Initial Boundary Value Problems Involving PM}axwell's Equations in Isotropic Media},
+title = {Numerical Solution of Initial Boundary Value Problems Involving {M}axwell's Equations in Isotropic Media},
 author = {Kane Yee},
 year = {1966},
 month = may,
