### 2026-01-13 2beae84
updates JOSS paper references (#224)
* updates JOSS paper references

* Fix formatting of title in paper.bib

---------

Co-authored-by: Yannik Mahlau <59509701+ymahlau@users.noreply.github.com>
