+++ paper.bib
@@ -1,6 +1,6 @@
 @article{kaneyeeNumericalSolutionInitial1966,
-title = {Numerical Solution of Initial Boundary Value Problems Involving Maxwell's Equations in Isotropic Media},
-author = {{Kane Yee}},
+title = {Numerical Solution of Initial Boundary Value Problems Involving {M}axwell's Equations in Isotropic Media},
+author = {Kane Yee},
 year = {1966},
 month = may,
 journal = {IEEE Transactions on Antennas and Propagation},
@@ -11,7 +11,7 @@ doi = {10.1109/TAP.1966.1138693}
 }
 
 @article{ceviche,
-title={Forward-Mode Differentiation of Maxwell’s Equations},
+title={Forward-Mode Differentiation of {M}axwell’s Equations},
 author={Hughes, Tyler W and Williamson, Ian AD and Minkov, Momchil and Fan, Shanhui},
 journal={ACS Photonics},
 volume={6},
@@ -24,9 +24,9 @@ doi = {10.1021/acsphotonics.9b01238}
 
 @misc{fdtdz,
 author = {Lu, Jesse and Vučković, Jelena},
-title = {fdtd-z : A systolic scheme for GPU-accelerated nanophotonic simulation},
+title = {fdtd-z: A systolic scheme for {GPU}-accelerated nanophotonic simulation},
 url = {https://github.com/spinsphotonics/fdtdz},
-howpublished = {https://github.com/spinsphotonics/fdtdz},
+howpublished = {GitHub repository},
 year = {2024}
 }
 
@@ -38,7 +38,7 @@ year={2022}
 }
 
 @article{meep,
-title = {Meep: A flexible free-software package for electromagnetic simulations by the FDTD method},
+title = {Meep: A flexible free-software package for electromagnetic simulations by the {FDTD} method},
 journal = {Computer Physics Communications},
 volume = {181},
 number = {3},
@@ -61,16 +61,17 @@ year = {2024}
 
 @misc{fdtd_laporte,
 author = {Laporte, Floris},
-title = {Python 3D FDTD Simulator},
+title = {Python {3D} {FDTD} Simulator},
 url = {https://github.com/flaport/fdtd},
-howpublished = {https://github.com/flaport/fdtd},
+howpublished = {GitHub repository},
 year = {2024}
 }
 
 @misc{khronos,
 author = {Alec Hammond},
 title = {Khronos},
-howpublished = {https://github.com/facebookresearch/Khronos.jl},
+url = {https://github.com/facebookresearch/Khronos.jl},
+howpublished = {GitHub repository},
 year = {2024},
 }
 
@@ -80,7 +81,7 @@ author = {Taflove, Allen and Hagness, Susan C.},
 edition = {3rd},
 keywords = {EM FDTD textbook},
 publisher = {Artech House},
-title = {Computational electrodynamics: the finite-difference time-domain method},
+title = {Computational Electrodynamics: The Finite-Difference Time-Domain Method},
 year = {2005},
 isbn = {9781580538329},
 doi = {10.1016/B978-012170960-0/50046-3}
@@ -99,7 +100,7 @@ doi = {10.1016/B978-012170960-0/50046-3}
 }
 
 @inproceedings{mahlau2025flexible,
-title={A flexible framework for large-scale FDTD simulations: open-source inverse design for 3D nanostructures},
+title={A flexible framework for large-scale {FDTD} simulations: open-source inverse design for {3D} nanostructures},
[…]
