### 2026-01-13 51f743a
more tests and example
