+++ tests/test_anisotropic_materials.py
@@ -57,6 +57,19 @@ class TestMaterialClass:
         assert mat.electric_conductivity == (0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9)
         assert mat.magnetic_conductivity == (0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09)
 
+    def test_fully_anisotropic_material_nested_tuple_input(self):
+        """Test that nested tuple inputs are stored correctly."""
+        mat = Material(
+            permittivity=((2.0, 2.5, 3.0), (3.5, 4.0, 4.5), (5.0, 5.5, 6.0)),
+            permeability=((1.0, 1.2, 1.3), (1.4, 1.5, 1.6), (1.7, 1.8, 1.9)),
+            electric_conductivity=((0.1, 0.2, 0.3), (0.4, 0.5, 0.6), (0.7, 0.8, 0.9)),
+            magnetic_conductivity=((0.01, 0.02, 0.03), (0.04, 0.05, 0.06), (0.07, 0.08, 0.09)),
+        )
+        assert mat.permittivity == (2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0)
+        assert mat.permeability == (1.0, 1.2, 1.3, 1.4, 1.5, 1.6, 1.7, 1.8, 1.9)
+        assert mat.electric_conductivity == (0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9)
+        assert mat.magnetic_conductivity == (0.01, 0.02, 0.03, 0.04, 0.05, 0.06, 0.07, 0.08, 0.09)
+
     def test_mixed_isotropic_and_diagonally_anisotropic(self):
         """Test material with some isotropic and some anisotropic properties."""
         mat = Material(
@@ -75,10 +88,28 @@ class TestMaterialClass:
         assert mat.permittivity == (2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0)
         assert mat.permeability == (1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0)
 
+    def test_mixed_isotropic_and_nested_fully_anisotropic(self):
+        """Test material with some isotropic and some nested fully anisotropic properties."""
+        mat = Material(
+            permittivity=((2.0, 2.5, 3.0), (3.5, 4.0, 4.5), (5.0, 5.5, 6.0)),
+            permeability=1.0,
+        )
+        assert mat.permittivity == (2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0)
+        assert mat.permeability == (1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0)
+
     def test_mixed_diagonally_anisotropic_and_fully_anisotropic(self):
         """Test material with some diagonally anisotropic and some fully anisotropic properties."""
         mat = Material(
             permittivity=(2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0),
+            permeability=(1.0, 1.5, 2.0),
+        )
+        assert mat.permittivity == (2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0)
+        assert mat.permeability == (1.0, 0.0, 0.0, 0.0, 1.5, 0.0, 0.0, 0.0, 2.0)
+
+    def test_mixed_diagonally_anisotropic_and_nested_fully_anisotropic(self):
[…]
