### 2026-01-13 #8912 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
