### 2026-01-14 #8912 @arfon
The PDF should be rendering now (I can see it here: https://www.theoj.org/joss-papers/joss.08912/10.21105.joss.08912.pdf) but browser caching might make it take a while to show up properly for you @ymahlau .
